Split expenses view into listing and creation
Closes #589
