+++ public/js/app.js
@@ -1,1 +1,1 @@
-(window.webpackJsonp=window.webpackJsonp||[]).push([[0],{"+KkT":function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\n#users-list[data-v-21c16506] .el-table .success-row {\n  background: #f0f9eb;\n}\n@media (max-width: 575px) {\n#users-list[data-v-21c16506] .el-dialog {\n    width: 100%;\n}\n}\n",""])},"/BTM":function(t,e,n){var r=n("fysm");"string"==typeof r&&(r=[[t.i,r,""]]);var s={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(r,s);r.locals&&(t.exports=r.locals)},0:function(t,e,n){n("bUC5"),t.exports=n("pyCd")},"0P3G":function(t,e,n){"use strict";var r=n("/BTM");n.n(r).a},"1XD5":function(t,e,n){"use strict";var r=n("KJzQ");n.n(r).a},"69hy":function(t,e,n){var r=n("JnLb");"string"==typeof r&&(r=[[t.i,r,""]]);var s={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(r,s);r.locals&&(t.exports=r.locals)},"8O32":function(t,e,n){var r=n("eGZ0");"string"==typeof r&&(r=[[t.i,r,""]]);var s={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(r,s);r.locals&&(t.exports=r.locals)},"9bXS":function(t,e,n){var r=n("gE6I");"string"==typeof r&&(r=[[t.i,r,""]]);var s={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(r,s);r.locals&&(t.exports=r.locals)},A19t:function(t,e,n){var r=n("+KkT");"string"==typeof r&&(r=[[t.i,r,""]]);var s={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(r,s);r.locals&&(t.exports=r.locals)},"D1a/":function(t,e,n){var r=n("WGSt");"string"==typeof r&&(r=[[t.i,r,""]]);var s={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(r,s);r.locals&&(t.exports=r.locals)},F4tQ:function(t,e,n){"use strict";var r=n("WwcV");n.n(r).a},GFEm:function(t,e,n){var r=n("yx/N");"string"==typeof r&&(r=[[t.i,r,""]]);var s={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(r,s);r.locals&&(t.exports=r.locals)},GYTP:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\n.slide-fade-enter-active[data-v-764fdbfc] {\n  transition: all .3s ease;\n}\n.slide-fade-leave-active[data-v-764fdbfc] {\n  transition: all .5s cubic-bezier(1.0, 0.5, 0.8, 1.0);\n}\n.slide-fade-enter[data-v-764fdbfc], .slide-fade-leave-to[data-v-764fdbfc] {\n  background-color: #ffef0029;\n}\n",""])},HF9f:function(t,e,n){"use strict";var r=n("69hy");n.n(r).a},IUMi:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n/* Chrome, Safari, Edge, Opera */\n[data-v-8243289e] input::-webkit-outer-spin-button,[data-v-8243289e] input::-webkit-inner-spin-button {\n  -webkit-appearance: none;\n  margin: 0;\n}\n\n/* Firefox */\n[data-v-8243289e] input[type=number] {\n  -moz-appearance: textfield;\n}\n\n",""])},IqpQ:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\ntextarea[data-v-21a1e722] {\n  outline: 0;\n  background: #f4f4f4;\n  box-shadow: 1px 1px 0 #ccc;\n  border-color: transparent;\n  border-radius: 0;\n  font-family: 'Inconsolata', monospace;\n}\ntextarea[data-v-21a1e722]:focus {\n  box-shadow: 1px 1px 0 #ccc;\n  outline: 0;\n  background: #f4f4f4;\n  border-color: transparent;\n}\n.new-tag-group[data-v-21a1e722] {\n  font-size: 0;\n}\n.new-tag-group .input-new-tag[data-v-21a1e722], .new-tag-group .button-new-tag[data-v-21a1e722] {\n  vertical-align: middle;\n  width: 120px;\n}\n#notes-create[data-v-21a1e722] .el-checkbox-button {\n  border-radius: 4px;\n  margin-right: 0 !important;\n}\n#notes-create[data-v-21a1e722] .el-checkbox-button.is-checked .el-checkbox-button__inner {\n  border-color: #409EFF;\n}\n",""])},J82J:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\n.slide-enter-active[data-v-1a92697e] {\n  transition-duration: 0.3s;\n  transition-timing-function: ease-in;\n}\n.slide-leave-active[data-v-1a92697e] {\n  transition-duration: 0.3s;\n  transition-timing-function: cubic-bezier(0, 1, 0.5, 1);\n}\n.slide-enter-to[data-v-1a92697e], .slide-leave[data-v-1a92697e] {\n  max-height: 1000px;\n  overflow: hidden;\n}\n.slide-enter[data-v-1a92697e], .slide-leave-to[data-v-1a92697e] {\n  overflow: hidden;\n  max-height: 0;\n}\n.text-line-through[data-v-1a92697e] {\n  text-decoration: line-through;\n}\n.btn-collapse[data-v-1a92697e] {\n  font-family: 'Inconsolata', monospace;\n  padding: 0;\n  height: 23px;\n}\n",""])},JnLb:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"#main-navigation .router-link-active[data-v-590a6ef5] {\n  color: #3490dc;\n}",""])},KJzQ:function(t,e,n){var r=n("IqpQ");"string"==typeof r&&(r=[[t.i,r,""]]);var s={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(r,s);r.locals&&(t.exports=r.locals)},Ofak:function(t,e,n){"use strict";var r=n("V4XD");n.n(r).a},P9EP:function(t,e,n){"use strict";var r=n("avtZ");n.n(r).a},RnhZ:function(t,e,n){var r={"./af":"K/tc","./af.js":"K/tc","./ar":"jnO4","./ar-dz":"o1bE","./ar-dz.js":"o1bE","./ar-kw":"Qj4J","./ar-kw.js":"Qj4J","./ar-ly":"HP3h","./ar-ly.js":"HP3h","./ar-ma":"CoRJ","./ar-ma.js":"CoRJ","./ar-sa":"gjCT","./ar-sa.js":"gjCT","./ar-tn":"bYM6","./ar-tn.js":"bYM6","./ar.js":"jnO4","./az":"SFxW","./az.js":"SFxW","./be":"H8ED","./be.js":"H8ED","./bg":"hKrs","./bg.js":"hKrs","./bm":"p/rL","./bm.js":"p/rL","./bn":"kEOa","./bn.js":"kEOa","./bo":"0mo+","./bo.js":"0mo+","./br":"aIdf","./br.js":"aIdf","./bs":"JVSJ","./bs.js":"JVSJ","./ca":"1xZ4","./ca.js":"1xZ4","./cs":"PA2r","./cs.js":"PA2r","./cv":"A+xa","./cv.js":"A+xa","./cy":"l5ep","./cy.js":"l5ep","./da":"DxQv","./da.js":"DxQv","./de":"tGlX","./de-at":"s+uk","./de-at.js":"s+uk","./de-ch":"u3GI","./de-ch.js":"u3GI","./de.js":"tGlX","./dv":"WYrj","./dv.js":"WYrj","./el":"jUeY","./el.js":"jUeY","./en-au":"Dmvi","./en-au.js":"Dmvi","./en-ca":"OIYi","./en-ca.js":"OIYi","./en-gb":"Oaa7","./en-gb.js":"Oaa7","./en-ie":"4dOw","./en-ie.js":"4dOw","./en-il":"czMo","./en-il.js":"czMo","./en-in":"7C5Q","./en-in.js":"7C5Q","./en-nz":"b1Dy","./en-nz.js":"b1Dy","./en-sg":"t+mt","./en-sg.js":"t+mt","./eo":"Zduo","./eo.js":"Zduo","./es":"iYuL","./es-do":"CjzT","./es-do.js":"CjzT","./es-us":"Vclq","./es-us.js":"Vclq","./es.js":"iYuL","./et":"7BjC","./et.js":"7BjC","./eu":"D/JM","./eu.js":"D/JM","./fa":"jfSC","./fa.js":"jfSC","./fi":"gekB","./fi.js":"gekB","./fil":"1ppg","./fil.js":"1ppg","./fo":"ByF4","./fo.js":"ByF4","./fr":"nyYc","./fr-ca":"2fjn","./fr-ca.js":"2fjn","./fr-ch":"Dkky","./fr-ch.js":"Dkky","./fr.js":"nyYc","./fy":"cRix","./fy.js":"cRix","./ga":"USCx","./ga.js":"USCx","./gd":"9rRi","./gd.js":"9rRi","./gl":"iEDd","./gl.js":"iEDd","./gom-deva":"qvJo","./gom-deva.js":"qvJo","./gom-latn":"DKr+","./gom-latn.js":"DKr+","./gu":"4MV3","./gu.js":"4MV3","./he":"x6pH","./he.js":"x6pH","./hi":"3E1r","./hi.js":"3E1r","./hr":"S6ln","./hr.js":"S6ln","./hu":"WxRl","./hu.js":"WxRl","./hy-am":"1rYy","./hy-am.js":"1rYy","./id":"UDhR","./id.js":"UDhR","./is":"BVg3","./is.js":"BVg3","./it":"bpih","./it-ch":"bxKX","./it-ch.js":"bxKX","./it.js":"bpih","./ja":"B55N","./ja.js":"B55N","./jv":"tUCv","./jv.js":"tUCv","./ka":"IBtZ","./ka.js":"IBtZ","./kk":"bXm7","./kk.js":"bXm7","./km":"6B0Y","./km.js":"6B0Y","./kn":"PpIw","./kn.js":"PpIw","./ko":"Ivi+","./ko.js":"Ivi+","./ku":"JCF/","./ku.js":"JCF/","./ky":"lgnt","./ky.js":"lgnt","./lb":"RAwQ","./lb.js":"RAwQ","./lo":"sp3z","./lo.js":"sp3z","./lt":"JvlW","./lt.js":"JvlW","./lv":"uXwI","./lv.js":"uXwI","./me":"KTz0","./me.js":"KTz0","./mi":"aIsn","./mi.js":"aIsn","./mk":"aQkU","./mk.js":"aQkU","./ml":"AvvY","./ml.js":"AvvY","./mn":"lYtQ","./mn.js":"lYtQ","./mr":"Ob0Z","./mr.js":"Ob0Z","./ms":"6+QB","./ms-my":"ZAMP","./ms-my.js":"ZAMP","./ms.js":"6+QB","./mt":"G0Uy","./mt.js":"G0Uy","./my":"honF","./my.js":"honF","./nb":"bOMt","./nb.js":"bOMt","./ne":"OjkT","./ne.js":"OjkT","./nl":"+s0g","./nl-be":"2ykv","./nl-be.js":"2ykv","./nl.js":"+s0g","./nn":"uEye","./nn.js":"uEye","./oc-lnc":"Fnuy","./oc-lnc.js":"Fnuy","./pa-in":"8/+R","./pa-in.js":"8/+R","./pl":"jVdC","./pl.js":"jVdC","./pt":"8mBD","./pt-br":"0tRk","./pt-br.js":"0tRk","./pt.js":"8mBD","./ro":"lyxo","./ro.js":"lyxo","./ru":"lXzo","./ru.js":"lXzo","./sd":"Z4QM","./sd.js":"Z4QM","./se":"//9w","./se.js":"//9w","./si":"7aV9","./si.js":"7aV9","./sk":"e+ae","./sk.js":"e+ae","./sl":"gVVK","./sl.js":"gVVK","./sq":"yPMs","./sq.js":"yPMs","./sr":"zx6S","./sr-cyrl":"E+lV","./sr-cyrl.js":"E+lV","./sr.js":"zx6S","./ss":"Ur1D","./ss.js":"Ur1D","./sv":"X709","./sv.js":"X709","./sw":"dNwA","./sw.js":"dNwA","./ta":"PeUW","./ta.js":"PeUW","./te":"XLvN","./te.js":"XLvN","./tet":"V2x9","./tet.js":"V2x9","./tg":"Oxv6","./tg.js":"Oxv6","./th":"EOgW","./th.js":"EOgW","./tl-ph":"Dzi0","./tl-ph.js":"Dzi0","./tlh":"z3Vd","./tlh.js":"z3Vd","./tr":"DoHr","./tr.js":"DoHr","./tzl":"z1FC","./tzl.js":"z1FC","./tzm":"wQk9","./tzm-latn":"tT3J","./tzm-latn.js":"tT3J","./tzm.js":"wQk9","./ug-cn":"YRex","./ug-cn.js":"YRex","./uk":"raLr","./uk.js":"raLr","./ur":"UpQW","./ur.js":"UpQW","./uz":"Loxo","./uz-latn":"AQ68","./uz-latn.js":"AQ68","./uz.js":"Loxo","./vi":"KSF8","./vi.js":"KSF8","./x-pseudo":"/X5v","./x-pseudo.js":"/X5v","./yo":"fzPg","./yo.js":"fzPg","./zh-cn":"XDpg","./zh-cn.js":"XDpg","./zh-hk":"SatO","./zh-hk.js":"SatO","./zh-mo":"OmwH","./zh-mo.js":"OmwH","./zh-tw":"kOpN","./zh-tw.js":"kOpN"};function s(t){var e=a(t);return n(e)}function a(t){if(!n.o(r,t)){var e=new Error("Cannot find module '"+t+"'");throw e.code="MODULE_NOT_FOUND",e}return r[t]}s.keys=function(){return Object.keys(r)},s.resolve=a,t.exports=s,s.id="RnhZ"},V4XD:function(t,e,n){var r=n("IUMi");"string"==typeof r&&(r=[[t.i,r,""]]);var s={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(r,s);r.locals&&(t.exports=r.locals)},WGSt:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\n.bookmark-story[data-v-cdd94d26] {\n  position: relative;\n  top: 2px;\n  text-decoration: none;\n}\n.loader-text[data-v-cdd94d26] {\n  background: #808080;\n  border-radius: 2px;\n}\n.loader-text--1x3[data-v-cdd94d26] {\n  width: 33%;\n}\n.loader-text--2x3[data-v-cdd94d26] {\n  width: 66%;\n}\n",""])},WwcV:function(t,e,n){var r=n("GYTP");"string"==typeof r&&(r=[[t.i,r,""]]);var s={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(r,s);r.locals&&(t.exports=r.locals)},avtZ:function(t,e,n){var r=n("J82J");"string"==typeof r&&(r=[[t.i,r,""]]);var s={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(r,s);r.locals&&(t.exports=r.locals)},bUC5:function(t,e,n){"use strict";n.r(e);var r=n("XuX8"),s=n.n(r),a=n("L2JU"),o=n("o0o1"),i=n.n(o),c=n("vDqi"),u=n.n(c),l=n("yDJ3"),d=n.n(l);function m(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}function p(t){return function(){var e=this,n=arguments;return new Promise((function(r,s){var a=t.apply(e,n);function o(t){m(a,r,s,o,i,"next",t)}function i(t){m(a,r,s,o,i,"throw",t)}o(void 0)}))}}var f={state:{notes:[],searchQuery:""},getters:{notesList:function(t){return t.notes},searchQuery:function(t){return t.searchQuery}},actions:{fetchNotesList:function(t){var e=arguments;return p(i.a.mark((function n(){var r,s,a,o;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:return r=t.commit,s=e.length>1&&void 0!==e[1]?e[1]:"",r("setSearchQuery",s),n.next=5,u.a.get("/api/notes?query=".concat(s));case 5:return a=n.sent,o=d()(a,"data.data",[]),r("setNotesList",o),n.abrupt("return",!0);case 9:case"end":return n.stop()}}),n)})))()},addNote:function(t,e){return p(i.a.mark((function n(){var r,s,a;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:return t.commit,n.next=3,u.a.post("/api/notecreate",{body:e.body,tags:e.tags});case 3:if(r=n.sent,s=d()(r,"data.data.id"),a=d()(r,"data.data.success"),!s||!a){n.next=8;break}return n.abrupt("return",s);case 8:return n.abrupt("return",!1);case 9:case"end":return n.stop()}}),n)})))()},updateSearchQuery:function(t,e){(0,t.commit)("setSearchQuery",e)}},mutations:{setNotesList:function(t,e){return t.notes=e},setSearchQuery:function(t,e){return t.searchQuery=e}}};function v(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}function h(t){return function(){var e=this,n=arguments;return new Promise((function(r,s){var a=t.apply(e,n);function o(t){v(a,r,s,o,i,"next",t)}function i(t){v(a,r,s,o,i,"throw",t)}o(void 0)}))}}var b={state:{expenses:[]},getters:{expensesList:function(t){return t.expenses}},actions:{fetchExpenses:function(t){return h(i.a.mark((function e(){var n,r,s;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return n=t.commit,e.next=3,u.a.get("/api/expenses");case 3:return r=e.sent,s=d()(r,"data.data",[]),n("setExpenses",s),e.abrupt("return",!0);case 7:case"end":return e.stop()}}),e)})))()},addExpense:function(t,e){return h(i.a.mark((function n(){var r,s,a,o;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:return r=t.commit,n.next=3,u.a.post("api/expenses/create",e);case 3:return s=n.sent,a=d()(s,"data.data"),(o=d()(a,"expense"))&&r("newExpense",o),n.abrupt("return",d()(a,"success",!1));case 8:case"end":return n.stop()}}),n)})))()}},mutations:{setExpenses:function(t,e){return t.expenses=e},newExpense:function(t,e){return t.expenses.unshift(e)}}};function g(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}function y(t){return function(){var e=this,n=arguments;return new Promise((function(r,s){var a=t.apply(e,n);function o(t){g(a,r,s,o,i,"next",t)}function i(t){g(a,r,s,o,i,"throw",t)}o(void 0)}))}}var x={state:{users:[],currentUserId:null},getters:{usersList:function(t){return t.users},currentUserId:function(t){return t.currentUserId}},actions:{fetchUsers:function(t){return y(i.a.mark((function e(){var n,r,s,a;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return n=t.commit,e.next=3,u.a.get("/api/users");case 3:return r=e.sent,s=d()(r,"data.data.users",[]),a=d()(r,"data.data.currentUserId"),n("setUsersList",s),n("setCurrentUserId",a),e.abrupt("return",!0);case 9:case"end":return e.stop()}}),e)})))()},addUser:function(t,e){return y(i.a.mark((function n(){var r,s,a,o;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:return r=t.commit,n.next=3,u.a.post("/api/users/create",e);case 3:return s=n.sent,a=d()(s,"data.data"),(o=d()(a,"user"))&&r("newUser",o),n.abrupt("return",d()(a,"success",!1));case 8:case"end":return n.stop()}}),n)})))()},updateUser:function(t,e){return y(i.a.mark((function n(){var r,s,a,o;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:return r=t.commit,s=!1,n.prev=2,a={id:e.id,name:e.name},e.password&&(a.password=e.password),n.next=7,u.a.put("/api/users/update",a);case 7:o=n.sent,(s=d()(o,"data.data.success",!1))&&r("updateUser",e),n.next=16;break;case 12:n.prev=12,n.t0=n.catch(2),s=!1,console.error(n.t0);case 16:return n.abrupt("return",s);case 17:case"end":return n.stop()}}),n,null,[[2,12]])})))()},destroyUser:function(t,e){return y(i.a.mark((function n(){var r,s,a;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:return r=t.commit,s=!1,n.prev=2,n.next=5,u.a.delete("/api/users/destroy",{data:{id:e}});case 5:a=n.sent,(s=d()(a,"data.data.success",!1))&&r("deleteUser",e),n.next=14;break;case 10:n.prev=10,n.t0=n.catch(2),s=!1,console.error(n.t0);case 14:return n.abrupt("return",s);case 15:case"end":return n.stop()}}),n,null,[[2,10]])})))()}},mutations:{setUsersList:function(t,e){return t.users=e},newUser:function(t,e){return t.user.push(e)},updateUser:function(t,e){t.users.filter((function(t){return t.id===e.id})).name=e.name},setCurrentUserId:function(t,e){return t.currentUserId=e},deleteUser:function(t,e){var n=t.users.findIndex((function(t){return t.id===e}));t.users.splice(n,1)}}};function w(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}function _(t){return function(){var e=this,n=arguments;return new Promise((function(r,s){var a=t.apply(e,n);function o(t){w(a,r,s,o,i,"next",t)}function i(t){w(a,r,s,o,i,"throw",t)}o(void 0)}))}}function k(){return j.apply(this,arguments)}function j(){return(j=_(i.a.mark((function t(){var e,n,r;return i.a.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return e=200,t.prev=1,t.next=4,u.a.get("/api/users/currentUser");case 4:r=t.sent,n=r.status===e,t.next=11;break;case 8:t.prev=8,t.t0=t.catch(1),n=!1;case 11:return t.abrupt("return",n);case 12:case"end":return t.stop()}}),t,null,[[1,8]])})))).apply(this,arguments)}function C(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}function O(t){return function(){var e=this,n=arguments;return new Promise((function(r,s){var a=t.apply(e,n);function o(t){C(a,r,s,o,i,"next",t)}function i(t){C(a,r,s,o,i,"throw",t)}o(void 0)}))}}var I={state:{user:function(){var t;try{t=JSON.parse(window.localStorage.getItem("ZE_UserInfo"))}catch(e){t=null}return t}()},getters:{userInfo:function(t){return t.user}},mutations:{setUserInfo:function(t,e){return t.user=e}},actions:{authenticate:function(t,e){return O(i.a.mark((function n(){var r,s,a;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:return t.commit,r=t.dispatch,s=204,n.next=4,u.a.post("/login",e);case 4:if((a=n.sent).status!==s){n.next=8;break}return n.next=8,r("fetchCurrentUser");case 8:return n.abrupt("return",a);case 9:case"end":return n.stop()}}),n)})))()},fetchCurrentUser:function(t){return O(i.a.mark((function e(){var n,r,s;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return n=t.commit,e.next=3,u.a.get("/api/users/currentUser");case 3:r=e.sent,s=d()(r,"data.data"),a=s,o=void 0,o=JSON.stringify(a),window.localStorage.setItem("ZE_UserInfo",o),n("setUserInfo",s);case 7:case"end":return e.stop()}var a,o}),e)})))()}}};s.a.use(a.a);var $=new a.a.Store({modules:{notes:f,expenses:b,users:x,auth:I}}),E=n("jE9Z");function P(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}function T(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(t);e&&(r=r.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,r)}return n}function S(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?T(Object(n),!0).forEach((function(e){N(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):T(Object(n)).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}function N(t,e,n){return e in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}var D={name:"Authenticate",data:function(){return{userForm:{email:"",password:""},pollingPasswordFieldInterval:null,errorMessage:null}},methods:S(S({},Object(a.b)(["authenticate"])),{},{pollingPasswordField:function(){var t=this;this.pollingPasswordFieldInterval=setInterval((function(){var e=document.createEvent("HTMLEvents");e.initEvent("change",!1,!0),t.$refs.pwf.dispatchEvent(e)}),500)},passwordChanged:function(){this.userForm.password=this.$refs.pwf.value},attemptLogin:function(t){var e,n=this;return(e=i.a.mark((function e(){var r;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return t.preventDefault(),e.next=3,u.a.get("/sanctum/csrf-cookie");case 3:return e.next=5,n.authenticate({email:n.userForm.email,password:n.userForm.password});case 5:r=e.sent,n.checkLoginResponse(r);case 7:case"end":return e.stop()}}),e)})),function(){var t=this,n=arguments;return new Promise((function(r,s){var a=e.apply(t,n);function o(t){P(a,r,s,o,i,"next",t)}function i(t){P(a,r,s,o,i,"throw",t)}o(void 0)}))})()},checkLoginResponse:function(t){var e=t.status;422===e?this.showsErrors(d()(t,"data.errors")):302!==e&&204!==e||(clearInterval(this.pollingPasswordFieldInterval),this.$router.push(Ie))},showsErrors:function(t){var e="";Object.keys(t).forEach((function(n){t[n].forEach((function(t){e+=t+"<br/>"}))})),this.$message({type:"error",dangerouslyUseHTMLString:!0,message:e})}}),mounted:function(){this.pollingPasswordField()}},U=n("KHd+"),F=Object(U.a)(D,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"pt-4",attrs:{id:"view-authenticate"}},[n("div",{staticClass:"row justify-content-center"},[n("div",{staticClass:"col-md-8 no-gutter-xs"},[n("div",{staticClass:"card shadow"},[n("div",{staticClass:"card-header"},[t._v(t._s(t.$I18n.trans("auth.login")))]),t._v(" "),n("div",{staticClass:"card-body"},[n("form",{attrs:{method:"POST",model:t.userForm},on:{submit:function(e){return t.attemptLogin(e)}}},[n("div",{staticClass:"form-group row"},[n("label",{staticClass:"col-md-4 col-form-label text-md-right",attrs:{for:"email"}},[t._v(t._s(t.$I18n.trans("auth.email")))]),t._v(" "),n("div",{staticClass:"col-md-6"},[n("input",{directives:[{name:"model",rawName:"v-model",value:t.userForm.email,expression:"userForm.email"}],staticClass:"form-control",attrs:{id:"email",inputmode:"email",type:"email",name:"email",autocomplete:"email",required:""},domProps:{value:t.userForm.email},on:{input:function(e){e.target.composing||t.$set(t.userForm,"email",e.target.value)}}})])]),t._v(" "),n("div",{staticClass:"form-group row"},[n("label",{staticClass:"col-md-4 col-form-label text-md-right",attrs:{for:"password"}},[t._v(t._s(t.$I18n.trans("auth.password")))]),t._v(" "),n("div",{staticClass:"col-md-6"},[n("input",{directives:[{name:"model",rawName:"v-model",value:t.userForm.password,expression:"userForm.password"}],ref:"pwf",staticClass:"form-control",attrs:{id:"password",name:"password",type:"password",autocomplete:"current-password",required:""},domProps:{value:t.userForm.password},on:{change:t.passwordChanged,input:function(e){e.target.composing||t.$set(t.userForm,"password",e.target.value)}}})])]),t._v(" "),n("div",{staticClass:"form-group row"},[n("div",{staticClass:"col-md-6 offset-md-4"},[n("div",{staticClass:"form-check"},[n("input",{staticClass:"form-check-input",attrs:{type:"checkbox",name:"remember",id:"remember",checked:""}}),t._v(" "),n("label",{staticClass:"form-check-label",attrs:{for:"remember"}},[t._v(t._s(t.$I18n.trans("auth.remember_me")))])])])]),t._v(" "),n("div",{staticClass:"form-group row mb-0"},[n("div",{staticClass:"col-md-8 offset-md-4"},[n("button",{staticClass:"btn btn-primary",attrs:{type:"submit",disabled:!t.userForm.email&&!t.userForm.password}},[t._v(t._s(t.$I18n.trans("auth.login")))])])])])])])])])])}),[],!1,null,null,null).exports,B={name:"NotFound",props:{message:{type:String,required:!1,default:"We could not find what you where looking for!"},linksTo:{type:Array,required:!1}},data:function(){return{dashboardRoute:Ie}}},A=Object(U.a)(B,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"pt-4",attrs:{id:"NotFound"}},[n("h1",{staticClass:"text-center text-capitalize"},[t._v("404 - page not found")]),t._v(" "),n("h5",{staticClass:"text-center text-black-50 mt-4"},[t._v(t._s(t.message))]),t._v(" "),n("div",{staticClass:"px-5 pt-3 text-center"},[n("router-link",{staticClass:"btn btn-lg btn-link",attrs:{to:t.dashboardRoute}},[t._v("Home")]),t._v(" "),t._l(t.linksTo,(function(e){return n("router-link",{key:e.route.name,staticClass:"btn btn-lg btn-link",attrs:{to:e.route}},[t._v(t._s(t.route.linkText))])}))],2)])}),[],!1,null,null,null).exports,R=n("wd/R"),z=n.n(R);function L(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}var M={name:"Weather",data:function(){return{loading:!0,weather:{icon:"",description:""}}},methods:{fetchWeather:function(){var t,e=this;return(t=i.a.mark((function t(){var n,r,s,a,o;return i.a.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:n="openweather--currentweather",s=localStorage.getItem(n);try{r=JSON.parse(s)}catch(t){r=null}if(r&&r.data&&!(r.saved+6e5<Date.now())){t.next=12;break}return r={data:{},saved:Date.now()},t.next=7,fetch("http://api.openweathermap.org/data/2.5/weather?q=Funchal,PT&appid=fe66d26d5b01501290be42a4809bbc4e&units=metric&lang=pt");case 7:return a=t.sent,t.next=10,a.json();case 10:r.data=t.sent,localStorage.setItem(n,JSON.stringify(r));case 12:o=d()(r,"data.weather[0].icon"),e.weather.icon="http://openweathermap.org/img/wn/".concat(o,".png"),e.weather.temp=d()(r,"data.main.temp"),e.weather.description=d()(r,"data.weather[0].description"),setTimeout((function(){return e.loading=!1}),500);case 17:case"end":return t.stop()}}),t)})),function(){var e=this,n=arguments;return new Promise((function(r,s){var a=t.apply(e,n);function o(t){L(a,r,s,o,i,"next",t)}function i(t){L(a,r,s,o,i,"throw",t)}o(void 0)}))})()}},filters:{capitalize:function(t){if(t){var e=t.toString();return e.charAt(0).toUpperCase()+e.slice(1)}},localTimeFromUnixTimestamp:function(t){if(t)return z.a.unix(t).format("HH:mm")}},created:function(){this.fetchWeather(),setInterval(this.fetchWeather,6e5)}},q=(n("wdM6"),Object(U.a)(M,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}],attrs:{id:"weather"}},[n("div",{staticClass:"card mb-3 shadow"},[n("div",{staticClass:"card-body d-flex flex-wrap"},[n("div",{staticClass:"align-self-center"},[n("a",{attrs:{href:"https://openweathermap.org/city/2267827",target:"_blank"}},[n("img",{attrs:{src:t.weather.icon,alt:t.weather.description}})])]),t._v(" "),n("div",{staticClass:"divider align-items-stretch mx-2"}),t._v(" "),n("div",{staticClass:"align-self-center ml-2"},[t._v("\n        "+t._s(t._f("capitalize")(t.weather.description))+"\n      ")]),t._v(" "),n("div",{staticClass:"divider align-items-stretch mx-2"}),t._v(" "),n("div",{staticClass:"align-self-center"},[t._v("\n        "+t._s(t.weather.temp)+" ℃\n      ")])])])])}),[],!1,null,"3d91250f",null).exports);function Q(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}var H={name:"SystemInfo",data:function(){return{loading:!0,memory:{free:"",used:"",total:""},nQueueWorkersRunning:0,up:"",upSince:"",systemInfoLoop:null}},created:function(){this.fetchSystemInfo(),this.systemInfoLoop=setInterval(this.fetchSystemInfo,1e4)},methods:{fetchSystemInfo:function(){var t,e=this;return(t=i.a.mark((function t(){var n,r;return i.a.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.next=2,u.a.get("/api/system-info");case 2:n=t.sent,(r=d()(n,"data.data",null))&&(e.memory.free=d()(r,"memory.free"),e.memory.used=d()(r,"memory.used"),e.memory.total=d()(r,"memory.total"),e.nQueueWorkersRunning=d()(r,"nQueueWorkersRunning"),e.up=d()(r,"up"),e.upSince=d()(r,"upSince")),setTimeout((function(){return e.loading=!1}),500);case 6:case"end":return t.stop()}}),t)})),function(){var e=this,n=arguments;return new Promise((function(r,s){var a=t.apply(e,n);function o(t){Q(a,r,s,o,i,"next",t)}function i(t){Q(a,r,s,o,i,"throw",t)}o(void 0)}))})()}},filters:{capitalize:function(t){if(t)return(t=t.toString()).charAt(0).toUpperCase()+t.slice(1)}}};n("F4tQ");function V(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}var Y={name:"Inspire",data:function(){return{inspire:null,fallbackLines:["The hardest part of building software is deciding what to build.","Everything not specified will be specified by the programmer, on the fly.","Design is also specification.","Everything that is public-facing defines the significant part of the business logic.","The open office plan is the best way to destroy productivity.","Don’t multitask. Focus on one task at a time.","There is work time and there is personal time. Don’t mix the two.","Everyone writes bad code except for you. /s"]}},computed:{fallback:function(){return this.fallbackLines[Math.floor(Math.random()*this.fallbackLines.length)]}},created:function(){this.fetchInspire()},methods:{fetchInspire:function(){var t,e=this;return(t=i.a.mark((function t(){var n;return i.a.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.next=2,u.a.get("/api/inspire");case 2:n=t.sent,e.inspire=d()(n,"data.data");case 4:case"end":return t.stop()}}),t)})),function(){var e=this,n=arguments;return new Promise((function(r,s){var a=t.apply(e,n);function o(t){V(a,r,s,o,i,"next",t)}function i(t){V(a,r,s,o,i,"throw",t)}o(void 0)}))})()}}},J={name:"Dashboard",components:{Weather:q,SystemInfo:Object(U.a)(H,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"card mb-3 shadow"},[n("div",{staticClass:"card-header"},[t._v(t._s(t.$I18n.trans("system.info")))]),t._v(" "),n("div",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}],staticClass:"card-body"},[n("transition",{attrs:{name:"slide-fade",mode:"out-in"}},[n("p",{key:t.up,staticClass:"alert-info"},[t._v("\n        "+t._s(t._f("capitalize")(t.up))+" "+t._s(t.$I18n.trans("system.since"))+" "+t._s(t.upSince)+"\n      ")])]),t._v(" "),n("p",{class:{"alert-success":t.nQueueWorkersRunning,"alert-danger":!t.nQueueWorkersRunning}},[t._v("\n      "+t._s(t.$I18n.trans("system.number_queue_workers"))+": "+t._s(t.nQueueWorkersRunning)+"\n    ")]),t._v(" "),n("div",{staticClass:"pt-4"},[n("table",{staticClass:"table table-sm"},[n("caption",[t._v(t._s(t.$I18n.trans("system.since")))]),t._v(" "),n("thead",[n("tr",[n("th",[t._v(t._s(t.$I18n.trans("system.used")))]),t._v(" "),n("th",[t._v(t._s(t.$I18n.trans("system.free")))]),t._v(" "),n("th",[t._v(t._s(t.$I18n.trans("system.total")))])])]),t._v(" "),n("tbody",[n("transition",{attrs:{name:"slide-fade",mode:"out-in"}},[n("tr",{key:t.memory.used},[n("td",[t.memory.used?t._e():n("span",[t._v(" ")]),t._v(t._s(t.memory.used))]),t._v(" "),n("td",[t._v(t._s(t.memory.free))]),t._v(" "),n("td",[t._v(t._s(t.memory.total))])])])],1)])])],1)])}),[],!1,null,"764fdbfc",null).exports,Inspire:Object(U.a)(Y,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"card mb-3 shadow"},[n("div",{staticClass:"card-body"},[n("small",[t.inspire?t._e():n("span",[t._v(t._s(t.fallback))]),t._v(t._s(t.inspire))])])])}),[],!1,null,null,null).exports},data:function(){return{}},beforeRouteLeave:function(t,e,n){clearInterval(this.$refs.refSystemInfo.systemInfoLoop),n()}},W=Object(U.a)(J,(function(){var t=this.$createElement,e=this._self._c||t;return e("div",{staticClass:"pt-4",attrs:{id:"view-dashboard"}},[e("div",{staticClass:"row justify-content-center"},[e("div",{staticClass:"col-sm-6 no-gutter-xs"},[e("inspire"),this._v(" "),e("system-info",{ref:"refSystemInfo"})],1),this._v(" "),e("div",{staticClass:"col-sm-6 no-gutter-xs"},[e("weather")],1)])])}),[],!1,null,null,null).exports,X={name:"Navigation",data:function(){return{notesRoute:Te,notesCreateRoute:$e}}},Z={name:"Notes",components:{Navigation:Object(U.a)(X,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"navigation"}},[n("div",{staticClass:"row justify-content-center top-submenu"},[n("div",{staticClass:"col-12 no-gutter-xs"},[n("div",{staticClass:"btn-group d-flex mb-2"},[n("router-link",{staticClass:"btn btn-group-sm w-100 inner-nav btn-secondary",attrs:{to:t.notesRoute}},[t._v(t._s(t.$I18n.trans("notes.notes")))]),t._v(" "),n("router-link",{staticClass:"btn btn-group-sm w-100 inner-nav btn-secondary",attrs:{to:t.notesCreateRoute}},[t._v(t._s(t.$I18n.trans("notes.new_note")))])],1)])])])}),[],!1,null,null,null).exports}},K=Object(U.a)(Z,(function(){var t=this.$createElement,e=this._self._c||t;return e("div",{attrs:{id:"notes"}},[e("navigation"),this._v(" "),e("router-view")],1)}),[],!1,null,null,null).exports,G=n("9/5/"),tt=n.n(G);function et(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(t);e&&(r=r.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,r)}return n}function nt(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?et(Object(n),!0).forEach((function(e){rt(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):et(Object(n)).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}function rt(t,e,n){return e in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}var st={name:"AlgoliaSearch",computed:{searchQuery:{get:function(){return this.$store.state.notes.searchQuery},set:function(t){this.$store.commit("setSearchQuery",t)}}},watch:{searchQuery:function(){(0===this.searchQuery.length||this.searchQuery.length>1)&&this.submit()}},methods:nt(nt({},Object(a.b)(["fetchNotesList"])),{},{submit:tt()((function(){this.fetchNotesList(this.searchQuery)}),400),clearSearch:function(){this.searchQuery="",this.submit()}})},at=(n("mgPi"),Object(U.a)(st,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"col-12 no-gutter-xs",attrs:{id:"algolia-search"}},[n("div",{staticClass:"justify-content-center form-inline d-flex"},[n("label",{staticClass:"sr-only",attrs:{for:"query"}},[t._v(t._s(t.$I18n.trans("notes.search")))]),t._v(" "),n("input",{directives:[{name:"model",rawName:"v-model",value:t.searchQuery,expression:"searchQuery"}],staticClass:"form-control flex-grow-1",attrs:{type:"text",id:"query",name:"query",placeholder:t.$I18n.trans("notes.min_2_chars")},domProps:{value:t.searchQuery},on:{keyup:function(e){return!e.type.indexOf("key")&&t._k(e.keyCode,"esc",27,e.key,["Esc","Escape"])?null:t.clearSearch(e)},input:function(e){e.target.composing||(t.searchQuery=e.target.value)}}}),t._v("\n     \n    "),t._m(0)])])}),[function(){var t=this.$createElement,e=this._self._c||t;return e("picture",[e("source",{attrs:{srcset:"search-by-algolia.svg",media:"(min-width: 800px)"}}),this._v(" "),e("img",{staticStyle:{"max-height":"30px"},attrs:{src:"algolia-blue-mark.svg",alt:"search by algolia"}})])}],!1,null,"45329d92",null).exports);function ot(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(t);e&&(r=r.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,r)}return n}function it(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?ot(Object(n),!0).forEach((function(e){ct(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):ot(Object(n)).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}function ct(t,e,n){return e in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}var ut={name:"NotesList",data:function(){return{notesShowRoute:Ee,loading:!0}},computed:Object(a.c)(["notesList"]),methods:it(it({},Object(a.b)(["fetchNotesList","updateSearchQuery"])),{},{close:function(){var t=this;this.updateSearchQuery(""),this.loading=!0,this.fetchNotesList().then((function(e){t.loading=!1}))}}),created:function(){var t=this;this.fetchNotesList().then((function(e){t.loading=!1}))}},lt={name:"NotesListing",components:{AlgoliaSearch:at,NotesList:Object(U.a)(ut,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"notes-list"}},[n("ul",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}],staticClass:"list-group list-group-flush"},[t.notesList.length?t._e():n("el-alert",{attrs:{title:t.$I18n.trans("notes.alert_failed_to_find_records"),type:"info",center:"","close-text":t.$I18n.trans("notes.alert_close_text")},on:{close:t.close}}),t._v(" "),t._l(t.notesList,(function(e){return n("li",{staticClass:"list-group-item list-group-item-action p-2 p-sm-3"},[n("router-link",{staticClass:"text-secondary",attrs:{to:{name:t.notesShowRoute.name,params:{id:e.id}}}},[n("div",{staticClass:"d-flex w-100 justify-content-between"},[n("h5",{staticClass:"mb-1"},[t._v(t._s(e.title))]),t._v(" "),n("small",{staticClass:"d-none d-sm-block"},[t._v(t._s(t._f("diffForHumans")(e.updated_at)))])])]),t._v(" "),n("div",t._l(e.tags,(function(e){return n("el-tag",{key:e.id,staticClass:"mr-1 mb-1",attrs:{size:"small"}},[t._v(t._s(e.name))])})),1)],1)}))],2)])}),[],!1,null,null,null).exports}},dt=Object(U.a)(lt,(function(){var t=this.$createElement,e=this._self._c||t;return e("div",{attrs:{id:"notes-listing"}},[e("div",{staticClass:"row justify-content-center"},[e("algolia-search")],1),this._v(" "),e("div",{staticClass:"row"},[e("div",{staticClass:"col-12 mt-3 no-gutter-xs"},[e("notes-list")],1)])])}),[],!1,null,null,null).exports,mt={methods:{textareaCharInserter:function(t){var e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:"  ";t.preventDefault();var n=this.note.body,r=t.target.selectionStart,s=n.slice(0,r),a=n.slice(r);this.note.body="".concat(s).concat(e).concat(a),t.target.value=this.note.body,t.target.selectionEnd=t.target.selectionStart=r+1}}};function pt(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}function ft(t){return function(){var e=this,n=arguments;return new Promise((function(r,s){var a=t.apply(e,n);function o(t){pt(a,r,s,o,i,"next",t)}function i(t){pt(a,r,s,o,i,"throw",t)}o(void 0)}))}}function vt(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(t);e&&(r=r.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,r)}return n}function ht(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?vt(Object(n),!0).forEach((function(e){bt(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):vt(Object(n)).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}function bt(t,e,n){return e in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}var gt={name:"NotesCreate",mixins:[mt],data:function(){return{newTagInputVisible:!1,newTag:"",errors:[],note:{body:""},tags:[],selectedTags:[],inputEl:null}},methods:ht(ht({},Object(a.b)(["addNote"])),{},{insertCodeBlockBackTicks:function(){if(this.inputEl||(this.inputEl=document.querySelector("#note-create textarea")),!document.execCommand("insertText",!1,"```")&&"function"==typeof this.inputEl.setRangeText){var t=this.inputEl.selectionStart;this.inputEl.setRangeText("```"),this.inputEl.selectionStart=this.inputEl.selectionEnd=t+"```".length;var e=document.createEvent("UIEvent");e.initEvent("input",!0,!1),this.inputEl.dispatchEvent(e)}},addNewTag:function(){var t=this;return ft(i.a.mark((function e(){var n,r,s,a;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:if(!t.newTag){e.next=15;break}return t.errors=[],e.prev=2,e.next=5,u.a.post("/api/notetagcreate",{tag:t.newTag});case 5:n=e.sent,r=d()(n,"data.data.success"),s=d()(n,"data.data.id"),r&&s&&(t.tags.push({id:s,name:t.newTag}),t.newTagInputVisible=!1,t.newTag=""),e.next=15;break;case 11:e.prev=11,e.t0=e.catch(2),a=d()(e.t0,"response.data.errors.tag[0]"),t.errors.push({field:"tagCreate",text:a});case 15:case"end":return e.stop()}}),e,null,[[2,11]])})))()},showInput:function(){var t=this;this.newTagInputVisible=!0,this.$nextTick((function(e){t.$refs.saveTagInput.$refs.input.focus()}))},fetchTags:function(){var t=this;return ft(i.a.mark((function e(){var n;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,u.a.get("/api/notes/tags");case 2:n=e.sent,t.tags=d()(n,"data.data",[]);case 4:case"end":return e.stop()}}),e)})))()},submitForm:function(t){var e=this;return ft(i.a.mark((function n(){var r;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:if(t.preventDefault(),!e.newTagInputVisible){n.next=3;break}return n.abrupt("return",!1);case 3:if(e.errors=[],e.note.body){n.next=8;break}return e.errors.push({field:"body",text:"Note text is empty."}),e.note.body=null,n.abrupt("return");case 8:return n.prev=8,n.next=11,e.addNote({body:e.note.body,tags:e.selectedTags});case 11:(r=n.sent)?(e.note.body="",e.selectedTags=[],e.$notify.success(e.$I18n.trans("notes.confirmation_success_note_create")),setTimeout((function(){return e.$router.push(Object.assign({},Ee,{params:{id:r}}))}),400)):e.errors.push({field:"na",text:"Failed to create the note."}),n.next=18;break;case 15:n.prev=15,n.t0=n.catch(8),e.errors.push({field:"submit",text:n.t0.message});case 18:case"end":return n.stop()}}),n,null,[[8,15]])})))()}}),created:function(){this.fetchTags()}},yt=(n("1XD5"),Object(U.a)(gt,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"notes-create"}},[n("div",{staticClass:"row justify-content-center"},[n("div",{staticClass:"col-sm no-gutter-xs"},[n("div",{staticClass:"card shadow"},[t.errors.length?n("div",{staticClass:"p-1"},t._l(t.errors,(function(t){return n("el-alert",{key:t.field,attrs:{title:t.text,type:"error"}})})),1):t._e(),t._v(" "),n("form",{attrs:{method:"post"},on:{submit:t.submitForm}},[n("div",{staticClass:"textarea-container"},[n("textarea",{directives:[{name:"model",rawName:"v-model",value:t.note.body,expression:"note.body"}],staticClass:"form-control form-text m-0",attrs:{placeholder:t.$I18n.trans("notes.placeholder_note_title"),name:"body",cols:"100",rows:"15"},domProps:{value:t.note.body},on:{keydown:function(e){return!e.type.indexOf("key")&&t._k(e.keyCode,"tab",9,e.key,"Tab")?null:(e.preventDefault(),t.textareaCharInserter(e))},input:function(e){e.target.composing||t.$set(t.note,"body",e.target.value)}}})]),t._v(" "),n("div",{staticClass:"d-flex justify-content-between"},[n("small",{staticClass:"form-text text-muted ml-3"},[t._m(0)]),t._v(" "),n("el-button",{attrs:{circle:"",size:"small",type:"info"},on:{click:t.insertCodeBlockBackTicks}},[t._v("```")])],1),t._v(" "),n("div",{staticClass:"mt-3 ml-3"},[n("p",[t._v(t._s(t.$I18n.trans("notes.tags"))+":")]),t._v(" "),n("el-checkbox-group",{staticClass:"d-inline-block",attrs:{size:"small"},model:{value:t.selectedTags,callback:function(e){t.selectedTags=e},expression:"selectedTags"}},t._l(t.tags,(function(e){return n("el-checkbox-button",{key:e.id,staticClass:"mr-1",attrs:{label:e.id}},[t._v(t._s(e.name))])})),1),t._v(" "),n("div",{staticClass:"d-inline-block new-tag-group"},[t.newTagInputVisible?n("el-input",{ref:"saveTagInput",staticClass:"input-new-tag",attrs:{size:"small"},on:{blur:t.addNewTag},nativeOn:{keyup:function(e){return!e.type.indexOf("key")&&t._k(e.keyCode,"enter",13,e.key,"Enter")?null:t.addNewTag(e)}},model:{value:t.newTag,callback:function(e){t.newTag=e},expression:"newTag"}}):n("el-button",{staticClass:"button-new-tag",attrs:{type:"info",plain:"",size:"small"},on:{click:t.showInput}},[t._v(t._s(t.$I18n.trans("notes.new_tag")))])],1)],1),t._v(" "),n("div",{staticClass:"px-3 pb-3 pt-3"},[n("button",{staticClass:"btn btn-primary",attrs:{type:"submit"}},[t._v(t._s(t.$I18n.trans("common.add")))])])])])])])])}),[function(){var t=this.$createElement;return(this._self._c||t)("a",{staticClass:"text-muted",attrs:{href:"https://commonmark.org/help/",tabindex:"-1",target:"_CommonMark"}},[this._v(this._s(this.$I18n.trans("notes.commonmark")))])}],!1,null,"21a1e722",null).exports);n("WxZA");function xt(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}function wt(t){return function(){var e=this,n=arguments;return new Promise((function(r,s){var a=t.apply(e,n);function o(t){xt(a,r,s,o,i,"next",t)}function i(t){xt(a,r,s,o,i,"throw",t)}o(void 0)}))}}var _t={name:"NotesShow",props:["id"],data:function(){return{loading:!0,note:{},notesUpdateRoute:Pe}},created:function(){this.fetchNote(this.id)},methods:{fetchNote:function(t){var e=this;return wt(i.a.mark((function n(){var r;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:if(t){n.next=2;break}return n.abrupt("return",!1);case 2:return n.prev=2,n.next=5,u.a.get("/api/notes/".concat(t,"?html=1"));case 5:r=n.sent,n.next=14;break;case 8:if(n.prev=8,n.t0=n.catch(2),404!==n.t0.response.status){n.next=14;break}return n.next=13,e.$router.push(Object.assign({},ze,{params:{message:"Note could not be found.",linksTo:[{route:Te,linkText:"Notes"}]}}));case 13:return n.abrupt("return");case 14:e.note=d()(r,"data.data"),setTimeout((function(){return e.loading=!1}),400);case 16:case"end":return n.stop()}}),n,null,[[2,8]])})))()},confirmDelete:function(t){var e=this;return wt(i.a.mark((function n(){var r,s,a,o;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:r=e.$I18n.trans("notes.confirmation_ask_delete",{id:t}),s=e.$I18n.trans("notes.confirmation_success_deleted",{id:t}),a=e.$I18n.trans("notes.delete"),o=e.$I18n.trans("notes.cancel"),e.$confirm(r,"Warning",{confirmButtonText:a,cancelButtonText:o,type:"warning",center:!0}).then((function(){e.deleteNote(t).then((function(t){t&&(e.$notify.success(s),setTimeout((function(){return e.$router.push(Te)}),400))})).catch((function(n){e.$notify.warning(e.$I18n.trans("notes.failed_delete",{id:t}))}))})).catch((function(){e.$notify.info(e.$I18n.trans("notes.delete_canceled",{id:t}))}));case 5:case"end":return n.stop()}}),n)})))()},deleteNote:function(t){return wt(i.a.mark((function e(){var n,r;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,u.a.delete("/api/notedestroy/".concat(t),{_method:"delete"});case 2:if(n=e.sent,!(r=d()(n,"data.data.success",!1))){e.next=6;break}return e.abrupt("return",r);case 6:throw"Delete failed";case 7:case"end":return e.stop()}}),e)})))()}}},kt=(n("nNO0"),Object(U.a)(_t,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"note"},[n("div",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}],staticClass:"row justify-content-center"},[n("div",{staticClass:"col-12 no-gutter-xs"},[t.note.id?t._e():n("div",{staticClass:"mt-3 "},[n("p",{staticClass:"loader-text pb-2 pt-3"},[t._v(" ")]),t._v(" "),n("p",{staticClass:"loader-text"},[t._v(" ")]),t._v(" "),n("p",{staticClass:"loader-text"},[t._v(" ")]),t._v(" "),n("p",{staticClass:"loader-text loader-text--1x3 d-inline-block"},[t._v("   ")])]),t._v(" "),n("div",{staticClass:"mt-3",domProps:{innerHTML:t._s(t.note.body)}}),t._v(" "),n("div",t._l(t.note.tags,(function(e){return n("el-tag",{key:e.id,staticClass:"mr-1 mb-1",attrs:{size:"big"}},[t._v(t._s(e.name))])})),1)])]),t._v(" "),n("div",{staticClass:"row"},[n("div",{staticClass:"col-12"},[n("div",{staticClass:"mt-3"},[t.note.id?n("router-link",{staticClass:"btn btn-primary",attrs:{to:Object.assign({},t.notesUpdateRoute,{params:{id:t.note.id}})}},[t._v(t._s(t.$I18n.trans("notes.edit")))]):t._e(),t._v(" "),n("button",{staticClass:"btn btn-danger ml-3",on:{click:function(e){return t.confirmDelete(t.note.id)}}},[t._v(t._s(t.$I18n.trans("notes.delete")))])],1)])])])}),[],!1,null,"7a7ca61a",null).exports);function jt(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}function Ct(t){return function(){var e=this,n=arguments;return new Promise((function(r,s){var a=t.apply(e,n);function o(t){jt(a,r,s,o,i,"next",t)}function i(t){jt(a,r,s,o,i,"throw",t)}o(void 0)}))}}var Ot={name:"NotesUpdate",mixins:[mt],props:["id"],data:function(){return{loading:!0,newTagInputVisible:!1,newTag:"",errors:[],note:{},tags:[{id:1,name:""}],selectedTags:[],inputEl:null}},methods:{insertCodeBlockBackTicks:function(){if(this.inputEl||(this.inputEl=document.querySelector("#note-update textarea")),!document.execCommand("insertText",!1,"```")&&"function"==typeof this.inputEl.setRangeText){var t=this.inputEl.selectionStart;this.inputEl.setRangeText("```"),this.inputEl.selectionStart=this.inputEl.selectionEnd=t+"```".length;var e=document.createEvent("UIEvent");e.initEvent("input",!0,!1),this.inputEl.dispatchEvent(e)}},addNewTag:function(){var t=this;return Ct(i.a.mark((function e(){var n,r,s,a;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:if(!t.newTag){e.next=15;break}return t.errors=[],e.prev=2,e.next=5,u.a.post("/api/notetagcreate",{tag:t.newTag});case 5:n=e.sent,r=d()(n,"data.data.success"),s=d()(n,"data.data.id"),r&&s&&(t.tags.push({id:s,name:t.newTag}),t.newTagInputVisible=!1,t.newTag=""),e.next=15;break;case 11:e.prev=11,e.t0=e.catch(2),a=d()(e.t0,"response.data.errors.tag[0]"),t.errors.push({field:"tagCreate",text:a});case 15:case"end":return e.stop()}}),e,null,[[2,11]])})))()},showInput:function(){var t=this;this.newTagInputVisible=!0,this.$nextTick((function(e){t.$refs.saveTagInput.$refs.input.focus()}))},fetchNote:function(t){var e=this;return Ct(i.a.mark((function n(){var r;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:if(t){n.next=2;break}return n.abrupt("return",!1);case 2:return n.next=4,u.a.get("/api/notes/".concat(t,"?html=0"));case 4:r=n.sent,e.note=d()(r,"data.data",[]),e.selectedTags=e.note.tags.map((function(t){return t.id})),setTimeout((function(){return e.loading=!1}),400);case 8:case"end":return n.stop()}}),n)})))()},fetchTags:function(){var t=this;return Ct(i.a.mark((function e(){var n;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,u.a.get("/api/notes/tags");case 2:n=e.sent,t.tags=d()(n,"data.data",[]);case 4:case"end":return e.stop()}}),e)})))()},submitForm:function(t){var e=this;return Ct(i.a.mark((function n(){var r;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:if(t.preventDefault(),!e.newTagInputVisible){n.next=3;break}return n.abrupt("return",!1);case 3:if(e.errors=[],e.note.body){n.next=7;break}return e.errors.push({field:"body",text:"Note text is empty."}),n.abrupt("return");case 7:return n.prev=7,n.next=10,u.a.put("/api/noteupdate/".concat(e.note.id),{body:e.note.body,tags:e.selectedTags,_method:"put"});case 10:r=n.sent,d()(r,"data.data.success")?(e.selectedTags=[],e.$notify.success(e.$I18n.trans("notes.confirmation_success_note_updated")),setTimeout((function(){return e.$router.push(Object.assign({},Ee,{params:{id:e.note.id}}))}),400)):e.errors.push({field:"na",text:"Failed to update note."}),n.next=17;break;case 14:n.prev=14,n.t0=n.catch(7),e.errors.push({field:"submit",text:n.t0.message});case 17:case"end":return n.stop()}}),n,null,[[7,14]])})))()}},created:function(){var t=this;this.fetchTags().then((function(){t.fetchNote(t.id)}))}},It=(n("iqeS"),Object(U.a)(Ot,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"notes-update"}},[n("div",{staticClass:"row justify-content-center"},[n("div",{staticClass:"col-sm no-gutter-xs"},[n("div",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}],staticClass:"card shadow"},[t.errors.length?n("div",{staticClass:"p-1"},t._l(t.errors,(function(t){return n("el-alert",{key:t.field,attrs:{title:t.text,type:"error"}})})),1):t._e(),t._v(" "),n("form",{attrs:{method:"post"},on:{submit:t.submitForm}},[n("div",{staticClass:"textarea-container"},[n("textarea",{directives:[{name:"model",rawName:"v-model",value:t.note.body,expression:"note.body"}],staticClass:"form-control form-text m-0",attrs:{placeholder:t.$I18n.trans("notes.placeholder_note_title"),name:"body",cols:"100",rows:"15"},domProps:{value:t.note.body},on:{keydown:function(e){return!e.type.indexOf("key")&&t._k(e.keyCode,"tab",9,e.key,"Tab")?null:(e.preventDefault(),t.textareaCharInserter(e))},input:function(e){e.target.composing||t.$set(t.note,"body",e.target.value)}}})]),t._v(" "),n("div",{staticClass:"d-flex justify-content-between"},[n("small",{staticClass:"form-text text-muted ml-3"},[t._m(0)]),t._v(" "),n("el-button",{attrs:{circle:"",size:"small",type:"info"},on:{click:t.insertCodeBlockBackTicks}},[t._v("```")])],1),t._v(" "),n("div",{staticClass:"mt-3 ml-3"},[n("p",[t._v("Tags:")]),t._v(" "),n("el-checkbox-group",{staticClass:"d-inline-block",attrs:{size:"small"},model:{value:t.selectedTags,callback:function(e){t.selectedTags=e},expression:"selectedTags"}},t._l(t.tags,(function(e){return n("el-checkbox-button",{key:e.id,staticClass:"mr-1",attrs:{label:e.id}},[t._v(t._s(e.name))])})),1),t._v(" "),n("div",{staticClass:"d-inline-block new-tag-group"},[t.newTagInputVisible?n("el-input",{ref:"saveTagInput",staticClass:"input-new-tag",attrs:{size:"small"},on:{blur:t.addNewTag},nativeOn:{keyup:function(e){return!e.type.indexOf("key")&&t._k(e.keyCode,"enter",13,e.key,"Enter")?null:t.addNewTag(e)}},model:{value:t.newTag,callback:function(e){t.newTag=e},expression:"newTag"}}):n("el-button",{staticClass:"button-new-tag",attrs:{type:"info",plain:"",size:"small"},on:{click:t.showInput}},[t._v(t._s(t.$I18n.trans("notes.new_tag")))])],1)],1),t._v(" "),n("div",{staticClass:"px-3 pb-3 pt-3"},[n("button",{staticClass:"btn btn-primary",attrs:{type:"submit"}},[t._v(t._s(t.$I18n.trans("common.update")))])])])])])])])}),[function(){var t=this.$createElement;return(this._self._c||t)("a",{staticClass:"text-muted",attrs:{href:"https://commonmark.org/help/",tabindex:"-1",target:"_CommonMark"}},[this._v(this._s(this.$I18n.trans("notes.commonmark")))])}],!1,null,"401ef6a5",null).exports),$t={name:"Navigation",props:{numberOfBookmarks:{type:null|Number,required:!0}},data:function(){return{hackerNewsTopPostsRoute:Se,hackerNewsBestPostsRoute:Ne,hackerNewsBookmarkedPostsRoute:De}}},Et=Object(U.a)($t,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"navigation"}},[n("div",{staticClass:"row justify-content-center top-submenu"},[n("div",{staticClass:"col-12 no-gutter-xs"},[n("div",{staticClass:"btn-group d-flex mb-2"},[n("router-link",{staticClass:"btn btn-group-sm w-100 inner-nav btn-secondary",attrs:{to:t.hackerNewsTopPostsRoute}},[t._v(t._s(t.$I18n.trans("hackernews.top")))]),t._v(" "),n("router-link",{staticClass:"btn btn-group-sm w-100 inner-nav btn-secondary",attrs:{to:t.hackerNewsBestPostsRoute}},[t._v(t._s(t.$I18n.trans("hackernews.best")))]),t._v(" "),n("router-link",{staticClass:"btn btn-group-sm w-100 inner-nav btn-secondary",attrs:{to:t.hackerNewsBookmarkedPostsRoute}},[t._v(t._s(t.$I18n.trans("hackernews.bookmarks"))+" "),n("span",{staticClass:"badge badge-light"},[t._v(t._s(t.numberOfBookmarks))])])],1)])])])}),[],!1,null,null,null).exports,Pt={name:"HackerNews",components:{Navigation:Et},data:function(){return{numberOfBookmarks:null}},methods:{updateNumberOfBookmarks:function(t){this.numberOfBookmarks=t}}},Tt=Object(U.a)(Pt,(function(){var t=this.$createElement,e=this._self._c||t;return e("div",{attrs:{id:"hacker-news"}},[e("navigation",{attrs:{"number-of-bookmarks":this.numberOfBookmarks}}),this._v(" "),e("router-view",{on:{nBookmarksChangedEvent:this.updateNumberOfBookmarks}})],1)}),[],!1,null,null,null).exports;function St(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}var Nt={data:function(){return{nBookmarks:null,notification:null}},methods:{bookmarkPost:function(t){var e,n=this;return(e=i.a.mark((function e(){var r,s,a,o;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return t.status.saving=!0,r={postId:t.id},(s=!t.status.bookmarked)||(r._method="delete"),e.next=6,u.a.post("/api/bookmarks",r);case 6:a=e.sent,(o=d()(a,"data.data.success",!1))&&(t.status.bookmarked=!t.status.bookmarked,n.nBookmarks+=s?1:-1),n.notifyUserOfBookmarkStatusChange(o,s,t);case 10:case"end":return e.stop()}}),e)})),function(){var t=this,n=arguments;return new Promise((function(r,s){var a=e.apply(t,n);function o(t){St(a,r,s,o,i,"next",t)}function i(t){St(a,r,s,o,i,"throw",t)}o(void 0)}))})()},notifyUserOfBookmarkStatusChange:function(t,e,n){var r={message:this.$I18n.trans("hackernews.add_failure"),type:"error"};t&&(r=e?{message:this.$I18n.trans("hackernews.added_to_bookmarks",{title:n.title}),type:"success"}:{message:this.$I18n.trans("hackernews.remove_from_bookmarks",{title:n.title}),type:"warning"}),this.notification&&!this.notification.closed&&this.notification.close(),this.notification=this.$notify(r),setTimeout((function(){return n.status.saving=!1}),400)}}},Dt=n("Wcq6"),Ut=n.n(Dt),Ft=(n("Zs65"),Ut.a.initializeApp({databaseURL:"hacker-news.firebaseio.com"}).database().ref("/v0"));function Bt(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}function At(t){return function(t){if(Array.isArray(t))return Rt(t)}(t)||function(t){if("undefined"!=typeof Symbol&&Symbol.iterator in Object(t))return Array.from(t)}(t)||function(t,e){if(!t)return;if("string"==typeof t)return Rt(t,e);var n=Object.prototype.toString.call(t).slice(8,-1);"Object"===n&&t.constructor&&(n=t.constructor.name);if("Map"===n||"Set"===n)return Array.from(t);if("Arguments"===n||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n))return Rt(t,e)}(t)||function(){throw new TypeError("Invalid attempt to spread non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}()}function Rt(t,e){(null==e||e>t.length)&&(e=t.length);for(var n=0,r=new Array(e);n<e;n++)r[n]=t[n];return r}var zt={name:"HNPosts",mixins:[Nt],props:{idList:{type:Array,default:[],required:!0}},data:function(){return{loading:!0,posts:[],nBookmarks:null,firstBatchPosition:10,hackerNewsPostRoute:Ue}},methods:{fetchItems:function(t){var e=this,n=this.firstBatchPosition;t.length<n&&(n=Math.floor(t.length/2));var r=t.slice(0,n),s=t.slice(n);Promise.all(r.map((function(t){return e.fetchItem(t)}))).then((function(t){e.posts=t.filter(e.isStoryValid),e.attachBookmarked(),e.loading=!1})),this.$nextTick((function(){Promise.all(s.map((function(t){return e.fetchItem(t)}))).then((function(t){var n;(n=e.posts).push.apply(n,At(t.filter(e.isStoryValid))),e.attachBookmarked()}))}))},isStoryValid:function(t){return"story"===t.type&&!!t.id},fetchItem:function(t){return new Promise((function(e,n){Ft.child("item/".concat(t)).on("value",(function(t){var n=t.val(),r={id:d()(n,"id"),title:d()(n,"title"),score:d()(n,"score"),time:z.a.unix(d()(n,"time",0)).format("YYYY-MM-DD HH:mm:ss"),url:d()(n,"url"),type:d()(n,"type"),nComments:d()(n,"descendants"),kids:d()(n,"kids",[]),status:{bookmarked:!1,saving:!1}};e(r)})),Ft.onDisconnect((function(t,e){this.$alert("Firebase connection lost","Lost connection",{confirmButtonText:"OK"}),console.log(t,e)}))}))},attachBookmarked:function(){var t,e=this;return(t=i.a.mark((function t(){var n,r;return i.a.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.next=2,u.a.get("/api/bookmarks");case 2:n=t.sent,r=d()(n,"data.data",[]),e.nBookmarks=r.length,e.posts=e.posts.map((function(t){return t.status.bookmarked=r.includes(t.id),t}));case 6:case"end":return t.stop()}}),t)})),function(){var e=this,n=arguments;return new Promise((function(r,s){var a=t.apply(e,n);function o(t){Bt(a,r,s,o,i,"next",t)}function i(t){Bt(a,r,s,o,i,"throw",t)}o(void 0)}))})()}},watch:{idList:function(t,e){var n=this;this.loading=!0,setTimeout((function(){return n.fetchItems(t)}),400)},nBookmarks:function(t,e){this.$emit("nBookmarksChangedEvent",t)}}};n("0P3G");function Lt(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}var Mt={name:"HackerNewsList",components:{Navigation:Et,HnPosts:Object(U.a)(zt,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("ul",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}],staticClass:"list-group",attrs:{id:"hn-posts"}},[t._l(Array(t.firstBatchPosition).fill(null),(function(e){return t.loading?n("li",{staticClass:"list-group-item list-group-item-action flex-column align-items-start"},[t._m(0,!0),t._v(" "),t._m(1,!0)]):t._e()})),t._v(" "),t._l(t.posts,(function(e,r){return n("li",{key:e.id,staticClass:"list-group-item list-group-item-action flex-column align-items-start",attrs:{"data-index":r+1}},[n("div",{staticClass:"d-flex w-100 justify-content-between"},[n("span",[n("router-link",{staticClass:"text-body",attrs:{to:Object.assign({},t.hackerNewsPostRoute,{params:{id:e.id}})}},[t._v(t._s(e.title))]),t._v(" "),e.url?n("a",{staticClass:"text-body",attrs:{href:e.url,target:"_blank"}},[n("small",{staticClass:"text-muted"},[t._v("("+t._s(t._f("domainFromUrl")(e.url))+")")])]):t._e()],1),t._v(" "),n("span",{directives:[{name:"loading",rawName:"v-loading",value:e.status.saving,expression:"post.status.saving"}],staticClass:"d-block d-md-none"},[n("span",{staticClass:"bookmark-story pointer text-primary",on:{click:function(n){return t.bookmarkPost(e)}}},[t._v(t._s(e.status.bookmarked?"⚫":"⚪️")+"️")])]),t._v(" "),n("span",{staticClass:"badge d-none d-md-block",attrs:{title:e.time}},[t._v(t._s(t._f("diffForHumans")(e.time)))])]),t._v(" "),n("div",{directives:[{name:"loading",rawName:"v-loading",value:e.status.saving,expression:"post.status.saving"}],staticClass:"d-none d-md-block"},[n("small",{staticClass:"text-muted"},[t._v(t._s(t.$I18n.trans("hackernews.points",{points:e.score||""})))]),t._v("\n      |\n      "),n("small",[t._v(t._s(t.$I18n.trans("hackernews.comments",{comments:e.nComments||""})))]),t._v("\n      |\n      "),n("span",{staticClass:"bookmark-story pointer text-primary",on:{click:function(n){return t.bookmarkPost(e)}}},[t._v(t._s(e.status.bookmarked?"⚫":"⚪️"))])])])}))],2)}),[function(){var t=this.$createElement,e=this._self._c||t;return e("div",{staticClass:"d-flex w-100 justify-content-between"},[e("span",{staticClass:"loader-text loader-text-top d-block w-100"},[this._v(" ")])])},function(){var t=this.$createElement,e=this._self._c||t;return e("div",{staticClass:"d-none d-md-block"},[e("span",{staticClass:"loader-text loader-text-bottom loader-text--1x3 d-inline-block"},[e("small",[this._v(" ")])])])}],!1,null,"a624b0ec",null).exports},props:{type:{type:String,validator:function(t){return-1!==["top","best","bookmarks"].indexOf(t)},default:"top",required:!0}},data:function(){return{idList:[],limit:100,types:{HackerNewsBookmarks:"bookmarks"},numberOfBookmarks:null}},methods:{fetchStories:function(){this.type===this.types.HackerNewsBookmarks?this.fetchIdsBookmarkedFromBackend():this.fetchIdsFromFirebase(this.type)},fetchIdsFromFirebase:function(t){var e=this;Ft.child("".concat(t,"stories")).limitToFirst(this.limit).once("value",(function(t,n){var r=t.val();e.idList=Array.isArray(r)?r:[]}))},fetchIdsBookmarkedFromBackend:function(){var t,e=this;return(t=i.a.mark((function t(){var n;return i.a.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.next=2,u.a.get("/api/bookmarks");case 2:n=t.sent,e.idList=d()(n,"data.data",[]);case 4:case"end":return t.stop()}}),t)})),function(){var e=this,n=arguments;return new Promise((function(r,s){var a=t.apply(e,n);function o(t){Lt(a,r,s,o,i,"next",t)}function i(t){Lt(a,r,s,o,i,"throw",t)}o(void 0)}))})()},updateNumberOfBookmarks:function(t){this.numberOfBookmarks=t}},created:function(){this.fetchStories()},watch:{$route:function(t,e){this.fetchStories()}}},qt=Object(U.a)(Mt,(function(){var t=this.$createElement,e=this._self._c||t;return e("div",{attrs:{id:"hacker-news-list"}},[e("div",{staticClass:"row"},[e("div",{staticClass:"col-12 no-gutter-xs"},[e("hn-posts",this._g({attrs:{"id-list":this.idList}},this.$listeners))],1)])])}),[],!1,null,null,null).exports,Qt={name:"HnComment",props:{comment:{type:Object,required:!0},handleCollapseToggle:{type:Function,required:!0}},data:function(){return{isShow:!0}}};n("P9EP");function Ht(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}var Vt={name:"HackerNewsPost",mixins:[Nt],components:{Navigation:Et,HnComment:Object(U.a)(Qt,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"hn-comment card mt-3"},[n("div",{staticClass:"card-body px-2 px-md-3 shadow"},[n("h6",{staticClass:"card-subtitle bg-light text-muted d-flex justify-content-between align-items-center",class:{"opacity-5 text-line-through":t.comment.deleted}},[t.comment.id?n("span",{staticClass:"align-middle"},[t._v(t._s(t.$I18n.trans("hackernews.by",{by:t.comment.by||""})))]):t._e(),t._v(" "),n("small",{staticClass:"text-muted",attrs:{title:t.comment.time}},[t._v(t._s(t._f("diffForHumans")(t.comment.time)))]),t._v(" "),t.comment.id?n("a",{staticClass:"btn btn-sm pointer btn-collapse",attrs:{role:"button","data-toggle":"collapse"},on:{click:function(e){return t.handleCollapseToggle(t.comment)}}},[n("b",[t._v("["),t.comment.collapsed?n("span",[t._v(t._s(t.comment.kids.length))]):n("span",[t._v("-")]),t._v("]\n        ")])]):t._e()]),t._v(" "),n("transition",{attrs:{name:"slide"}},[t.comment.collapsed?t._e():n("div",{staticClass:"mt-2"},[n("p",{domProps:{innerHTML:t._s(t.comment.text)}}),t._v(" "),t._l(t.comment.comments,(function(e){return t.comment.comments.length?n("hn-comment",{key:e.id,attrs:{comment:e,"handle-collapse-toggle":t.handleCollapseToggle}}):t._e()}))],2)])],1)])}),[],!1,null,"1a92697e",null).exports},props:{id:{type:Number|String,required:!0}},data:function(){return{loading:!0,loadingComments:!0,collapsedCommentsKey:"collapsedCommentsHackerNewsPost",collapsedComments:[],post:{id:null,by:null,title:null,text:null,score:null,time:null,url:null,type:null,nComments:null,status:{bookmarked:null,saving:null},kids:[],comments:[{id:null,by:null,kids:[],parent:null,text:null,time:null,comments:[]}]},numberOfBookmarks:null}},methods:{fetchPost:function(t){var e=this;t&&(this.fetchCollapsedComments(),Ft.child("item/".concat(t)).once("value",(function(t){var n=t.val();e.post.id=d()(n,"id"),e.post.by=d()(n,"by"),e.post.title=d()(n,"title"),e.post.text=d()(n,"text"),e.post.score=d()(n,"score"),e.post.time=z.a.unix(d()(n,"time",0)).format("YYYY-MM-DD HH:mm:ss"),e.post.url=d()(n,"url"),e.post.type=d()(n,"type"),e.post.nComments=d()(n,"descendants"),e.post.bookmarked=null,e.post.kids=d()(n,"kids",[]),e.post.comments=[],e.setBookmarkStatus(e.post),e.fetchComments(e.post),e.loading=!1})))},fetchComments:function(t){var e=this;t.kids.length&&t.kids.map((function(n){e.fetchComment(n).then((function(n){var r=d()(n,"kids",[]),s=d()(n,"id"),a={id:s,deleted:d()(n,"deleted"),by:d()(n,"by"),kids:r,text:d()(n,"text"),time:z.a.unix(d()(n,"time",0)).format("YYYY-MM-DD HH:mm:ss"),comments:[],collapsed:e.collapsedComments.includes(s)};e.fetchComments(a),t.comments.push(a),e.loadingComments=!1}))}))},fetchComment:function(t){return new Promise((function(e,n){Ft.child("item/".concat(t)).once("value",(function(t){var n=t.val();e(n)}))}))},setBookmarkStatus:function(t){var e,n=this;return(e=i.a.mark((function e(){var r,s;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,u.a.get("/api/bookmarks");case 2:r=e.sent,s=d()(r,"data.data",[]),n.numberOfBookmarks=s.length,t.status.bookmarked=s.includes(t.id);case 6:case"end":return e.stop()}}),e)})),function(){var t=this,n=arguments;return new Promise((function(r,s){var a=e.apply(t,n);function o(t){Ht(a,r,s,o,i,"next",t)}function i(t){Ht(a,r,s,o,i,"throw",t)}o(void 0)}))})()},fetchCollapsedComments:function(){var t,e=localStorage.getItem(this.collapsedCommentsKey);try{t=JSON.parse(e)}catch(t){console.error(t)}t=Array.isArray(t)?t:[],this.collapsedComments=t},setCollapsedComments:function(){var t=JSON.stringify(this.collapsedComments);localStorage.setItem(this.collapsedCommentsKey,t)},handleCollapseToggle:function(t){t.collapsed=!t.collapsed,t.collapsed?this.collapsedComments.push(t.id):this.collapsedComments=this.collapsedComments.filter((function(e){return e!==t.id})),this.setCollapsedComments()}},created:function(){this.fetchPost(this.id)}},Yt=(n("zTjN"),Object(U.a)(Vt,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"hacker-news-post"}},[n("div",{staticClass:"row"},[n("div",{staticClass:"col-12 no-gutter-xs"},[n("div",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}]},[t.loading?n("p",{staticClass:"lead loader-text d-block loader-text--2x3"},[t._v(" ")]):t._e(),t._v(" "),t.post.title?n("p",{staticClass:"lead"},[t._v("\n          "+t._s(t.post.title)+"\n          "),t.post.url?n("a",{staticClass:"text-body",attrs:{href:t.post.url,target:"_blank"}},[n("small",{staticClass:"text-muted"},[t._v("("+t._s(t._f("domainFromUrl")(t.post.url))+") [↗]")])]):t._e()]):t._e(),t._v(" "),t.post.text?n("p",{domProps:{innerHTML:t._s(t.post.text)}}):t._e(),t._v(" "),t.loading?n("div",{staticClass:"loader-text d-block loader-text--1x3"},[t._v(" ")]):t._e(),t._v(" "),t.post.title?n("div",{directives:[{name:"loading",rawName:"v-loading",value:t.post.status.saving,expression:"post.status.saving"}]},[n("small",{staticClass:"text-muted"},[t._v(t._s(t.$I18n.trans("hackernews.points",{points:t.post.score})))]),t._v(" "),n("span",{staticClass:"text-muted"},[t._v("|")]),t._v(" "),n("small",{staticClass:"text-muted"},[t._v(t._s(t.$I18n.trans("hackernews.comments",{comments:t.post.nComments})))]),t._v(" "),n("span",{staticClass:"text-muted"},[t._v("|")]),t._v(" "),n("small",{staticClass:"text-muted",attrs:{title:t.post.time}},[t._v(t._s(t._f("diffForHumans")(t.post.time)))]),t._v(" "),n("span",{staticClass:"text-muted"},[t._v("|")]),t._v(" "),n("small",{staticClass:"text-muted"},[t._v(t._s(t.$I18n.trans("hackernews.by",{by:t.post.by})))]),t._v(" "),n("span",{staticClass:"text-muted"},[t._v("|")]),t._v(" "),n("span",{staticClass:"bookmark-story pointer text-primary",on:{click:function(e){return t.bookmarkPost(t.post)}}},[t._v(t._s(t.post.status.bookmarked?"⚫":"⚪️")+"️")])]):t._e()])])]),t._v(" "),t.post.comments.length?n("div",{staticClass:"row"},[n("div",{staticClass:"col-12 no-gutter-xs"},[t.loadingComments?t._e():n("div",t._l(t.post.comments,(function(e){return n("hn-comment",{key:e.id,attrs:{comment:e,"handle-collapse-toggle":t.handleCollapseToggle}})})),1)])]):t._e()])}),[],!1,null,"cdd94d26",null).exports),Jt={name:"Navigation",data:function(){return{usersRoute:Ae,usersCreateRoute:Be}}},Wt={name:"Users",components:{Navigation:Object(U.a)(Jt,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"navigation"}},[n("div",{staticClass:"row justify-content-center top-submenu"},[n("div",{staticClass:"col-12 no-gutter-xs"},[n("div",{staticClass:"btn-group d-flex mb-2"},[n("router-link",{staticClass:"btn btn-group-sm w-100 inner-nav btn-secondary",attrs:{to:t.usersRoute}},[t._v(t._s(t.$I18n.trans("users.users")))]),t._v(" "),n("router-link",{staticClass:"btn btn-group-sm w-100 inner-nav btn-secondary",attrs:{to:t.usersCreateRoute}},[t._v(t._s(t.$I18n.trans("users.new-user")))])],1)])])])}),[],!1,null,null,null).exports}},Xt=Object(U.a)(Wt,(function(){var t=this.$createElement,e=this._self._c||t;return e("div",{attrs:{id:"users"}},[e("navigation"),this._v(" "),e("router-view")],1)}),[],!1,null,null,null).exports;function Zt(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}function Kt(t){return function(){var e=this,n=arguments;return new Promise((function(r,s){var a=t.apply(e,n);function o(t){Zt(a,r,s,o,i,"next",t)}function i(t){Zt(a,r,s,o,i,"throw",t)}o(void 0)}))}}function Gt(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(t);e&&(r=r.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,r)}return n}function te(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?Gt(Object(n),!0).forEach((function(e){ee(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):Gt(Object(n)).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}function ee(t,e,n){return e in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}var ne={name:"Users",data:function(){return{loading:!0,userToEdit:{},dialogVisible:!1,dialogFormRules:{name:[{required:!0,message:this.$I18n.trans("users.name_cannot_be_empty"),trigger:"blur"},{min:2,max:255,message:this.$I18n.trans("users.min_2_max_255"),trigger:"blur"}],password:[{min:12,max:100,message:this.$I18n.trans("users.please_set_valid_password"),trigger:"blur"}]}}},computed:Object(a.c)(["usersList","currentUserId"]),methods:te(te({},Object(a.b)(["fetchUsers","updateUser","destroyUser"])),{},{highlightCurrentUser:function(t){var e=t.row;t.rowIndex;return e.id===this.currentUserId?"table-success":null},formatDate:function(t){return this.$options.filters.diffForHumans(t)},formatUpdatedAt:function(t,e){return this.formatDate(t.updatedAt)},closeDialog:function(t){this.dialogVisible=!1,this.$refs[t].resetFields()},editDialogShow:function(t){this.userToEdit={id:t.id,name:t.name,email:t.email},this.dialogVisible=!0},editDialogSubmit:function(t){var e=this;return Kt(i.a.mark((function n(){var r,s;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:return n.next=2,e.validateDialogSubmit(t);case 2:if(!n.sent){n.next=10;break}return r={id:e.userToEdit.id,name:e.userToEdit.name},e.userToEdit.password&&(r.password=e.userToEdit.password),n.next=8,e.updateUser(r);case 8:s=n.sent,e.notifyActionSuccess(s);case 10:case"end":return n.stop()}}),n)})))()},validateDialogSubmit:function(t){var e=this;return Kt(i.a.mark((function n(){return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:return n.abrupt("return",new Promise((function(n,r){e.$refs[t].validate((function(t){return t?n(!0):r(!1)}))})));case 1:case"end":return n.stop()}}),n)})))()},notifyActionSuccess:function(t){var e=this,n=arguments.length>1&&void 0!==arguments[1]?arguments[1]:"users.user_updated",r=arguments.length>2&&void 0!==arguments[2]?arguments[2]:"users.failed_to_update";if(t){this.$notify.success(this.$I18n.trans(n));var s=this.usersList.find((function(t){return t.id===e.userToEdit.id}));s&&(s.name=this.userToEdit.name.trim()),this.dialogVisible=!1}else this.$notify.error(this.$I18n.trans(r))},deleteUser:function(){var t=this;return Kt(i.a.mark((function e(){var n;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,t.askUserToConfirmDelete();case 2:if(!e.sent){e.next=8;break}return e.next=6,t.destroyUser(t.userToEdit.id);case 6:n=e.sent,t.notifyActionSuccess(n,"users.delete_success","users.failed_to_delete");case 8:case"end":return e.stop()}}),e)})))()},askUserToConfirmDelete:function(){var t=this;return Kt(i.a.mark((function e(){var n;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.prev=0,e.next=3,t.$confirm(t.$I18n.trans("users.confirm_delete_request",{name:t.userToEdit.name}),t.$I18n.trans("common.please_confirm"),{confirmButtonText:t.$I18n.trans("users.delete"),cancelButtonText:t.$I18n.trans("common.cancel"),type:"warning"});case 3:e.t0=e.sent,n="confirm"===e.t0,e.next=10;break;case 7:e.prev=7,e.t1=e.catch(0),n=!1;case 10:return e.abrupt("return",n);case 11:case"end":return e.stop()}}),e,null,[[0,7]])})))()}}),created:function(){var t=this;this.fetchUsers().then((function(e){return t.loading=!1}))}},re=(n("pM8S"),Object(U.a)(ne,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"users-list"}},[n("div",{staticClass:"row"},[n("div",{staticClass:"col-12 mt-3 no-gutter-xs"},[n("el-table",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}],staticStyle:{width:"100%"},attrs:{data:t.usersList,"row-class-name":t.highlightCurrentUser,"empty-text":t.$I18n.trans("users.failed_to_obtain_user_list")}},[n("el-table-column",{attrs:{prop:"id",label:"#"}}),t._v(" "),n("el-table-column",{attrs:{prop:"name",label:t.$I18n.trans("users.name")}}),t._v(" "),n("el-table-column",{attrs:{prop:"email",label:t.$I18n.trans("users.email")}}),t._v(" "),n("el-table-column",{attrs:{prop:"updatedAt",label:t.$I18n.trans("users.updated_at"),formatter:t.formatUpdatedAt}}),t._v(" "),n("el-table-column",{attrs:{fixed:"right",align:"right",width:"60",prop:"updatedAt"},scopedSlots:t._u([{key:"default",fn:function(e){return[n("el-button",{attrs:{type:"text",size:"small"},on:{click:function(n){return t.editDialogShow(e.row)}}},[t._v(t._s(t.$I18n.trans("users.edit")))])]}}])})],1)],1)]),t._v(" "),n("el-dialog",{attrs:{title:t.$I18n.trans("users.user_edit",{id:t.userToEdit.id||""}),visible:t.dialogVisible},on:{"update:visible":function(e){t.dialogVisible=e}}},[n("el-form",{ref:"formRef",attrs:{model:t.userToEdit,rules:t.dialogFormRules},on:{submit:t.editDialogSubmit},nativeOn:{submit:function(t){t.preventDefault()}}},[n("el-form-item",{attrs:{label:t.$I18n.trans("users.name"),prop:"name"}},[n("el-input",{attrs:{placeholder:t.$I18n.trans("users.name_cannot_be_empty")},model:{value:t.userToEdit.name,callback:function(e){t.$set(t.userToEdit,"name",e)},expression:"userToEdit.name"}})],1),t._v(" "),n("el-form-item",{attrs:{label:t.$I18n.trans("users.password"),prop:"password"}},[n("el-input",{attrs:{type:"password",autocomplete:"off"},model:{value:t.userToEdit.password,callback:function(e){t.$set(t.userToEdit,"password",e)},expression:"userToEdit.password"}})],1),t._v(" "),n("el-form-item",{attrs:{label:t.$I18n.trans("users.email")}},[n("el-input",{attrs:{type:"email",value:t.userToEdit.email,disabled:""}})],1)],1),t._v(" "),n("span",{staticClass:"dialog-footer",attrs:{slot:"footer"},slot:"footer"},[n("el-button",{on:{click:function(e){return t.closeDialog("formRef")}}},[t._v(t._s(t.$I18n.trans("common.cancel")))]),t._v(" "),n("el-button",{attrs:{type:"danger",disabled:t.userToEdit.id===t.currentUserId},on:{click:function(e){return t.deleteUser()}}},[t._v(t._s(t.$I18n.trans("users.delete")))]),t._v(" "),n("el-button",{attrs:{type:"primary"},on:{click:function(e){return t.editDialogSubmit("formRef")}}},[t._v(t._s(t.$I18n.trans("common.confirm")))])],1)],1)],1)}),[],!1,null,"21c16506",null).exports);function se(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}function ae(t){return function(){var e=this,n=arguments;return new Promise((function(r,s){var a=t.apply(e,n);function o(t){se(a,r,s,o,i,"next",t)}function i(t){se(a,r,s,o,i,"throw",t)}o(void 0)}))}}function oe(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(t);e&&(r=r.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,r)}return n}function ie(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?oe(Object(n),!0).forEach((function(e){ce(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):oe(Object(n)).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}function ce(t,e,n){return e in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}var ue={name:"UsersCreate",data:function(){return{rules:{name:[{required:!0,message:this.$I18n.trans("users.name_cannot_be_empty"),trigger:"blur"},{min:2,max:255,message:this.$I18n.trans("users.min_2_max_255"),trigger:"blur"}],email:[{required:!0,type:"email",message:this.$I18n.trans("users.please_set_valid_email"),trigger:"blur"}],password:[{required:!0,message:this.$I18n.trans("users.please_set_valid_password"),trigger:"blur"},{min:12,max:100,message:this.$I18n.trans("users.please_set_valid_password"),trigger:"blur"}]},user:{name:"",email:"",password:""}}},methods:ie(ie({},Object(a.b)(["addUser"])),{},{submitForm:function(t){var e=this;return ae(i.a.mark((function n(){return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:return n.next=2,e.validateForm(t);case 2:n.sent&&e.addUser(e.user).then((function(t){return e.notifyActionResult(t)}));case 4:case"end":return n.stop()}}),n)})))()},validateForm:function(t){var e=this;return ae(i.a.mark((function n(){return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:return n.abrupt("return",new Promise((function(n,r){e.$refs[t].validate((function(t){t&&n(!0),r(!1)}))})));case 1:case"end":return n.stop()}}),n)})))()},notifyActionResult:function(t){var e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:"users.created_success",n=arguments.length>2&&void 0!==arguments[2]?arguments[2]:"users.failed_to_create";t?(this.$notify.success(this.$I18n.trans(e)),this.$router.push(Ae)):this.$notify.error(this.$I18n.trans(n))}})},le=Object(U.a)(ue,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"users-create"}},[n("div",{staticClass:"row"},[n("div",{staticClass:"col-12 mt-3 no-gutter-xs"},[n("el-form",{ref:"refCreateUserForm",attrs:{model:t.user,rules:t.rules}},[n("el-form-item",{attrs:{label:t.$I18n.trans("users.name"),prop:"name"}},[n("el-input",{model:{value:t.user.name,callback:function(e){t.$set(t.user,"name",e)},expression:"user.name"}})],1),t._v(" "),n("el-form-item",{attrs:{label:t.$I18n.trans("users.email"),prop:"email"}},[n("el-input",{attrs:{type:"email",autocomplete:"off"},model:{value:t.user.email,callback:function(e){t.$set(t.user,"email",e)},expression:"user.email"}})],1),t._v(" "),n("el-form-item",{attrs:{label:t.$I18n.trans("users.password"),prop:"password"}},[n("el-input",{attrs:{type:"password",autocomplete:"off"},model:{value:t.user.password,callback:function(e){t.$set(t.user,"password",e)},expression:"user.password"}})],1),t._v(" "),n("el-form-item",[n("el-button",{attrs:{type:"primary"},on:{click:function(e){return t.submitForm("refCreateUserForm")}}},[t._v(t._s(t.$I18n.trans("users.create")))])],1)],1)],1)])])}),[],!1,null,null,null).exports;function de(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}function me(t){return function(){var e=this,n=arguments;return new Promise((function(r,s){var a=t.apply(e,n);function o(t){de(a,r,s,o,i,"next",t)}function i(t){de(a,r,s,o,i,"throw",t)}o(void 0)}))}}function pe(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(t);e&&(r=r.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,r)}return n}function fe(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?pe(Object(n),!0).forEach((function(e){ve(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):pe(Object(n)).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}function ve(t,e,n){return e in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}var he={name:"ExpenseAdd",data:function(){return{expense:{description:null,amount:null},rules:{description:[{required:!0,message:this.$I18n.trans("expenses.expense_description_required"),trigger:"change"},{min:2,max:255,message:this.$I18n.trans("expenses.min_2_max_255"),trigger:"change"}],amount:[{required:!0,message:this.$I18n.trans("expenses.amount_is_required"),trigger:"change"},{type:"regexp",pattern:"/^\\d*\\.?\\d*$/",message:this.$I18n.trans("expenses.please_set_valid_amount"),trigger:"change"}]}}},methods:fe(fe({},Object(a.b)(["addExpense"])),{},{onSubmit:function(t){var e=this;return me(i.a.mark((function n(){return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:e.$refs[t].validate(function(){var n=me(i.a.mark((function n(r){return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:if(!r){n.next=7;break}return n.next=3,e.createExpense();case 3:return e.resetForm(t),n.abrupt("return",!0);case 7:return n.abrupt("return",!1);case 8:case"end":return n.stop()}}),n)})));return function(t){return n.apply(this,arguments)}}());case 1:case"end":return n.stop()}}),n)})))()},createExpense:function(){var t=this;return me(i.a.mark((function e(){var n,r;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,t.addExpense(t.expense);case 2:e.sent?(n=t.$I18n.trans("expenses.success_expense_added"),t.$notify.success(n),t.expense.description="",t.expense.amount=""):(r=t.$I18n.trans("expenses.fail_expense_added"),t.$notify.error(r));case 4:case"end":return e.stop()}}),e)})))()},resetForm:function(t){this.$refs[t].resetFields()}})},be=(n("Ofak"),Object(U.a)(he,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"add-expense"},[n("el-form",{ref:"refExpenseForm",attrs:{inline:!0,model:t.expense,rules:t.rules}},[n("el-form-item",{attrs:{label:t.$I18n.trans("expenses.what"),prop:"description"}},[n("el-input",{attrs:{clearable:""},model:{value:t.expense.description,callback:function(e){t.$set(t.expense,"description",e)},expression:"expense.description"}})],1),t._v(" "),n("el-form-item",{attrs:{label:t.$I18n.trans("expenses.amount"),prop:"amount"}},[n("el-input",{attrs:{type:"number",inputmode:"decimal",decimal:"true",clearable:"",min:"0",step:"0.01"},model:{value:t.expense.amount,callback:function(e){t.$set(t.expense,"amount",e)},expression:"expense.amount"}})],1),t._v(" "),n("el-form-item",[n("el-button",{attrs:{type:"primary"},on:{click:function(e){return t.onSubmit("refExpenseForm")}}},[t._v(t._s(t.$I18n.trans("common.add")))])],1)],1)],1)}),[],!1,null,"8243289e",null).exports),ge=n("ZhIB"),ye=n.n(ge);function xe(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(t);e&&(r=r.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,r)}return n}function we(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?xe(Object(n),!0).forEach((function(e){_e(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):xe(Object(n)).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}function _e(t,e,n){return e in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}var ke={name:"Expenses",components:{ExpenseAdd:be},data:function(){return{loading:!0}},computed:Object(a.c)(["expensesList"]),methods:we(we({},Object(a.b)(["fetchExpenses"])),{},{getSummaries:function(t){var e=this,n=t.columns,r=t.data,s=[];return n.forEach((function(t,n){if(0===n)s[n]=e.$I18n.trans("expenses.total");else if(1===n)s[n]="";else{var a=r.map((function(e){return Number(e[t.property])})).reduce((function(t,e){var n=Number(e);return isNaN(n)?t:t+e}),0);s[n]=ye()(a).format("$ 0,0[.]00")}})),s},formatCurr:function(t,e,n,r){return ye()(Number(n)).format("$ 0,0[.]00")},formatDate:function(t,e,n,r){var s=z()(n),a=s.format("YYYY-MM-DD"),o=s.fromNow();return"".concat(o," on ").concat(a)}}),mounted:function(){var t=this;this.fetchExpenses().then((function(e){return t.loading=!1}))}},je=Object(U.a)(ke,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"expenses"}},[n("div",{staticClass:"row"},[n("div",{staticClass:"col-12 mt-3 no-gutter-xs"},[n("expense-add"),t._v(" "),n("div",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}]},[n("el-table",{staticStyle:{width:"100%"},attrs:{data:t.expensesList,"show-summary":"","summary-method":t.getSummaries,"empty-text":t.$I18n.trans("expenses.failed_to_obtain_expenses")}},[n("el-table-column",{attrs:{prop:"description","min-width":"250",label:t.$I18n.trans("expenses.what")}}),t._v(" "),n("el-table-column",{attrs:{prop:"updated_at",formatter:t.formatDate,"min-width":"220",label:t.$I18n.trans("expenses.when")}}),t._v(" "),n("el-table-column",{attrs:{prop:"amount",align:"right",width:"100",label:t.$I18n.trans("expenses.amount"),formatter:t.formatCurr,fixed:"right"}})],1)],1)],1)])])}),[],!1,null,null,null).exports;function Ce(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}s.a.use(E.a);var Oe={path:"/authenticate",name:"Authenticate",component:F,meta:{requiresAuth:!1}},Ie={path:"/dashboard",alias:"/",name:"Dashboard",component:W,meta:{requiresAuth:!0}},$e={path:"new",name:"NotesCreate",component:yt,meta:{requiresAuth:!0}},Ee={path:":id",name:"NotesShow",component:kt,props:!0,meta:{requiresAuth:!0}},Pe={path:"edit/:id",name:"NotesUpdate",component:It,props:!0,meta:{requiresAuth:!0}},Te={path:"/notes",component:K,children:[{path:"",name:"Notes",component:dt,meta:{requiresAuth:!0}},$e,Ee,Pe]},Se={path:"",alias:"top",name:"HackerNewsTopPostsRoute",component:qt,props:{type:"top"},meta:{requiresAuth:!0}},Ne={path:"best",name:"HackerNewsBestPostsRoute",component:qt,props:{type:"best"},meta:{requiresAuth:!0}},De={path:"bookmarks",name:"HackerNewsBookmarkedPostsRoute",component:qt,props:{type:"bookmarks"},meta:{requiresAuth:!0}},Ue={path:":id",name:"HackerNewsPost",component:Yt,props:!0,meta:{requiresAuth:!0}},Fe={path:"/hn",component:Tt,children:[Se,Ne,De,Ue]},Be={path:"new",name:"UsersCreate",component:le,meta:{requiresAuth:!0}},Ae={path:"/users",component:Xt,children:[{path:"",name:"UsersList",component:re,meta:{requiresAuth:!0}},Be]},Re={path:"/expenses",name:"Expenses",component:je,meta:{requiresAuth:!0}},ze={path:"/404",name:"NotFound",component:A,props:!0,meta:{requiresAuth:!0}},Le=new E.a({routes:[Oe,Ie,Te,Fe,Ae,Re,ze,{path:"*",redirect:"/404",meta:{requiresAuth:!0}}]});Le.beforeEach(function(){var t,e=(t=i.a.mark((function t(e,n,r){return i.a.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:if(e.matched.some((function(t){return t.meta.requiresAuth}))){t.next=4;break}return r(),t.abrupt("return");case 4:return t.next=6,k();case 6:if(!t.sent){t.next=11;break}return e.name===Oe.name&&r(Ie),r(),t.abrupt("return");case 11:r(Oe);case 12:case"end":return t.stop()}}),t)})),function(){var e=this,n=arguments;return new Promise((function(r,s){var a=t.apply(e,n);function o(t){Ce(a,r,s,o,i,"next",t)}function i(t){Ce(a,r,s,o,i,"throw",t)}o(void 0)}))});return function(t,n,r){return e.apply(this,arguments)}}());var Me=n("XJYT"),qe=n.n(Me),Qe=(n("D66Q"),n("stYL")),He=n.n(Qe);function Ve(t){return(Ve="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t})(t)}function Ye(t,e){return function(t){if(Array.isArray(t))return t}(t)||function(t,e){if("undefined"==typeof Symbol||!(Symbol.iterator in Object(t)))return;var n=[],r=!0,s=!1,a=void 0;try{for(var o,i=t[Symbol.iterator]();!(r=(o=i.next()).done)&&(n.push(o.value),!e||n.length!==e);r=!0);}catch(t){s=!0,a=t}finally{try{r||null==i.return||i.return()}finally{if(s)throw a}}return n}(t,e)||function(t,e){if(!t)return;if("string"==typeof t)return Je(t,e);var n=Object.prototype.toString.call(t).slice(8,-1);"Object"===n&&t.constructor&&(n=t.constructor.name);if("Map"===n||"Set"===n)return Array.from(t);if("Arguments"===n||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n))return Je(t,e)}(t,e)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}()}function Je(t,e){(null==e||e>t.length)&&(e=t.length);for(var n=0,r=new Array(e);n<e;n++)r[n]=t[n];return r}function We(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}function Xe(t,e){for(var n=0;n<e.length;n++){var r=e[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(t,r.key,r)}}s.a.filter("diffForHumans",(function(t){if(t)return z()(t,"YYYY-MM-DD hh:mm:ss").fromNow()})),s.a.filter("domainFromUrl",(function(t){if(t)return new URL(t).hostname.replace("www.","")}));var Ze=function(){function t(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:"translations";We(this,t),this.key=e}var e,n,r;return e=t,(n=[{key:"trans",value:function(t){var e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{};return this._replace(this._extract(t),e)}},{key:"trans_choice",value:function(t){var e,n=this,r=arguments.length>1&&void 0!==arguments[1]?arguments[1]:1,s=arguments.length>2&&void 0!==arguments[2]?arguments[2]:{},a=this._extract(t,"|").split("|");return a.some((function(t){return e=n._match(t,r)})),e=(e=e||(r>1?a[1]:a[0])).replace(/\[.*?\]|\{.*?\}/,""),this._replace(e,s)}},{key:"_match",value:function(t,e){var n=t.match(/^[\{\[]([^\[\]\{\}]*)[\}\]](.*)/);if(n){if(n[1].includes(",")){var r=Ye(n[1].split(",",2),2),s=r[0],a=r[1];if("*"===a&&e>=s)return n[2];if("*"===s&&e<=a)return n[2];if(e>=s&&e<=a)return n[2]}return n[1]==e?n[2]:null}}},{key:"_replace",value:function(t,e){if("object"===Ve(t))return t;for(var n in e)t=t.toString().replace(":".concat(n),e[n]).replace(":".concat(n.toUpperCase()),e[n].toString().toUpperCase()).replace(":".concat(n.charAt(0).toUpperCase()).concat(n.slice(1)),e[n].toString().charAt(0).toUpperCase()+e[n].toString().slice(1));return t.toString().trim()}},{key:"_extract",value:function(t){var e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:null,n=t.toString().split("::"),r=n.pop().toString().split(".");return n.length>0&&(n[0]+="::"),n.concat(r).reduce((function(n,r){return n[r]||e||t}),window[this.key])}}])&&Xe(e.prototype,n),r&&Xe(e,r),t}(),Ke=n("D/Jt"),Ge={name:"MainNavigation",data:function(){return{dashboardRoute:Ie,menuOptions:[{route:Te,text:this.$I18n.trans("notes.notes")},{route:Fe,text:this.$I18n.trans("hackernews.hackernews")},{route:Re,text:this.$I18n.trans("expenses.expenses")},{route:Ae,text:this.$I18n.trans("users.users")}],showMenu:!0,menuCollapsed:!0,appRoutes:[]}},methods:{logout:function(t){var e=this;t.preventDefault(),u.a.post("/logout",{}).then((function(){window.localStorage.removeItem("ZE_UserInfo"),e.showMenu=!1,e.$router.push(Oe),e.showMenu=!1}))},toggleMenuVisibility:function(t){this.showMenu=Oe.name!==t.name}},watch:{$route:function(t,e){this.toggleMenuVisibility(t)}},mounted:function(){this.toggleMenuVisibility(this.$route)}},tn=(n("HF9f"),Object(U.a)(Ge,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return t.showMenu?n("nav",{staticClass:"navbar navbar-expand-md navbar-light navbar-laravel",attrs:{id:"main-navigation"}},[n("div",{staticClass:"container"},[n("router-link",{staticClass:"navbar-brand",attrs:{exact:"",to:t.dashboardRoute}},[t._v("Zetetic Elench")]),t._v(" "),n("button",{staticClass:"navbar-toggler",attrs:{type:"button"},on:{click:function(e){t.menuCollapsed=!t.menuCollapsed}}},[n("span",{staticClass:"navbar-toggler-icon"})]),t._v(" "),n("div",{staticClass:"navbar-collapse",class:{collapse:t.menuCollapsed}},[n("ul",{staticClass:"navbar-nav mr-auto"},t._l(t.menuOptions,(function(e){return n("li",{staticClass:"nav-item"},[n("router-link",{staticClass:"nav-link text-right text-sm-left",attrs:{to:e.route}},[t._v(t._s(e.text))])],1)})),0),t._v(" "),n("ul",{staticClass:"navbar-nav ml-auto"},[n("li",{staticClass:"nav-item"},[n("a",{staticClass:"nav-link text-right text-sm-left",attrs:{href:"#"},on:{click:function(e){return t.logout(e)}}},[t._v(t._s(t.$I18n.trans("users.logout")))])])])])],1)]):t._e()}),[],!1,null,"590a6ef5",null).exports);n("XDMY");ye.a.locale("pt-pt"),u.a.defaults.headers.common["X-Requested-With"]="XMLHttpRequest",u.a.defaults.withCredentials=!0,s.a.component("main-navigation",tn),s.a.use(qe.a,{locale:He.a}),s.a.use(Ke.a),s.a.prototype.$I18n=new Ze;new s.a({store:$,router:Le,el:"#app"})},deJl:function(t,e,n){var r=n("fSEL");"string"==typeof r&&(r=[[t.i,r,""]]);var s={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(r,s);r.locals&&(t.exports=r.locals)},eGZ0:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\n.divider[data-v-3d91250f] {\n  border-left: 1px solid #ccc;\n}\n",""])},fSEL:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,'\n.loader-text[data-v-7a7ca61a] {\n  background: #808080;\n  border-radius: 2px;\n}\n.loader-text--1x3[data-v-7a7ca61a] {\n  width: 33%;\n}\n.note[data-v-7a7ca61a] .footnotes {\n  font-size: 0.8rem;\n}\n.note[data-v-7a7ca61a] .footnotes p {\n  margin-bottom: 0.5rem;\n}\n.note[data-v-7a7ca61a] table.table th[align="left"] {\n  text-align: left;\n}\n.note[data-v-7a7ca61a] table.table th[align="center"] {\n  text-align: center;\n}\n.note[data-v-7a7ca61a] table.table th[align="right"] {\n  text-align: right;\n}\n',""])},fysm:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\n.bookmark-story[data-v-a624b0ec] {\n  position: relative;\n  top: 2px;\n  text-decoration: none;\n}\n.loader-text[data-v-a624b0ec] {\n  background: #808080;\n  border-radius: 2px;\n}\n.loader-text--1x3[data-v-a624b0ec] {\n  width: 33%;\n}\n.loader-text-top[data-v-a624b0ec] {\n  height: 22px;\n  margin-bottom: 6px;\n}\n.loader-text-bottom[data-v-a624b0ec] {\n  height: 18px;\n}\n",""])},gE6I:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\n.form-control[data-v-45329d92] {\n  width: auto;\n}\n",""])},iqeS:function(t,e,n){"use strict";var r=n("GFEm");n.n(r).a},mgPi:function(t,e,n){"use strict";var r=n("9bXS");n.n(r).a},nNO0:function(t,e,n){"use strict";var r=n("deJl");n.n(r).a},pM8S:function(t,e,n){"use strict";var r=n("A19t");n.n(r).a},pyCd:function(t,e){},wdM6:function(t,e,n){"use strict";var r=n("8O32");n.n(r).a},"yx/N":function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\ntextarea[data-v-401ef6a5] {\n  outline: 0;\n  background: #f4f4f4;\n  box-shadow: 1px 1px 0 #ccc;\n  border-color: transparent;\n  border-radius: 0;\n  font-family: 'Inconsolata', monospace;\n}\ntextarea[data-v-401ef6a5]:focus {\n  box-shadow: 1px 1px 0 #ccc;\n  outline: 0;\n  background: #f4f4f4;\n  border-color: transparent;\n}\n.new-tag-group[data-v-401ef6a5] {\n  font-size: 0;\n}\n.new-tag-group .input-new-tag[data-v-401ef6a5], .new-tag-group .button-new-tag[data-v-401ef6a5] {\n  vertical-align: middle;\n  width: 120px;\n}\n#note-update[data-v-401ef6a5] .el-checkbox-button--small .el-checkbox-button__inner {\n  border-radius: 4px;\n  border: 1px solid #DCDFE6;\n}\n#note-update[data-v-401ef6a5] .el-checkbox-button.is-checked .el-checkbox-button__inner {\n  border-color: #409EFF;\n}\n",""])},zTjN:function(t,e,n){"use strict";var r=n("D1a/");n.n(r).a}},[[0,1,2]]]);
+(window.webpackJsonp=window.webpackJsonp||[]).push([[0],{"+KkT":function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\n#users-list[data-v-21c16506] .el-table .success-row {\n  background: #f0f9eb;\n}\n@media (max-width: 575px) {\n#users-list[data-v-21c16506] .el-dialog {\n    width: 100%;\n}\n}\n",""])},"/BTM":function(t,e,n){var r=n("fysm");"string"==typeof r&&(r=[[t.i,r,""]]);var s={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(r,s);r.locals&&(t.exports=r.locals)},0:function(t,e,n){n("bUC5"),t.exports=n("pyCd")},"0P3G":function(t,e,n){"use strict";var r=n("/BTM");n.n(r).a},"1XD5":function(t,e,n){"use strict";var r=n("KJzQ");n.n(r).a},"69hy":function(t,e,n){var r=n("JnLb");"string"==typeof r&&(r=[[t.i,r,""]]);var s={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(r,s);r.locals&&(t.exports=r.locals)},"8O32":function(t,e,n){var r=n("eGZ0");"string"==typeof r&&(r=[[t.i,r,""]]);var s={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(r,s);r.locals&&(t.exports=r.locals)},"9bXS":function(t,e,n){var r=n("gE6I");"string"==typeof r&&(r=[[t.i,r,""]]);var s={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(r,s);r.locals&&(t.exports=r.locals)},A19t:function(t,e,n){var r=n("+KkT");"string"==typeof r&&(r=[[t.i,r,""]]);var s={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(r,s);r.locals&&(t.exports=r.locals)},AF6d:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n/* Chrome, Safari, Edge, Opera */\n[data-v-0a4f35a8] input::-webkit-outer-spin-button,[data-v-0a4f35a8] input::-webkit-inner-spin-button {\n  -webkit-appearance: none;\n  margin: 0;\n}\n\n/* Firefox */\n[data-v-0a4f35a8] input[type=number] {\n  -moz-appearance: textfield;\n}\n\n",""])},"D1a/":function(t,e,n){var r=n("WGSt");"string"==typeof r&&(r=[[t.i,r,""]]);var s={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(r,s);r.locals&&(t.exports=r.locals)},F4tQ:function(t,e,n){"use strict";var r=n("WwcV");n.n(r).a},GFEm:function(t,e,n){var r=n("yx/N");"string"==typeof r&&(r=[[t.i,r,""]]);var s={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(r,s);r.locals&&(t.exports=r.locals)},GYTP:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\n.slide-fade-enter-active[data-v-764fdbfc] {\n  transition: all .3s ease;\n}\n.slide-fade-leave-active[data-v-764fdbfc] {\n  transition: all .5s cubic-bezier(1.0, 0.5, 0.8, 1.0);\n}\n.slide-fade-enter[data-v-764fdbfc], .slide-fade-leave-to[data-v-764fdbfc] {\n  background-color: #ffef0029;\n}\n",""])},HF9f:function(t,e,n){"use strict";var r=n("69hy");n.n(r).a},IqpQ:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\ntextarea[data-v-21a1e722] {\n  outline: 0;\n  background: #f4f4f4;\n  box-shadow: 1px 1px 0 #ccc;\n  border-color: transparent;\n  border-radius: 0;\n  font-family: 'Inconsolata', monospace;\n}\ntextarea[data-v-21a1e722]:focus {\n  box-shadow: 1px 1px 0 #ccc;\n  outline: 0;\n  background: #f4f4f4;\n  border-color: transparent;\n}\n.new-tag-group[data-v-21a1e722] {\n  font-size: 0;\n}\n.new-tag-group .input-new-tag[data-v-21a1e722], .new-tag-group .button-new-tag[data-v-21a1e722] {\n  vertical-align: middle;\n  width: 120px;\n}\n#notes-create[data-v-21a1e722] .el-checkbox-button {\n  border-radius: 4px;\n  margin-right: 0 !important;\n}\n#notes-create[data-v-21a1e722] .el-checkbox-button.is-checked .el-checkbox-button__inner {\n  border-color: #409EFF;\n}\n",""])},J82J:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\n.slide-enter-active[data-v-1a92697e] {\n  transition-duration: 0.3s;\n  transition-timing-function: ease-in;\n}\n.slide-leave-active[data-v-1a92697e] {\n  transition-duration: 0.3s;\n  transition-timing-function: cubic-bezier(0, 1, 0.5, 1);\n}\n.slide-enter-to[data-v-1a92697e], .slide-leave[data-v-1a92697e] {\n  max-height: 1000px;\n  overflow: hidden;\n}\n.slide-enter[data-v-1a92697e], .slide-leave-to[data-v-1a92697e] {\n  overflow: hidden;\n  max-height: 0;\n}\n.text-line-through[data-v-1a92697e] {\n  text-decoration: line-through;\n}\n.btn-collapse[data-v-1a92697e] {\n  font-family: 'Inconsolata', monospace;\n  padding: 0;\n  height: 23px;\n}\n",""])},JnLb:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"#main-navigation .router-link-active[data-v-590a6ef5] {\n  color: #3490dc;\n}",""])},KJzQ:function(t,e,n){var r=n("IqpQ");"string"==typeof r&&(r=[[t.i,r,""]]);var s={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(r,s);r.locals&&(t.exports=r.locals)},P9EP:function(t,e,n){"use strict";var r=n("avtZ");n.n(r).a},RG2k:function(t,e,n){"use strict";var r=n("Y1AS");n.n(r).a},RnhZ:function(t,e,n){var r={"./af":"K/tc","./af.js":"K/tc","./ar":"jnO4","./ar-dz":"o1bE","./ar-dz.js":"o1bE","./ar-kw":"Qj4J","./ar-kw.js":"Qj4J","./ar-ly":"HP3h","./ar-ly.js":"HP3h","./ar-ma":"CoRJ","./ar-ma.js":"CoRJ","./ar-sa":"gjCT","./ar-sa.js":"gjCT","./ar-tn":"bYM6","./ar-tn.js":"bYM6","./ar.js":"jnO4","./az":"SFxW","./az.js":"SFxW","./be":"H8ED","./be.js":"H8ED","./bg":"hKrs","./bg.js":"hKrs","./bm":"p/rL","./bm.js":"p/rL","./bn":"kEOa","./bn.js":"kEOa","./bo":"0mo+","./bo.js":"0mo+","./br":"aIdf","./br.js":"aIdf","./bs":"JVSJ","./bs.js":"JVSJ","./ca":"1xZ4","./ca.js":"1xZ4","./cs":"PA2r","./cs.js":"PA2r","./cv":"A+xa","./cv.js":"A+xa","./cy":"l5ep","./cy.js":"l5ep","./da":"DxQv","./da.js":"DxQv","./de":"tGlX","./de-at":"s+uk","./de-at.js":"s+uk","./de-ch":"u3GI","./de-ch.js":"u3GI","./de.js":"tGlX","./dv":"WYrj","./dv.js":"WYrj","./el":"jUeY","./el.js":"jUeY","./en-au":"Dmvi","./en-au.js":"Dmvi","./en-ca":"OIYi","./en-ca.js":"OIYi","./en-gb":"Oaa7","./en-gb.js":"Oaa7","./en-ie":"4dOw","./en-ie.js":"4dOw","./en-il":"czMo","./en-il.js":"czMo","./en-in":"7C5Q","./en-in.js":"7C5Q","./en-nz":"b1Dy","./en-nz.js":"b1Dy","./en-sg":"t+mt","./en-sg.js":"t+mt","./eo":"Zduo","./eo.js":"Zduo","./es":"iYuL","./es-do":"CjzT","./es-do.js":"CjzT","./es-us":"Vclq","./es-us.js":"Vclq","./es.js":"iYuL","./et":"7BjC","./et.js":"7BjC","./eu":"D/JM","./eu.js":"D/JM","./fa":"jfSC","./fa.js":"jfSC","./fi":"gekB","./fi.js":"gekB","./fil":"1ppg","./fil.js":"1ppg","./fo":"ByF4","./fo.js":"ByF4","./fr":"nyYc","./fr-ca":"2fjn","./fr-ca.js":"2fjn","./fr-ch":"Dkky","./fr-ch.js":"Dkky","./fr.js":"nyYc","./fy":"cRix","./fy.js":"cRix","./ga":"USCx","./ga.js":"USCx","./gd":"9rRi","./gd.js":"9rRi","./gl":"iEDd","./gl.js":"iEDd","./gom-deva":"qvJo","./gom-deva.js":"qvJo","./gom-latn":"DKr+","./gom-latn.js":"DKr+","./gu":"4MV3","./gu.js":"4MV3","./he":"x6pH","./he.js":"x6pH","./hi":"3E1r","./hi.js":"3E1r","./hr":"S6ln","./hr.js":"S6ln","./hu":"WxRl","./hu.js":"WxRl","./hy-am":"1rYy","./hy-am.js":"1rYy","./id":"UDhR","./id.js":"UDhR","./is":"BVg3","./is.js":"BVg3","./it":"bpih","./it-ch":"bxKX","./it-ch.js":"bxKX","./it.js":"bpih","./ja":"B55N","./ja.js":"B55N","./jv":"tUCv","./jv.js":"tUCv","./ka":"IBtZ","./ka.js":"IBtZ","./kk":"bXm7","./kk.js":"bXm7","./km":"6B0Y","./km.js":"6B0Y","./kn":"PpIw","./kn.js":"PpIw","./ko":"Ivi+","./ko.js":"Ivi+","./ku":"JCF/","./ku.js":"JCF/","./ky":"lgnt","./ky.js":"lgnt","./lb":"RAwQ","./lb.js":"RAwQ","./lo":"sp3z","./lo.js":"sp3z","./lt":"JvlW","./lt.js":"JvlW","./lv":"uXwI","./lv.js":"uXwI","./me":"KTz0","./me.js":"KTz0","./mi":"aIsn","./mi.js":"aIsn","./mk":"aQkU","./mk.js":"aQkU","./ml":"AvvY","./ml.js":"AvvY","./mn":"lYtQ","./mn.js":"lYtQ","./mr":"Ob0Z","./mr.js":"Ob0Z","./ms":"6+QB","./ms-my":"ZAMP","./ms-my.js":"ZAMP","./ms.js":"6+QB","./mt":"G0Uy","./mt.js":"G0Uy","./my":"honF","./my.js":"honF","./nb":"bOMt","./nb.js":"bOMt","./ne":"OjkT","./ne.js":"OjkT","./nl":"+s0g","./nl-be":"2ykv","./nl-be.js":"2ykv","./nl.js":"+s0g","./nn":"uEye","./nn.js":"uEye","./oc-lnc":"Fnuy","./oc-lnc.js":"Fnuy","./pa-in":"8/+R","./pa-in.js":"8/+R","./pl":"jVdC","./pl.js":"jVdC","./pt":"8mBD","./pt-br":"0tRk","./pt-br.js":"0tRk","./pt.js":"8mBD","./ro":"lyxo","./ro.js":"lyxo","./ru":"lXzo","./ru.js":"lXzo","./sd":"Z4QM","./sd.js":"Z4QM","./se":"//9w","./se.js":"//9w","./si":"7aV9","./si.js":"7aV9","./sk":"e+ae","./sk.js":"e+ae","./sl":"gVVK","./sl.js":"gVVK","./sq":"yPMs","./sq.js":"yPMs","./sr":"zx6S","./sr-cyrl":"E+lV","./sr-cyrl.js":"E+lV","./sr.js":"zx6S","./ss":"Ur1D","./ss.js":"Ur1D","./sv":"X709","./sv.js":"X709","./sw":"dNwA","./sw.js":"dNwA","./ta":"PeUW","./ta.js":"PeUW","./te":"XLvN","./te.js":"XLvN","./tet":"V2x9","./tet.js":"V2x9","./tg":"Oxv6","./tg.js":"Oxv6","./th":"EOgW","./th.js":"EOgW","./tl-ph":"Dzi0","./tl-ph.js":"Dzi0","./tlh":"z3Vd","./tlh.js":"z3Vd","./tr":"DoHr","./tr.js":"DoHr","./tzl":"z1FC","./tzl.js":"z1FC","./tzm":"wQk9","./tzm-latn":"tT3J","./tzm-latn.js":"tT3J","./tzm.js":"wQk9","./ug-cn":"YRex","./ug-cn.js":"YRex","./uk":"raLr","./uk.js":"raLr","./ur":"UpQW","./ur.js":"UpQW","./uz":"Loxo","./uz-latn":"AQ68","./uz-latn.js":"AQ68","./uz.js":"Loxo","./vi":"KSF8","./vi.js":"KSF8","./x-pseudo":"/X5v","./x-pseudo.js":"/X5v","./yo":"fzPg","./yo.js":"fzPg","./zh-cn":"XDpg","./zh-cn.js":"XDpg","./zh-hk":"SatO","./zh-hk.js":"SatO","./zh-mo":"OmwH","./zh-mo.js":"OmwH","./zh-tw":"kOpN","./zh-tw.js":"kOpN"};function s(t){var e=a(t);return n(e)}function a(t){if(!n.o(r,t)){var e=new Error("Cannot find module '"+t+"'");throw e.code="MODULE_NOT_FOUND",e}return r[t]}s.keys=function(){return Object.keys(r)},s.resolve=a,t.exports=s,s.id="RnhZ"},WGSt:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\n.bookmark-story[data-v-cdd94d26] {\n  position: relative;\n  top: 2px;\n  text-decoration: none;\n}\n.loader-text[data-v-cdd94d26] {\n  background: #808080;\n  border-radius: 2px;\n}\n.loader-text--1x3[data-v-cdd94d26] {\n  width: 33%;\n}\n.loader-text--2x3[data-v-cdd94d26] {\n  width: 66%;\n}\n",""])},WwcV:function(t,e,n){var r=n("GYTP");"string"==typeof r&&(r=[[t.i,r,""]]);var s={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(r,s);r.locals&&(t.exports=r.locals)},Y1AS:function(t,e,n){var r=n("AF6d");"string"==typeof r&&(r=[[t.i,r,""]]);var s={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(r,s);r.locals&&(t.exports=r.locals)},avtZ:function(t,e,n){var r=n("J82J");"string"==typeof r&&(r=[[t.i,r,""]]);var s={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(r,s);r.locals&&(t.exports=r.locals)},bUC5:function(t,e,n){"use strict";n.r(e);var r=n("XuX8"),s=n.n(r),a=n("L2JU"),o=n("o0o1"),i=n.n(o),c=n("vDqi"),u=n.n(c),l=n("yDJ3"),d=n.n(l);function p(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}function m(t){return function(){var e=this,n=arguments;return new Promise((function(r,s){var a=t.apply(e,n);function o(t){p(a,r,s,o,i,"next",t)}function i(t){p(a,r,s,o,i,"throw",t)}o(void 0)}))}}var f={state:{notes:[],searchQuery:""},getters:{notesList:function(t){return t.notes},searchQuery:function(t){return t.searchQuery}},actions:{fetchNotesList:function(t){var e=arguments;return m(i.a.mark((function n(){var r,s,a,o;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:return r=t.commit,s=e.length>1&&void 0!==e[1]?e[1]:"",r("setSearchQuery",s),n.next=5,u.a.get("/api/notes?query=".concat(s));case 5:return a=n.sent,o=d()(a,"data.data",[]),r("setNotesList",o),n.abrupt("return",!0);case 9:case"end":return n.stop()}}),n)})))()},addNote:function(t,e){return m(i.a.mark((function n(){var r,s,a;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:return t.commit,n.next=3,u.a.post("/api/notecreate",{body:e.body,tags:e.tags});case 3:if(r=n.sent,s=d()(r,"data.data.id"),a=d()(r,"data.data.success"),!s||!a){n.next=8;break}return n.abrupt("return",s);case 8:return n.abrupt("return",!1);case 9:case"end":return n.stop()}}),n)})))()},updateSearchQuery:function(t,e){(0,t.commit)("setSearchQuery",e)}},mutations:{setNotesList:function(t,e){return t.notes=e},setSearchQuery:function(t,e){return t.searchQuery=e}}};function v(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}function h(t){return function(){var e=this,n=arguments;return new Promise((function(r,s){var a=t.apply(e,n);function o(t){v(a,r,s,o,i,"next",t)}function i(t){v(a,r,s,o,i,"throw",t)}o(void 0)}))}}var b={state:{expenses:[]},getters:{expensesList:function(t){return t.expenses}},actions:{fetchExpenses:function(t){return h(i.a.mark((function e(){var n,r,s;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return n=t.commit,e.next=3,u.a.get("/api/expenses");case 3:return r=e.sent,s=d()(r,"data.data",[]),n("setExpenses",s),e.abrupt("return",!0);case 7:case"end":return e.stop()}}),e)})))()},addExpense:function(t,e){return h(i.a.mark((function n(){var r,s,a,o;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:return r=t.commit,n.next=3,u.a.post("api/expenses/create",e);case 3:return s=n.sent,a=d()(s,"data.data"),(o=d()(a,"expense"))&&r("newExpense",o),n.abrupt("return",d()(a,"success",!1));case 8:case"end":return n.stop()}}),n)})))()}},mutations:{setExpenses:function(t,e){return t.expenses=e},newExpense:function(t,e){return t.expenses.unshift(e)}}};function g(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}function x(t){return function(){var e=this,n=arguments;return new Promise((function(r,s){var a=t.apply(e,n);function o(t){g(a,r,s,o,i,"next",t)}function i(t){g(a,r,s,o,i,"throw",t)}o(void 0)}))}}var y={state:{users:[],currentUserId:null},getters:{usersList:function(t){return t.users},currentUserId:function(t){return t.currentUserId}},actions:{fetchUsers:function(t){return x(i.a.mark((function e(){var n,r,s,a;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return n=t.commit,e.next=3,u.a.get("/api/users");case 3:return r=e.sent,s=d()(r,"data.data.users",[]),a=d()(r,"data.data.currentUserId"),n("setUsersList",s),n("setCurrentUserId",a),e.abrupt("return",!0);case 9:case"end":return e.stop()}}),e)})))()},addUser:function(t,e){return x(i.a.mark((function n(){var r,s,a,o;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:return r=t.commit,n.next=3,u.a.post("/api/users/create",e);case 3:return s=n.sent,a=d()(s,"data.data"),(o=d()(a,"user"))&&r("newUser",o),n.abrupt("return",d()(a,"success",!1));case 8:case"end":return n.stop()}}),n)})))()},updateUser:function(t,e){return x(i.a.mark((function n(){var r,s,a,o;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:return r=t.commit,s=!1,n.prev=2,a={id:e.id,name:e.name},e.password&&(a.password=e.password),n.next=7,u.a.put("/api/users/update",a);case 7:o=n.sent,(s=d()(o,"data.data.success",!1))&&r("updateUser",e),n.next=16;break;case 12:n.prev=12,n.t0=n.catch(2),s=!1,console.error(n.t0);case 16:return n.abrupt("return",s);case 17:case"end":return n.stop()}}),n,null,[[2,12]])})))()},destroyUser:function(t,e){return x(i.a.mark((function n(){var r,s,a;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:return r=t.commit,s=!1,n.prev=2,n.next=5,u.a.delete("/api/users/destroy",{data:{id:e}});case 5:a=n.sent,(s=d()(a,"data.data.success",!1))&&r("deleteUser",e),n.next=14;break;case 10:n.prev=10,n.t0=n.catch(2),s=!1,console.error(n.t0);case 14:return n.abrupt("return",s);case 15:case"end":return n.stop()}}),n,null,[[2,10]])})))()}},mutations:{setUsersList:function(t,e){return t.users=e},newUser:function(t,e){return t.user.push(e)},updateUser:function(t,e){t.users.filter((function(t){return t.id===e.id})).name=e.name},setCurrentUserId:function(t,e){return t.currentUserId=e},deleteUser:function(t,e){var n=t.users.findIndex((function(t){return t.id===e}));t.users.splice(n,1)}}};function w(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}function _(t){return function(){var e=this,n=arguments;return new Promise((function(r,s){var a=t.apply(e,n);function o(t){w(a,r,s,o,i,"next",t)}function i(t){w(a,r,s,o,i,"throw",t)}o(void 0)}))}}function k(){return j.apply(this,arguments)}function j(){return(j=_(i.a.mark((function t(){var e,n,r;return i.a.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return e=200,t.prev=1,t.next=4,u.a.get("/api/users/currentUser");case 4:r=t.sent,n=r.status===e,t.next=11;break;case 8:t.prev=8,t.t0=t.catch(1),n=!1;case 11:return t.abrupt("return",n);case 12:case"end":return t.stop()}}),t,null,[[1,8]])})))).apply(this,arguments)}function C(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}function O(t){return function(){var e=this,n=arguments;return new Promise((function(r,s){var a=t.apply(e,n);function o(t){C(a,r,s,o,i,"next",t)}function i(t){C(a,r,s,o,i,"throw",t)}o(void 0)}))}}var I={state:{user:function(){var t;try{t=JSON.parse(window.localStorage.getItem("ZE_UserInfo"))}catch(e){t=null}return t}()},getters:{userInfo:function(t){return t.user}},mutations:{setUserInfo:function(t,e){return t.user=e}},actions:{authenticate:function(t,e){return O(i.a.mark((function n(){var r,s,a;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:return t.commit,r=t.dispatch,s=204,n.next=4,u.a.post("/login",e);case 4:if((a=n.sent).status!==s){n.next=8;break}return n.next=8,r("fetchCurrentUser");case 8:return n.abrupt("return",a);case 9:case"end":return n.stop()}}),n)})))()},fetchCurrentUser:function(t){return O(i.a.mark((function e(){var n,r,s;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return n=t.commit,e.next=3,u.a.get("/api/users/currentUser");case 3:r=e.sent,s=d()(r,"data.data"),a=s,o=void 0,o=JSON.stringify(a),window.localStorage.setItem("ZE_UserInfo",o),n("setUserInfo",s);case 7:case"end":return e.stop()}var a,o}),e)})))()}}};s.a.use(a.a);var $=new a.a.Store({modules:{notes:f,expenses:b,users:y,auth:I}}),E=n("jE9Z");function P(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}function T(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(t);e&&(r=r.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,r)}return n}function S(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?T(Object(n),!0).forEach((function(e){N(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):T(Object(n)).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}function N(t,e,n){return e in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}var D={name:"Authenticate",data:function(){return{userForm:{email:"",password:""},pollingPasswordFieldInterval:null,errorMessage:null}},methods:S(S({},Object(a.b)(["authenticate"])),{},{pollingPasswordField:function(){var t=this;this.pollingPasswordFieldInterval=setInterval((function(){var e=document.createEvent("HTMLEvents");e.initEvent("change",!1,!0),t.$refs.pwf.dispatchEvent(e)}),500)},passwordChanged:function(){this.userForm.password=this.$refs.pwf.value},attemptLogin:function(t){var e,n=this;return(e=i.a.mark((function e(){var r;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return t.preventDefault(),e.next=3,u.a.get("/sanctum/csrf-cookie");case 3:return e.next=5,n.authenticate({email:n.userForm.email,password:n.userForm.password});case 5:r=e.sent,n.checkLoginResponse(r);case 7:case"end":return e.stop()}}),e)})),function(){var t=this,n=arguments;return new Promise((function(r,s){var a=e.apply(t,n);function o(t){P(a,r,s,o,i,"next",t)}function i(t){P(a,r,s,o,i,"throw",t)}o(void 0)}))})()},checkLoginResponse:function(t){var e=t.status;422===e?this.showsErrors(d()(t,"data.errors")):302!==e&&204!==e||(clearInterval(this.pollingPasswordFieldInterval),this.$router.push(Pe))},showsErrors:function(t){var e="";Object.keys(t).forEach((function(n){t[n].forEach((function(t){e+=t+"<br/>"}))})),this.$message({type:"error",dangerouslyUseHTMLString:!0,message:e})}}),mounted:function(){this.pollingPasswordField()}},U=n("KHd+"),F=Object(U.a)(D,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"pt-4",attrs:{id:"view-authenticate"}},[n("div",{staticClass:"row justify-content-center"},[n("div",{staticClass:"col-md-8 no-gutter-xs"},[n("div",{staticClass:"card shadow"},[n("div",{staticClass:"card-header"},[t._v(t._s(t.$I18n.trans("auth.login")))]),t._v(" "),n("div",{staticClass:"card-body"},[n("form",{attrs:{method:"POST",model:t.userForm},on:{submit:function(e){return t.attemptLogin(e)}}},[n("div",{staticClass:"form-group row"},[n("label",{staticClass:"col-md-4 col-form-label text-md-right",attrs:{for:"email"}},[t._v(t._s(t.$I18n.trans("auth.email")))]),t._v(" "),n("div",{staticClass:"col-md-6"},[n("input",{directives:[{name:"model",rawName:"v-model",value:t.userForm.email,expression:"userForm.email"}],staticClass:"form-control",attrs:{id:"email",inputmode:"email",type:"email",name:"email",autocomplete:"email",required:""},domProps:{value:t.userForm.email},on:{input:function(e){e.target.composing||t.$set(t.userForm,"email",e.target.value)}}})])]),t._v(" "),n("div",{staticClass:"form-group row"},[n("label",{staticClass:"col-md-4 col-form-label text-md-right",attrs:{for:"password"}},[t._v(t._s(t.$I18n.trans("auth.password")))]),t._v(" "),n("div",{staticClass:"col-md-6"},[n("input",{directives:[{name:"model",rawName:"v-model",value:t.userForm.password,expression:"userForm.password"}],ref:"pwf",staticClass:"form-control",attrs:{id:"password",name:"password",type:"password",autocomplete:"current-password",required:""},domProps:{value:t.userForm.password},on:{change:t.passwordChanged,input:function(e){e.target.composing||t.$set(t.userForm,"password",e.target.value)}}})])]),t._v(" "),n("div",{staticClass:"form-group row"},[n("div",{staticClass:"col-md-6 offset-md-4"},[n("div",{staticClass:"form-check"},[n("input",{staticClass:"form-check-input",attrs:{type:"checkbox",name:"remember",id:"remember",checked:""}}),t._v(" "),n("label",{staticClass:"form-check-label",attrs:{for:"remember"}},[t._v(t._s(t.$I18n.trans("auth.remember_me")))])])])]),t._v(" "),n("div",{staticClass:"form-group row mb-0"},[n("div",{staticClass:"col-md-8 offset-md-4"},[n("button",{staticClass:"btn btn-primary",attrs:{type:"submit",disabled:!t.userForm.email&&!t.userForm.password}},[t._v(t._s(t.$I18n.trans("auth.login")))])])])])])])])])])}),[],!1,null,null,null).exports,B={name:"NotFound",props:{message:{type:String,required:!1,default:"We could not find what you where looking for!"},linksTo:{type:Array,required:!1}},data:function(){return{dashboardRoute:Pe}}},R=Object(U.a)(B,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"pt-4",attrs:{id:"NotFound"}},[n("h1",{staticClass:"text-center text-capitalize"},[t._v("404 - page not found")]),t._v(" "),n("h5",{staticClass:"text-center text-black-50 mt-4"},[t._v(t._s(t.message))]),t._v(" "),n("div",{staticClass:"px-5 pt-3 text-center"},[n("router-link",{staticClass:"btn btn-lg btn-link",attrs:{to:t.dashboardRoute}},[t._v("Home")]),t._v(" "),t._l(t.linksTo,(function(e){return n("router-link",{key:e.route.name,staticClass:"btn btn-lg btn-link",attrs:{to:e.route}},[t._v(t._s(t.route.linkText))])}))],2)])}),[],!1,null,null,null).exports,A=n("wd/R"),z=n.n(A);function L(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}var q={name:"Weather",data:function(){return{loading:!0,weather:{icon:"",description:""}}},methods:{fetchWeather:function(){var t,e=this;return(t=i.a.mark((function t(){var n,r,s,a,o;return i.a.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:n="openweather--currentweather",s=localStorage.getItem(n);try{r=JSON.parse(s)}catch(t){r=null}if(r&&r.data&&!(r.saved+6e5<Date.now())){t.next=12;break}return r={data:{},saved:Date.now()},t.next=7,fetch("http://api.openweathermap.org/data/2.5/weather?q=Funchal,PT&appid=fe66d26d5b01501290be42a4809bbc4e&units=metric&lang=pt");case 7:return a=t.sent,t.next=10,a.json();case 10:r.data=t.sent,localStorage.setItem(n,JSON.stringify(r));case 12:o=d()(r,"data.weather[0].icon"),e.weather.icon="http://openweathermap.org/img/wn/".concat(o,".png"),e.weather.temp=d()(r,"data.main.temp"),e.weather.description=d()(r,"data.weather[0].description"),setTimeout((function(){return e.loading=!1}),500);case 17:case"end":return t.stop()}}),t)})),function(){var e=this,n=arguments;return new Promise((function(r,s){var a=t.apply(e,n);function o(t){L(a,r,s,o,i,"next",t)}function i(t){L(a,r,s,o,i,"throw",t)}o(void 0)}))})()}},filters:{capitalize:function(t){if(t){var e=t.toString();return e.charAt(0).toUpperCase()+e.slice(1)}},localTimeFromUnixTimestamp:function(t){if(t)return z.a.unix(t).format("HH:mm")}},created:function(){this.fetchWeather(),setInterval(this.fetchWeather,6e5)}},M=(n("wdM6"),Object(U.a)(q,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}],attrs:{id:"weather"}},[n("div",{staticClass:"card mb-3 shadow"},[n("div",{staticClass:"card-body d-flex flex-wrap"},[n("div",{staticClass:"align-self-center"},[n("a",{attrs:{href:"https://openweathermap.org/city/2267827",target:"_blank"}},[n("img",{attrs:{src:t.weather.icon,alt:t.weather.description}})])]),t._v(" "),n("div",{staticClass:"divider align-items-stretch mx-2"}),t._v(" "),n("div",{staticClass:"align-self-center ml-2"},[t._v("\n        "+t._s(t._f("capitalize")(t.weather.description))+"\n      ")]),t._v(" "),n("div",{staticClass:"divider align-items-stretch mx-2"}),t._v(" "),n("div",{staticClass:"align-self-center"},[t._v("\n        "+t._s(t.weather.temp)+" ℃\n      ")])])])])}),[],!1,null,"3d91250f",null).exports);function Q(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}var Y={name:"SystemInfo",data:function(){return{loading:!0,memory:{free:"",used:"",total:""},nQueueWorkersRunning:0,up:"",upSince:"",systemInfoLoop:null}},created:function(){this.fetchSystemInfo(),this.systemInfoLoop=setInterval(this.fetchSystemInfo,1e4)},methods:{fetchSystemInfo:function(){var t,e=this;return(t=i.a.mark((function t(){var n,r;return i.a.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.next=2,u.a.get("/api/system-info");case 2:n=t.sent,(r=d()(n,"data.data",null))&&(e.memory.free=d()(r,"memory.free"),e.memory.used=d()(r,"memory.used"),e.memory.total=d()(r,"memory.total"),e.nQueueWorkersRunning=d()(r,"nQueueWorkersRunning"),e.up=d()(r,"up"),e.upSince=d()(r,"upSince")),setTimeout((function(){return e.loading=!1}),500);case 6:case"end":return t.stop()}}),t)})),function(){var e=this,n=arguments;return new Promise((function(r,s){var a=t.apply(e,n);function o(t){Q(a,r,s,o,i,"next",t)}function i(t){Q(a,r,s,o,i,"throw",t)}o(void 0)}))})()}},filters:{capitalize:function(t){if(t)return(t=t.toString()).charAt(0).toUpperCase()+t.slice(1)}}};n("F4tQ");function H(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}var V={name:"Inspire",data:function(){return{inspire:null,fallbackLines:["The hardest part of building software is deciding what to build.","Everything not specified will be specified by the programmer, on the fly.","Design is also specification.","Everything that is public-facing defines the significant part of the business logic.","The open office plan is the best way to destroy productivity.","Don’t multitask. Focus on one task at a time.","There is work time and there is personal time. Don’t mix the two.","Everyone writes bad code except for you. /s"]}},computed:{fallback:function(){return this.fallbackLines[Math.floor(Math.random()*this.fallbackLines.length)]}},created:function(){this.fetchInspire()},methods:{fetchInspire:function(){var t,e=this;return(t=i.a.mark((function t(){var n;return i.a.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.next=2,u.a.get("/api/inspire");case 2:n=t.sent,e.inspire=d()(n,"data.data");case 4:case"end":return t.stop()}}),t)})),function(){var e=this,n=arguments;return new Promise((function(r,s){var a=t.apply(e,n);function o(t){H(a,r,s,o,i,"next",t)}function i(t){H(a,r,s,o,i,"throw",t)}o(void 0)}))})()}}},J={name:"Dashboard",components:{Weather:M,SystemInfo:Object(U.a)(Y,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"card mb-3 shadow"},[n("div",{staticClass:"card-header"},[t._v(t._s(t.$I18n.trans("system.info")))]),t._v(" "),n("div",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}],staticClass:"card-body"},[n("transition",{attrs:{name:"slide-fade",mode:"out-in"}},[n("p",{key:t.up,staticClass:"alert-info"},[t._v("\n        "+t._s(t._f("capitalize")(t.up))+" "+t._s(t.$I18n.trans("system.since"))+" "+t._s(t.upSince)+"\n      ")])]),t._v(" "),n("p",{class:{"alert-success":t.nQueueWorkersRunning,"alert-danger":!t.nQueueWorkersRunning}},[t._v("\n      "+t._s(t.$I18n.trans("system.number_queue_workers"))+": "+t._s(t.nQueueWorkersRunning)+"\n    ")]),t._v(" "),n("div",{staticClass:"pt-4"},[n("table",{staticClass:"table table-sm"},[n("caption",[t._v(t._s(t.$I18n.trans("system.since")))]),t._v(" "),n("thead",[n("tr",[n("th",[t._v(t._s(t.$I18n.trans("system.used")))]),t._v(" "),n("th",[t._v(t._s(t.$I18n.trans("system.free")))]),t._v(" "),n("th",[t._v(t._s(t.$I18n.trans("system.total")))])])]),t._v(" "),n("tbody",[n("transition",{attrs:{name:"slide-fade",mode:"out-in"}},[n("tr",{key:t.memory.used},[n("td",[t.memory.used?t._e():n("span",[t._v(" ")]),t._v(t._s(t.memory.used))]),t._v(" "),n("td",[t._v(t._s(t.memory.free))]),t._v(" "),n("td",[t._v(t._s(t.memory.total))])])])],1)])])],1)])}),[],!1,null,"764fdbfc",null).exports,Inspire:Object(U.a)(V,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"card mb-3 shadow"},[n("div",{staticClass:"card-body"},[n("small",[t.inspire?t._e():n("span",[t._v(t._s(t.fallback))]),t._v(t._s(t.inspire))])])])}),[],!1,null,null,null).exports},data:function(){return{}},beforeRouteLeave:function(t,e,n){clearInterval(this.$refs.refSystemInfo.systemInfoLoop),n()}},W=Object(U.a)(J,(function(){var t=this.$createElement,e=this._self._c||t;return e("div",{staticClass:"pt-4",attrs:{id:"view-dashboard"}},[e("div",{staticClass:"row justify-content-center"},[e("div",{staticClass:"col-sm-6 no-gutter-xs"},[e("inspire"),this._v(" "),e("system-info",{ref:"refSystemInfo"})],1),this._v(" "),e("div",{staticClass:"col-sm-6 no-gutter-xs"},[e("weather")],1)])])}),[],!1,null,null,null).exports,X={name:"Navigation",data:function(){return{notesRoute:De,notesCreateRoute:Te}}},Z={name:"Notes",components:{Navigation:Object(U.a)(X,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"navigation"}},[n("div",{staticClass:"row justify-content-center top-submenu"},[n("div",{staticClass:"col-12 no-gutter-xs"},[n("div",{staticClass:"btn-group d-flex mb-2"},[n("router-link",{staticClass:"btn btn-group-sm w-100 inner-nav btn-secondary",attrs:{to:t.notesRoute}},[t._v(t._s(t.$I18n.trans("notes.notes")))]),t._v(" "),n("router-link",{staticClass:"btn btn-group-sm w-100 inner-nav btn-secondary",attrs:{to:t.notesCreateRoute}},[t._v(t._s(t.$I18n.trans("notes.new_note")))])],1)])])])}),[],!1,null,null,null).exports}},K=Object(U.a)(Z,(function(){var t=this.$createElement,e=this._self._c||t;return e("div",{attrs:{id:"notes"}},[e("navigation"),this._v(" "),e("router-view")],1)}),[],!1,null,null,null).exports,G=n("9/5/"),tt=n.n(G);function et(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(t);e&&(r=r.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,r)}return n}function nt(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?et(Object(n),!0).forEach((function(e){rt(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):et(Object(n)).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}function rt(t,e,n){return e in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}var st={name:"AlgoliaSearch",computed:{searchQuery:{get:function(){return this.$store.state.notes.searchQuery},set:function(t){this.$store.commit("setSearchQuery",t)}}},watch:{searchQuery:function(){(0===this.searchQuery.length||this.searchQuery.length>1)&&this.submit()}},methods:nt(nt({},Object(a.b)(["fetchNotesList"])),{},{submit:tt()((function(){this.fetchNotesList(this.searchQuery)}),400),clearSearch:function(){this.searchQuery="",this.submit()}})},at=(n("mgPi"),Object(U.a)(st,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"col-12 no-gutter-xs",attrs:{id:"algolia-search"}},[n("div",{staticClass:"justify-content-center form-inline d-flex"},[n("label",{staticClass:"sr-only",attrs:{for:"query"}},[t._v(t._s(t.$I18n.trans("notes.search")))]),t._v(" "),n("input",{directives:[{name:"model",rawName:"v-model",value:t.searchQuery,expression:"searchQuery"}],staticClass:"form-control flex-grow-1",attrs:{type:"text",id:"query",name:"query",placeholder:t.$I18n.trans("notes.min_2_chars")},domProps:{value:t.searchQuery},on:{keyup:function(e){return!e.type.indexOf("key")&&t._k(e.keyCode,"esc",27,e.key,["Esc","Escape"])?null:t.clearSearch(e)},input:function(e){e.target.composing||(t.searchQuery=e.target.value)}}}),t._v("\n     \n    "),t._m(0)])])}),[function(){var t=this.$createElement,e=this._self._c||t;return e("picture",[e("source",{attrs:{srcset:"search-by-algolia.svg",media:"(min-width: 800px)"}}),this._v(" "),e("img",{staticStyle:{"max-height":"30px"},attrs:{src:"algolia-blue-mark.svg",alt:"search by algolia"}})])}],!1,null,"45329d92",null).exports);function ot(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(t);e&&(r=r.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,r)}return n}function it(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?ot(Object(n),!0).forEach((function(e){ct(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):ot(Object(n)).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}function ct(t,e,n){return e in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}var ut={name:"NotesList",data:function(){return{notesShowRoute:Se,loading:!0}},computed:Object(a.c)(["notesList"]),methods:it(it({},Object(a.b)(["fetchNotesList","updateSearchQuery"])),{},{close:function(){var t=this;this.updateSearchQuery(""),this.loading=!0,this.fetchNotesList().then((function(e){t.loading=!1}))}}),created:function(){var t=this;this.fetchNotesList().then((function(e){t.loading=!1}))}},lt={name:"NotesListing",components:{AlgoliaSearch:at,NotesList:Object(U.a)(ut,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"notes-list"}},[n("ul",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}],staticClass:"list-group list-group-flush"},[t.notesList.length?t._e():n("el-alert",{attrs:{title:t.$I18n.trans("notes.alert_failed_to_find_records"),type:"info",center:"","close-text":t.$I18n.trans("notes.alert_close_text")},on:{close:t.close}}),t._v(" "),t._l(t.notesList,(function(e){return n("li",{staticClass:"list-group-item list-group-item-action p-2 p-sm-3"},[n("router-link",{staticClass:"text-secondary",attrs:{to:{name:t.notesShowRoute.name,params:{id:e.id}}}},[n("div",{staticClass:"d-flex w-100 justify-content-between"},[n("h5",{staticClass:"mb-1"},[t._v(t._s(e.title))]),t._v(" "),n("small",{staticClass:"d-none d-sm-block"},[t._v(t._s(t._f("diffForHumans")(e.updated_at)))])])]),t._v(" "),n("div",t._l(e.tags,(function(e){return n("el-tag",{key:e.id,staticClass:"mr-1 mb-1",attrs:{size:"small"}},[t._v(t._s(e.name))])})),1)],1)}))],2)])}),[],!1,null,null,null).exports}},dt=Object(U.a)(lt,(function(){var t=this.$createElement,e=this._self._c||t;return e("div",{attrs:{id:"notes-listing"}},[e("div",{staticClass:"row justify-content-center"},[e("algolia-search")],1),this._v(" "),e("div",{staticClass:"row"},[e("div",{staticClass:"col-12 mt-3 no-gutter-xs"},[e("notes-list")],1)])])}),[],!1,null,null,null).exports,pt={methods:{textareaCharInserter:function(t){var e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:"  ";t.preventDefault();var n=this.note.body,r=t.target.selectionStart,s=n.slice(0,r),a=n.slice(r);this.note.body="".concat(s).concat(e).concat(a),t.target.value=this.note.body,t.target.selectionEnd=t.target.selectionStart=r+1}}};function mt(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}function ft(t){return function(){var e=this,n=arguments;return new Promise((function(r,s){var a=t.apply(e,n);function o(t){mt(a,r,s,o,i,"next",t)}function i(t){mt(a,r,s,o,i,"throw",t)}o(void 0)}))}}function vt(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(t);e&&(r=r.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,r)}return n}function ht(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?vt(Object(n),!0).forEach((function(e){bt(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):vt(Object(n)).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}function bt(t,e,n){return e in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}var gt={name:"NotesCreate",mixins:[pt],data:function(){return{newTagInputVisible:!1,newTag:"",errors:[],note:{body:""},tags:[],selectedTags:[],inputEl:null}},methods:ht(ht({},Object(a.b)(["addNote"])),{},{insertCodeBlockBackTicks:function(){if(this.inputEl||(this.inputEl=document.querySelector("#note-create textarea")),!document.execCommand("insertText",!1,"```")&&"function"==typeof this.inputEl.setRangeText){var t=this.inputEl.selectionStart;this.inputEl.setRangeText("```"),this.inputEl.selectionStart=this.inputEl.selectionEnd=t+"```".length;var e=document.createEvent("UIEvent");e.initEvent("input",!0,!1),this.inputEl.dispatchEvent(e)}},addNewTag:function(){var t=this;return ft(i.a.mark((function e(){var n,r,s,a;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:if(!t.newTag){e.next=15;break}return t.errors=[],e.prev=2,e.next=5,u.a.post("/api/notetagcreate",{tag:t.newTag});case 5:n=e.sent,r=d()(n,"data.data.success"),s=d()(n,"data.data.id"),r&&s&&(t.tags.push({id:s,name:t.newTag}),t.newTagInputVisible=!1,t.newTag=""),e.next=15;break;case 11:e.prev=11,e.t0=e.catch(2),a=d()(e.t0,"response.data.errors.tag[0]"),t.errors.push({field:"tagCreate",text:a});case 15:case"end":return e.stop()}}),e,null,[[2,11]])})))()},showInput:function(){var t=this;this.newTagInputVisible=!0,this.$nextTick((function(e){t.$refs.saveTagInput.$refs.input.focus()}))},fetchTags:function(){var t=this;return ft(i.a.mark((function e(){var n;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,u.a.get("/api/notes/tags");case 2:n=e.sent,t.tags=d()(n,"data.data",[]);case 4:case"end":return e.stop()}}),e)})))()},submitForm:function(t){var e=this;return ft(i.a.mark((function n(){var r;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:if(t.preventDefault(),!e.newTagInputVisible){n.next=3;break}return n.abrupt("return",!1);case 3:if(e.errors=[],e.note.body){n.next=8;break}return e.errors.push({field:"body",text:"Note text is empty."}),e.note.body=null,n.abrupt("return");case 8:return n.prev=8,n.next=11,e.addNote({body:e.note.body,tags:e.selectedTags});case 11:(r=n.sent)?(e.note.body="",e.selectedTags=[],e.$notify.success(e.$I18n.trans("notes.confirmation_success_note_create")),setTimeout((function(){return e.$router.push(Object.assign({},Se,{params:{id:r}}))}),400)):e.errors.push({field:"na",text:"Failed to create the note."}),n.next=18;break;case 15:n.prev=15,n.t0=n.catch(8),e.errors.push({field:"submit",text:n.t0.message});case 18:case"end":return n.stop()}}),n,null,[[8,15]])})))()}}),created:function(){this.fetchTags()}},xt=(n("1XD5"),Object(U.a)(gt,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"notes-create"}},[n("div",{staticClass:"row justify-content-center"},[n("div",{staticClass:"col-sm no-gutter-xs"},[n("div",{staticClass:"card shadow"},[t.errors.length?n("div",{staticClass:"p-1"},t._l(t.errors,(function(t){return n("el-alert",{key:t.field,attrs:{title:t.text,type:"error"}})})),1):t._e(),t._v(" "),n("form",{attrs:{method:"post"},on:{submit:t.submitForm}},[n("div",{staticClass:"textarea-container"},[n("textarea",{directives:[{name:"model",rawName:"v-model",value:t.note.body,expression:"note.body"}],staticClass:"form-control form-text m-0",attrs:{placeholder:t.$I18n.trans("notes.placeholder_note_title"),name:"body",cols:"100",rows:"15"},domProps:{value:t.note.body},on:{keydown:function(e){return!e.type.indexOf("key")&&t._k(e.keyCode,"tab",9,e.key,"Tab")?null:(e.preventDefault(),t.textareaCharInserter(e))},input:function(e){e.target.composing||t.$set(t.note,"body",e.target.value)}}})]),t._v(" "),n("div",{staticClass:"d-flex justify-content-between"},[n("small",{staticClass:"form-text text-muted ml-3"},[t._m(0)]),t._v(" "),n("el-button",{attrs:{circle:"",size:"small",type:"info"},on:{click:t.insertCodeBlockBackTicks}},[t._v("```")])],1),t._v(" "),n("div",{staticClass:"mt-3 ml-3"},[n("p",[t._v(t._s(t.$I18n.trans("notes.tags"))+":")]),t._v(" "),n("el-checkbox-group",{staticClass:"d-inline-block",attrs:{size:"small"},model:{value:t.selectedTags,callback:function(e){t.selectedTags=e},expression:"selectedTags"}},t._l(t.tags,(function(e){return n("el-checkbox-button",{key:e.id,staticClass:"mr-1",attrs:{label:e.id}},[t._v(t._s(e.name))])})),1),t._v(" "),n("div",{staticClass:"d-inline-block new-tag-group"},[t.newTagInputVisible?n("el-input",{ref:"saveTagInput",staticClass:"input-new-tag",attrs:{size:"small"},on:{blur:t.addNewTag},nativeOn:{keyup:function(e){return!e.type.indexOf("key")&&t._k(e.keyCode,"enter",13,e.key,"Enter")?null:t.addNewTag(e)}},model:{value:t.newTag,callback:function(e){t.newTag=e},expression:"newTag"}}):n("el-button",{staticClass:"button-new-tag",attrs:{type:"info",plain:"",size:"small"},on:{click:t.showInput}},[t._v(t._s(t.$I18n.trans("notes.new_tag")))])],1)],1),t._v(" "),n("div",{staticClass:"px-3 pb-3 pt-3"},[n("button",{staticClass:"btn btn-primary",attrs:{type:"submit"}},[t._v(t._s(t.$I18n.trans("common.add")))])])])])])])])}),[function(){var t=this.$createElement;return(this._self._c||t)("a",{staticClass:"text-muted",attrs:{href:"https://commonmark.org/help/",tabindex:"-1",target:"_CommonMark"}},[this._v(this._s(this.$I18n.trans("notes.commonmark")))])}],!1,null,"21a1e722",null).exports);n("WxZA");function yt(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}function wt(t){return function(){var e=this,n=arguments;return new Promise((function(r,s){var a=t.apply(e,n);function o(t){yt(a,r,s,o,i,"next",t)}function i(t){yt(a,r,s,o,i,"throw",t)}o(void 0)}))}}var _t={name:"NotesShow",props:["id"],data:function(){return{loading:!0,note:{},notesUpdateRoute:Ne}},created:function(){this.fetchNote(this.id)},methods:{fetchNote:function(t){var e=this;return wt(i.a.mark((function n(){var r;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:if(t){n.next=2;break}return n.abrupt("return",!1);case 2:return n.prev=2,n.next=5,u.a.get("/api/notes/".concat(t,"?html=1"));case 5:r=n.sent,n.next=14;break;case 8:if(n.prev=8,n.t0=n.catch(2),404!==n.t0.response.status){n.next=14;break}return n.next=13,e.$router.push(Object.assign({},Qe,{params:{message:"Note could not be found.",linksTo:[{route:De,linkText:"Notes"}]}}));case 13:return n.abrupt("return");case 14:e.note=d()(r,"data.data"),setTimeout((function(){return e.loading=!1}),400);case 16:case"end":return n.stop()}}),n,null,[[2,8]])})))()},confirmDelete:function(t){var e=this;return wt(i.a.mark((function n(){var r,s,a,o;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:r=e.$I18n.trans("notes.confirmation_ask_delete",{id:t}),s=e.$I18n.trans("notes.confirmation_success_deleted",{id:t}),a=e.$I18n.trans("notes.delete"),o=e.$I18n.trans("notes.cancel"),e.$confirm(r,"Warning",{confirmButtonText:a,cancelButtonText:o,type:"warning",center:!0}).then((function(){e.deleteNote(t).then((function(t){t&&(e.$notify.success(s),setTimeout((function(){return e.$router.push(De)}),400))})).catch((function(n){e.$notify.warning(e.$I18n.trans("notes.failed_delete",{id:t}))}))})).catch((function(){e.$notify.info(e.$I18n.trans("notes.delete_canceled",{id:t}))}));case 5:case"end":return n.stop()}}),n)})))()},deleteNote:function(t){return wt(i.a.mark((function e(){var n,r;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,u.a.delete("/api/notedestroy/".concat(t),{_method:"delete"});case 2:if(n=e.sent,!(r=d()(n,"data.data.success",!1))){e.next=6;break}return e.abrupt("return",r);case 6:throw"Delete failed";case 7:case"end":return e.stop()}}),e)})))()}}},kt=(n("nNO0"),Object(U.a)(_t,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"note"},[n("div",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}],staticClass:"row justify-content-center"},[n("div",{staticClass:"col-12 no-gutter-xs"},[t.note.id?t._e():n("div",{staticClass:"mt-3 "},[n("p",{staticClass:"loader-text pb-2 pt-3"},[t._v(" ")]),t._v(" "),n("p",{staticClass:"loader-text"},[t._v(" ")]),t._v(" "),n("p",{staticClass:"loader-text"},[t._v(" ")]),t._v(" "),n("p",{staticClass:"loader-text loader-text--1x3 d-inline-block"},[t._v("   ")])]),t._v(" "),n("div",{staticClass:"mt-3",domProps:{innerHTML:t._s(t.note.body)}}),t._v(" "),n("div",t._l(t.note.tags,(function(e){return n("el-tag",{key:e.id,staticClass:"mr-1 mb-1",attrs:{size:"big"}},[t._v(t._s(e.name))])})),1)])]),t._v(" "),n("div",{staticClass:"row"},[n("div",{staticClass:"col-12"},[n("div",{staticClass:"mt-3"},[t.note.id?n("router-link",{staticClass:"btn btn-primary",attrs:{to:Object.assign({},t.notesUpdateRoute,{params:{id:t.note.id}})}},[t._v(t._s(t.$I18n.trans("notes.edit")))]):t._e(),t._v(" "),n("button",{staticClass:"btn btn-danger ml-3",on:{click:function(e){return t.confirmDelete(t.note.id)}}},[t._v(t._s(t.$I18n.trans("notes.delete")))])],1)])])])}),[],!1,null,"7a7ca61a",null).exports);function jt(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}function Ct(t){return function(){var e=this,n=arguments;return new Promise((function(r,s){var a=t.apply(e,n);function o(t){jt(a,r,s,o,i,"next",t)}function i(t){jt(a,r,s,o,i,"throw",t)}o(void 0)}))}}var Ot={name:"NotesUpdate",mixins:[pt],props:["id"],data:function(){return{loading:!0,newTagInputVisible:!1,newTag:"",errors:[],note:{},tags:[{id:1,name:""}],selectedTags:[],inputEl:null}},methods:{insertCodeBlockBackTicks:function(){if(this.inputEl||(this.inputEl=document.querySelector("#note-update textarea")),!document.execCommand("insertText",!1,"```")&&"function"==typeof this.inputEl.setRangeText){var t=this.inputEl.selectionStart;this.inputEl.setRangeText("```"),this.inputEl.selectionStart=this.inputEl.selectionEnd=t+"```".length;var e=document.createEvent("UIEvent");e.initEvent("input",!0,!1),this.inputEl.dispatchEvent(e)}},addNewTag:function(){var t=this;return Ct(i.a.mark((function e(){var n,r,s,a;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:if(!t.newTag){e.next=15;break}return t.errors=[],e.prev=2,e.next=5,u.a.post("/api/notetagcreate",{tag:t.newTag});case 5:n=e.sent,r=d()(n,"data.data.success"),s=d()(n,"data.data.id"),r&&s&&(t.tags.push({id:s,name:t.newTag}),t.newTagInputVisible=!1,t.newTag=""),e.next=15;break;case 11:e.prev=11,e.t0=e.catch(2),a=d()(e.t0,"response.data.errors.tag[0]"),t.errors.push({field:"tagCreate",text:a});case 15:case"end":return e.stop()}}),e,null,[[2,11]])})))()},showInput:function(){var t=this;this.newTagInputVisible=!0,this.$nextTick((function(e){t.$refs.saveTagInput.$refs.input.focus()}))},fetchNote:function(t){var e=this;return Ct(i.a.mark((function n(){var r;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:if(t){n.next=2;break}return n.abrupt("return",!1);case 2:return n.next=4,u.a.get("/api/notes/".concat(t,"?html=0"));case 4:r=n.sent,e.note=d()(r,"data.data",[]),e.selectedTags=e.note.tags.map((function(t){return t.id})),setTimeout((function(){return e.loading=!1}),400);case 8:case"end":return n.stop()}}),n)})))()},fetchTags:function(){var t=this;return Ct(i.a.mark((function e(){var n;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,u.a.get("/api/notes/tags");case 2:n=e.sent,t.tags=d()(n,"data.data",[]);case 4:case"end":return e.stop()}}),e)})))()},submitForm:function(t){var e=this;return Ct(i.a.mark((function n(){var r;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:if(t.preventDefault(),!e.newTagInputVisible){n.next=3;break}return n.abrupt("return",!1);case 3:if(e.errors=[],e.note.body){n.next=7;break}return e.errors.push({field:"body",text:"Note text is empty."}),n.abrupt("return");case 7:return n.prev=7,n.next=10,u.a.put("/api/noteupdate/".concat(e.note.id),{body:e.note.body,tags:e.selectedTags,_method:"put"});case 10:r=n.sent,d()(r,"data.data.success")?(e.selectedTags=[],e.$notify.success(e.$I18n.trans("notes.confirmation_success_note_updated")),setTimeout((function(){return e.$router.push(Object.assign({},Se,{params:{id:e.note.id}}))}),400)):e.errors.push({field:"na",text:"Failed to update note."}),n.next=17;break;case 14:n.prev=14,n.t0=n.catch(7),e.errors.push({field:"submit",text:n.t0.message});case 17:case"end":return n.stop()}}),n,null,[[7,14]])})))()}},created:function(){var t=this;this.fetchTags().then((function(){t.fetchNote(t.id)}))}},It=(n("iqeS"),Object(U.a)(Ot,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"notes-update"}},[n("div",{staticClass:"row justify-content-center"},[n("div",{staticClass:"col-sm no-gutter-xs"},[n("div",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}],staticClass:"card shadow"},[t.errors.length?n("div",{staticClass:"p-1"},t._l(t.errors,(function(t){return n("el-alert",{key:t.field,attrs:{title:t.text,type:"error"}})})),1):t._e(),t._v(" "),n("form",{attrs:{method:"post"},on:{submit:t.submitForm}},[n("div",{staticClass:"textarea-container"},[n("textarea",{directives:[{name:"model",rawName:"v-model",value:t.note.body,expression:"note.body"}],staticClass:"form-control form-text m-0",attrs:{placeholder:t.$I18n.trans("notes.placeholder_note_title"),name:"body",cols:"100",rows:"15"},domProps:{value:t.note.body},on:{keydown:function(e){return!e.type.indexOf("key")&&t._k(e.keyCode,"tab",9,e.key,"Tab")?null:(e.preventDefault(),t.textareaCharInserter(e))},input:function(e){e.target.composing||t.$set(t.note,"body",e.target.value)}}})]),t._v(" "),n("div",{staticClass:"d-flex justify-content-between"},[n("small",{staticClass:"form-text text-muted ml-3"},[t._m(0)]),t._v(" "),n("el-button",{attrs:{circle:"",size:"small",type:"info"},on:{click:t.insertCodeBlockBackTicks}},[t._v("```")])],1),t._v(" "),n("div",{staticClass:"mt-3 ml-3"},[n("p",[t._v("Tags:")]),t._v(" "),n("el-checkbox-group",{staticClass:"d-inline-block",attrs:{size:"small"},model:{value:t.selectedTags,callback:function(e){t.selectedTags=e},expression:"selectedTags"}},t._l(t.tags,(function(e){return n("el-checkbox-button",{key:e.id,staticClass:"mr-1",attrs:{label:e.id}},[t._v(t._s(e.name))])})),1),t._v(" "),n("div",{staticClass:"d-inline-block new-tag-group"},[t.newTagInputVisible?n("el-input",{ref:"saveTagInput",staticClass:"input-new-tag",attrs:{size:"small"},on:{blur:t.addNewTag},nativeOn:{keyup:function(e){return!e.type.indexOf("key")&&t._k(e.keyCode,"enter",13,e.key,"Enter")?null:t.addNewTag(e)}},model:{value:t.newTag,callback:function(e){t.newTag=e},expression:"newTag"}}):n("el-button",{staticClass:"button-new-tag",attrs:{type:"info",plain:"",size:"small"},on:{click:t.showInput}},[t._v(t._s(t.$I18n.trans("notes.new_tag")))])],1)],1),t._v(" "),n("div",{staticClass:"px-3 pb-3 pt-3"},[n("button",{staticClass:"btn btn-primary",attrs:{type:"submit"}},[t._v(t._s(t.$I18n.trans("common.update")))])])])])])])])}),[function(){var t=this.$createElement;return(this._self._c||t)("a",{staticClass:"text-muted",attrs:{href:"https://commonmark.org/help/",tabindex:"-1",target:"_CommonMark"}},[this._v(this._s(this.$I18n.trans("notes.commonmark")))])}],!1,null,"401ef6a5",null).exports),$t={name:"Navigation",props:{numberOfBookmarks:{type:null|Number,required:!0}},data:function(){return{hackerNewsTopPostsRoute:Ue,hackerNewsBestPostsRoute:Fe,hackerNewsBookmarkedPostsRoute:Be}}},Et=Object(U.a)($t,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"navigation"}},[n("div",{staticClass:"row justify-content-center top-submenu"},[n("div",{staticClass:"col-12 no-gutter-xs"},[n("div",{staticClass:"btn-group d-flex mb-2"},[n("router-link",{staticClass:"btn btn-group-sm w-100 inner-nav btn-secondary",attrs:{to:t.hackerNewsTopPostsRoute}},[t._v(t._s(t.$I18n.trans("hackernews.top")))]),t._v(" "),n("router-link",{staticClass:"btn btn-group-sm w-100 inner-nav btn-secondary",attrs:{to:t.hackerNewsBestPostsRoute}},[t._v(t._s(t.$I18n.trans("hackernews.best")))]),t._v(" "),n("router-link",{staticClass:"btn btn-group-sm w-100 inner-nav btn-secondary",attrs:{to:t.hackerNewsBookmarkedPostsRoute}},[t._v(t._s(t.$I18n.trans("hackernews.bookmarks"))+" "),n("span",{staticClass:"badge badge-light"},[t._v(t._s(t.numberOfBookmarks))])])],1)])])])}),[],!1,null,null,null).exports,Pt={name:"HackerNews",components:{Navigation:Et},data:function(){return{numberOfBookmarks:null}},methods:{updateNumberOfBookmarks:function(t){this.numberOfBookmarks=t}}},Tt=Object(U.a)(Pt,(function(){var t=this.$createElement,e=this._self._c||t;return e("div",{attrs:{id:"hacker-news"}},[e("navigation",{attrs:{"number-of-bookmarks":this.numberOfBookmarks}}),this._v(" "),e("router-view",{on:{nBookmarksChangedEvent:this.updateNumberOfBookmarks}})],1)}),[],!1,null,null,null).exports;function St(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}var Nt={data:function(){return{nBookmarks:null,notification:null}},methods:{bookmarkPost:function(t){var e,n=this;return(e=i.a.mark((function e(){var r,s,a,o;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return t.status.saving=!0,r={postId:t.id},(s=!t.status.bookmarked)||(r._method="delete"),e.next=6,u.a.post("/api/bookmarks",r);case 6:a=e.sent,(o=d()(a,"data.data.success",!1))&&(t.status.bookmarked=!t.status.bookmarked,n.nBookmarks+=s?1:-1),n.notifyUserOfBookmarkStatusChange(o,s,t);case 10:case"end":return e.stop()}}),e)})),function(){var t=this,n=arguments;return new Promise((function(r,s){var a=e.apply(t,n);function o(t){St(a,r,s,o,i,"next",t)}function i(t){St(a,r,s,o,i,"throw",t)}o(void 0)}))})()},notifyUserOfBookmarkStatusChange:function(t,e,n){var r={message:this.$I18n.trans("hackernews.add_failure"),type:"error"};t&&(r=e?{message:this.$I18n.trans("hackernews.added_to_bookmarks",{title:n.title}),type:"success"}:{message:this.$I18n.trans("hackernews.remove_from_bookmarks",{title:n.title}),type:"warning"}),this.notification&&!this.notification.closed&&this.notification.close(),this.notification=this.$notify(r),setTimeout((function(){return n.status.saving=!1}),400)}}},Dt=n("Wcq6"),Ut=n.n(Dt),Ft=(n("Zs65"),Ut.a.initializeApp({databaseURL:"hacker-news.firebaseio.com"}).database().ref("/v0"));function Bt(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}function Rt(t){return function(t){if(Array.isArray(t))return At(t)}(t)||function(t){if("undefined"!=typeof Symbol&&Symbol.iterator in Object(t))return Array.from(t)}(t)||function(t,e){if(!t)return;if("string"==typeof t)return At(t,e);var n=Object.prototype.toString.call(t).slice(8,-1);"Object"===n&&t.constructor&&(n=t.constructor.name);if("Map"===n||"Set"===n)return Array.from(t);if("Arguments"===n||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n))return At(t,e)}(t)||function(){throw new TypeError("Invalid attempt to spread non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}()}function At(t,e){(null==e||e>t.length)&&(e=t.length);for(var n=0,r=new Array(e);n<e;n++)r[n]=t[n];return r}var zt={name:"HNPosts",mixins:[Nt],props:{idList:{type:Array,default:[],required:!0}},data:function(){return{loading:!0,posts:[],nBookmarks:null,firstBatchPosition:10,hackerNewsPostRoute:Re}},methods:{fetchItems:function(t){var e=this,n=this.firstBatchPosition;t.length<n&&(n=Math.floor(t.length/2));var r=t.slice(0,n),s=t.slice(n);Promise.all(r.map((function(t){return e.fetchItem(t)}))).then((function(t){e.posts=t.filter(e.isStoryValid),e.attachBookmarked(),e.loading=!1})),this.$nextTick((function(){Promise.all(s.map((function(t){return e.fetchItem(t)}))).then((function(t){var n;(n=e.posts).push.apply(n,Rt(t.filter(e.isStoryValid))),e.attachBookmarked()}))}))},isStoryValid:function(t){return"story"===t.type&&!!t.id},fetchItem:function(t){return new Promise((function(e,n){Ft.child("item/".concat(t)).on("value",(function(t){var n=t.val(),r={id:d()(n,"id"),title:d()(n,"title"),score:d()(n,"score"),time:z.a.unix(d()(n,"time",0)).format("YYYY-MM-DD HH:mm:ss"),url:d()(n,"url"),type:d()(n,"type"),nComments:d()(n,"descendants"),kids:d()(n,"kids",[]),status:{bookmarked:!1,saving:!1}};e(r)})),Ft.onDisconnect((function(t,e){this.$alert("Firebase connection lost","Lost connection",{confirmButtonText:"OK"}),console.log(t,e)}))}))},attachBookmarked:function(){var t,e=this;return(t=i.a.mark((function t(){var n,r;return i.a.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.next=2,u.a.get("/api/bookmarks");case 2:n=t.sent,r=d()(n,"data.data",[]),e.nBookmarks=r.length,e.posts=e.posts.map((function(t){return t.status.bookmarked=r.includes(t.id),t}));case 6:case"end":return t.stop()}}),t)})),function(){var e=this,n=arguments;return new Promise((function(r,s){var a=t.apply(e,n);function o(t){Bt(a,r,s,o,i,"next",t)}function i(t){Bt(a,r,s,o,i,"throw",t)}o(void 0)}))})()}},watch:{idList:function(t,e){var n=this;this.loading=!0,setTimeout((function(){return n.fetchItems(t)}),400)},nBookmarks:function(t,e){this.$emit("nBookmarksChangedEvent",t)}}};n("0P3G");function Lt(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}var qt={name:"HackerNewsList",components:{Navigation:Et,HnPosts:Object(U.a)(zt,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("ul",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}],staticClass:"list-group",attrs:{id:"hn-posts"}},[t._l(Array(t.firstBatchPosition).fill(null),(function(e){return t.loading?n("li",{staticClass:"list-group-item list-group-item-action flex-column align-items-start"},[t._m(0,!0),t._v(" "),t._m(1,!0)]):t._e()})),t._v(" "),t._l(t.posts,(function(e,r){return n("li",{key:e.id,staticClass:"list-group-item list-group-item-action flex-column align-items-start",attrs:{"data-index":r+1}},[n("div",{staticClass:"d-flex w-100 justify-content-between"},[n("span",[n("router-link",{staticClass:"text-body",attrs:{to:Object.assign({},t.hackerNewsPostRoute,{params:{id:e.id}})}},[t._v(t._s(e.title))]),t._v(" "),e.url?n("a",{staticClass:"text-body",attrs:{href:e.url,target:"_blank"}},[n("small",{staticClass:"text-muted"},[t._v("("+t._s(t._f("domainFromUrl")(e.url))+")")])]):t._e()],1),t._v(" "),n("span",{directives:[{name:"loading",rawName:"v-loading",value:e.status.saving,expression:"post.status.saving"}],staticClass:"d-block d-md-none"},[n("span",{staticClass:"bookmark-story pointer text-primary",on:{click:function(n){return t.bookmarkPost(e)}}},[t._v(t._s(e.status.bookmarked?"⚫":"⚪️")+"️")])]),t._v(" "),n("span",{staticClass:"badge d-none d-md-block",attrs:{title:e.time}},[t._v(t._s(t._f("diffForHumans")(e.time)))])]),t._v(" "),n("div",{directives:[{name:"loading",rawName:"v-loading",value:e.status.saving,expression:"post.status.saving"}],staticClass:"d-none d-md-block"},[n("small",{staticClass:"text-muted"},[t._v(t._s(t.$I18n.trans("hackernews.points",{points:e.score||""})))]),t._v("\n      |\n      "),n("small",[t._v(t._s(t.$I18n.trans("hackernews.comments",{comments:e.nComments||""})))]),t._v("\n      |\n      "),n("span",{staticClass:"bookmark-story pointer text-primary",on:{click:function(n){return t.bookmarkPost(e)}}},[t._v(t._s(e.status.bookmarked?"⚫":"⚪️"))])])])}))],2)}),[function(){var t=this.$createElement,e=this._self._c||t;return e("div",{staticClass:"d-flex w-100 justify-content-between"},[e("span",{staticClass:"loader-text loader-text-top d-block w-100"},[this._v(" ")])])},function(){var t=this.$createElement,e=this._self._c||t;return e("div",{staticClass:"d-none d-md-block"},[e("span",{staticClass:"loader-text loader-text-bottom loader-text--1x3 d-inline-block"},[e("small",[this._v(" ")])])])}],!1,null,"a624b0ec",null).exports},props:{type:{type:String,validator:function(t){return-1!==["top","best","bookmarks"].indexOf(t)},default:"top",required:!0}},data:function(){return{idList:[],limit:100,types:{HackerNewsBookmarks:"bookmarks"},numberOfBookmarks:null}},methods:{fetchStories:function(){this.type===this.types.HackerNewsBookmarks?this.fetchIdsBookmarkedFromBackend():this.fetchIdsFromFirebase(this.type)},fetchIdsFromFirebase:function(t){var e=this;Ft.child("".concat(t,"stories")).limitToFirst(this.limit).once("value",(function(t,n){var r=t.val();e.idList=Array.isArray(r)?r:[]}))},fetchIdsBookmarkedFromBackend:function(){var t,e=this;return(t=i.a.mark((function t(){var n;return i.a.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.next=2,u.a.get("/api/bookmarks");case 2:n=t.sent,e.idList=d()(n,"data.data",[]);case 4:case"end":return t.stop()}}),t)})),function(){var e=this,n=arguments;return new Promise((function(r,s){var a=t.apply(e,n);function o(t){Lt(a,r,s,o,i,"next",t)}function i(t){Lt(a,r,s,o,i,"throw",t)}o(void 0)}))})()},updateNumberOfBookmarks:function(t){this.numberOfBookmarks=t}},created:function(){this.fetchStories()},watch:{$route:function(t,e){this.fetchStories()}}},Mt=Object(U.a)(qt,(function(){var t=this.$createElement,e=this._self._c||t;return e("div",{attrs:{id:"hacker-news-list"}},[e("div",{staticClass:"row"},[e("div",{staticClass:"col-12 no-gutter-xs"},[e("hn-posts",this._g({attrs:{"id-list":this.idList}},this.$listeners))],1)])])}),[],!1,null,null,null).exports,Qt={name:"HnComment",props:{comment:{type:Object,required:!0},handleCollapseToggle:{type:Function,required:!0}},data:function(){return{isShow:!0}}};n("P9EP");function Yt(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}var Ht={name:"HackerNewsPost",mixins:[Nt],components:{Navigation:Et,HnComment:Object(U.a)(Qt,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"hn-comment card mt-3"},[n("div",{staticClass:"card-body px-2 px-md-3 shadow"},[n("h6",{staticClass:"card-subtitle bg-light text-muted d-flex justify-content-between align-items-center",class:{"opacity-5 text-line-through":t.comment.deleted}},[t.comment.id?n("span",{staticClass:"align-middle"},[t._v(t._s(t.$I18n.trans("hackernews.by",{by:t.comment.by||""})))]):t._e(),t._v(" "),n("small",{staticClass:"text-muted",attrs:{title:t.comment.time}},[t._v(t._s(t._f("diffForHumans")(t.comment.time)))]),t._v(" "),t.comment.id?n("a",{staticClass:"btn btn-sm pointer btn-collapse",attrs:{role:"button","data-toggle":"collapse"},on:{click:function(e){return t.handleCollapseToggle(t.comment)}}},[n("b",[t._v("["),t.comment.collapsed?n("span",[t._v(t._s(t.comment.kids.length))]):n("span",[t._v("-")]),t._v("]\n        ")])]):t._e()]),t._v(" "),n("transition",{attrs:{name:"slide"}},[t.comment.collapsed?t._e():n("div",{staticClass:"mt-2"},[n("p",{domProps:{innerHTML:t._s(t.comment.text)}}),t._v(" "),t._l(t.comment.comments,(function(e){return t.comment.comments.length?n("hn-comment",{key:e.id,attrs:{comment:e,"handle-collapse-toggle":t.handleCollapseToggle}}):t._e()}))],2)])],1)])}),[],!1,null,"1a92697e",null).exports},props:{id:{type:Number|String,required:!0}},data:function(){return{loading:!0,loadingComments:!0,collapsedCommentsKey:"collapsedCommentsHackerNewsPost",collapsedComments:[],post:{id:null,by:null,title:null,text:null,score:null,time:null,url:null,type:null,nComments:null,status:{bookmarked:null,saving:null},kids:[],comments:[{id:null,by:null,kids:[],parent:null,text:null,time:null,comments:[]}]},numberOfBookmarks:null}},methods:{fetchPost:function(t){var e=this;t&&(this.fetchCollapsedComments(),Ft.child("item/".concat(t)).once("value",(function(t){var n=t.val();e.post.id=d()(n,"id"),e.post.by=d()(n,"by"),e.post.title=d()(n,"title"),e.post.text=d()(n,"text"),e.post.score=d()(n,"score"),e.post.time=z.a.unix(d()(n,"time",0)).format("YYYY-MM-DD HH:mm:ss"),e.post.url=d()(n,"url"),e.post.type=d()(n,"type"),e.post.nComments=d()(n,"descendants"),e.post.bookmarked=null,e.post.kids=d()(n,"kids",[]),e.post.comments=[],e.setBookmarkStatus(e.post),e.fetchComments(e.post),e.loading=!1})))},fetchComments:function(t){var e=this;t.kids.length&&t.kids.map((function(n){e.fetchComment(n).then((function(n){var r=d()(n,"kids",[]),s=d()(n,"id"),a={id:s,deleted:d()(n,"deleted"),by:d()(n,"by"),kids:r,text:d()(n,"text"),time:z.a.unix(d()(n,"time",0)).format("YYYY-MM-DD HH:mm:ss"),comments:[],collapsed:e.collapsedComments.includes(s)};e.fetchComments(a),t.comments.push(a),e.loadingComments=!1}))}))},fetchComment:function(t){return new Promise((function(e,n){Ft.child("item/".concat(t)).once("value",(function(t){var n=t.val();e(n)}))}))},setBookmarkStatus:function(t){var e,n=this;return(e=i.a.mark((function e(){var r,s;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,u.a.get("/api/bookmarks");case 2:r=e.sent,s=d()(r,"data.data",[]),n.numberOfBookmarks=s.length,t.status.bookmarked=s.includes(t.id);case 6:case"end":return e.stop()}}),e)})),function(){var t=this,n=arguments;return new Promise((function(r,s){var a=e.apply(t,n);function o(t){Yt(a,r,s,o,i,"next",t)}function i(t){Yt(a,r,s,o,i,"throw",t)}o(void 0)}))})()},fetchCollapsedComments:function(){var t,e=localStorage.getItem(this.collapsedCommentsKey);try{t=JSON.parse(e)}catch(t){console.error(t)}t=Array.isArray(t)?t:[],this.collapsedComments=t},setCollapsedComments:function(){var t=JSON.stringify(this.collapsedComments);localStorage.setItem(this.collapsedCommentsKey,t)},handleCollapseToggle:function(t){t.collapsed=!t.collapsed,t.collapsed?this.collapsedComments.push(t.id):this.collapsedComments=this.collapsedComments.filter((function(e){return e!==t.id})),this.setCollapsedComments()}},created:function(){this.fetchPost(this.id)}},Vt=(n("zTjN"),Object(U.a)(Ht,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"hacker-news-post"}},[n("div",{staticClass:"row"},[n("div",{staticClass:"col-12 no-gutter-xs"},[n("div",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}]},[t.loading?n("p",{staticClass:"lead loader-text d-block loader-text--2x3"},[t._v(" ")]):t._e(),t._v(" "),t.post.title?n("p",{staticClass:"lead"},[t._v("\n          "+t._s(t.post.title)+"\n          "),t.post.url?n("a",{staticClass:"text-body",attrs:{href:t.post.url,target:"_blank"}},[n("small",{staticClass:"text-muted"},[t._v("("+t._s(t._f("domainFromUrl")(t.post.url))+") [↗]")])]):t._e()]):t._e(),t._v(" "),t.post.text?n("p",{domProps:{innerHTML:t._s(t.post.text)}}):t._e(),t._v(" "),t.loading?n("div",{staticClass:"loader-text d-block loader-text--1x3"},[t._v(" ")]):t._e(),t._v(" "),t.post.title?n("div",{directives:[{name:"loading",rawName:"v-loading",value:t.post.status.saving,expression:"post.status.saving"}]},[n("small",{staticClass:"text-muted"},[t._v(t._s(t.$I18n.trans("hackernews.points",{points:t.post.score})))]),t._v(" "),n("span",{staticClass:"text-muted"},[t._v("|")]),t._v(" "),n("small",{staticClass:"text-muted"},[t._v(t._s(t.$I18n.trans("hackernews.comments",{comments:t.post.nComments})))]),t._v(" "),n("span",{staticClass:"text-muted"},[t._v("|")]),t._v(" "),n("small",{staticClass:"text-muted",attrs:{title:t.post.time}},[t._v(t._s(t._f("diffForHumans")(t.post.time)))]),t._v(" "),n("span",{staticClass:"text-muted"},[t._v("|")]),t._v(" "),n("small",{staticClass:"text-muted"},[t._v(t._s(t.$I18n.trans("hackernews.by",{by:t.post.by})))]),t._v(" "),n("span",{staticClass:"text-muted"},[t._v("|")]),t._v(" "),n("span",{staticClass:"bookmark-story pointer text-primary",on:{click:function(e){return t.bookmarkPost(t.post)}}},[t._v(t._s(t.post.status.bookmarked?"⚫":"⚪️")+"️")])]):t._e()])])]),t._v(" "),t.post.comments.length?n("div",{staticClass:"row"},[n("div",{staticClass:"col-12 no-gutter-xs"},[t.loadingComments?t._e():n("div",t._l(t.post.comments,(function(e){return n("hn-comment",{key:e.id,attrs:{comment:e,"handle-collapse-toggle":t.handleCollapseToggle}})})),1)])]):t._e()])}),[],!1,null,"cdd94d26",null).exports),Jt={name:"Navigation",data:function(){return{usersRoute:Le,usersCreateRoute:ze}}},Wt={name:"Users",components:{Navigation:Object(U.a)(Jt,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"navigation"}},[n("div",{staticClass:"row justify-content-center top-submenu"},[n("div",{staticClass:"col-12 no-gutter-xs"},[n("div",{staticClass:"btn-group d-flex mb-2"},[n("router-link",{staticClass:"btn btn-group-sm w-100 inner-nav btn-secondary",attrs:{to:t.usersRoute}},[t._v(t._s(t.$I18n.trans("users.users")))]),t._v(" "),n("router-link",{staticClass:"btn btn-group-sm w-100 inner-nav btn-secondary",attrs:{to:t.usersCreateRoute}},[t._v(t._s(t.$I18n.trans("users.new-user")))])],1)])])])}),[],!1,null,null,null).exports}},Xt=Object(U.a)(Wt,(function(){var t=this.$createElement,e=this._self._c||t;return e("div",{attrs:{id:"users"}},[e("navigation"),this._v(" "),e("router-view")],1)}),[],!1,null,null,null).exports;function Zt(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}function Kt(t){return function(){var e=this,n=arguments;return new Promise((function(r,s){var a=t.apply(e,n);function o(t){Zt(a,r,s,o,i,"next",t)}function i(t){Zt(a,r,s,o,i,"throw",t)}o(void 0)}))}}function Gt(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(t);e&&(r=r.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,r)}return n}function te(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?Gt(Object(n),!0).forEach((function(e){ee(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):Gt(Object(n)).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}function ee(t,e,n){return e in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}var ne={name:"Users",data:function(){return{loading:!0,userToEdit:{},dialogVisible:!1,dialogFormRules:{name:[{required:!0,message:this.$I18n.trans("users.name_cannot_be_empty"),trigger:"blur"},{min:2,max:255,message:this.$I18n.trans("users.min_2_max_255"),trigger:"blur"}],password:[{min:12,max:100,message:this.$I18n.trans("users.please_set_valid_password"),trigger:"blur"}]}}},computed:Object(a.c)(["usersList","currentUserId"]),methods:te(te({},Object(a.b)(["fetchUsers","updateUser","destroyUser"])),{},{highlightCurrentUser:function(t){var e=t.row;t.rowIndex;return e.id===this.currentUserId?"table-success":null},formatDate:function(t){return this.$options.filters.diffForHumans(t)},formatUpdatedAt:function(t,e){return this.formatDate(t.updatedAt)},closeDialog:function(t){this.dialogVisible=!1,this.$refs[t].resetFields()},editDialogShow:function(t){this.userToEdit={id:t.id,name:t.name,email:t.email},this.dialogVisible=!0},editDialogSubmit:function(t){var e=this;return Kt(i.a.mark((function n(){var r,s;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:return n.next=2,e.validateDialogSubmit(t);case 2:if(!n.sent){n.next=10;break}return r={id:e.userToEdit.id,name:e.userToEdit.name},e.userToEdit.password&&(r.password=e.userToEdit.password),n.next=8,e.updateUser(r);case 8:s=n.sent,e.notifyActionSuccess(s);case 10:case"end":return n.stop()}}),n)})))()},validateDialogSubmit:function(t){var e=this;return Kt(i.a.mark((function n(){return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:return n.abrupt("return",new Promise((function(n,r){e.$refs[t].validate((function(t){return t?n(!0):r(!1)}))})));case 1:case"end":return n.stop()}}),n)})))()},notifyActionSuccess:function(t){var e=this,n=arguments.length>1&&void 0!==arguments[1]?arguments[1]:"users.user_updated",r=arguments.length>2&&void 0!==arguments[2]?arguments[2]:"users.failed_to_update";if(t){this.$notify.success(this.$I18n.trans(n));var s=this.usersList.find((function(t){return t.id===e.userToEdit.id}));s&&(s.name=this.userToEdit.name.trim()),this.dialogVisible=!1}else this.$notify.error(this.$I18n.trans(r))},deleteUser:function(){var t=this;return Kt(i.a.mark((function e(){var n;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,t.askUserToConfirmDelete();case 2:if(!e.sent){e.next=8;break}return e.next=6,t.destroyUser(t.userToEdit.id);case 6:n=e.sent,t.notifyActionSuccess(n,"users.delete_success","users.failed_to_delete");case 8:case"end":return e.stop()}}),e)})))()},askUserToConfirmDelete:function(){var t=this;return Kt(i.a.mark((function e(){var n;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.prev=0,e.next=3,t.$confirm(t.$I18n.trans("users.confirm_delete_request",{name:t.userToEdit.name}),t.$I18n.trans("common.please_confirm"),{confirmButtonText:t.$I18n.trans("users.delete"),cancelButtonText:t.$I18n.trans("common.cancel"),type:"warning"});case 3:e.t0=e.sent,n="confirm"===e.t0,e.next=10;break;case 7:e.prev=7,e.t1=e.catch(0),n=!1;case 10:return e.abrupt("return",n);case 11:case"end":return e.stop()}}),e,null,[[0,7]])})))()}}),created:function(){var t=this;this.fetchUsers().then((function(e){return t.loading=!1}))}},re=(n("pM8S"),Object(U.a)(ne,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"users-list"}},[n("div",{staticClass:"row"},[n("div",{staticClass:"col-12 mt-3 no-gutter-xs"},[n("el-table",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}],staticStyle:{width:"100%"},attrs:{data:t.usersList,"row-class-name":t.highlightCurrentUser,"empty-text":t.$I18n.trans("users.failed_to_obtain_user_list")}},[n("el-table-column",{attrs:{prop:"id",label:"#"}}),t._v(" "),n("el-table-column",{attrs:{prop:"name",label:t.$I18n.trans("users.name")}}),t._v(" "),n("el-table-column",{attrs:{prop:"email",label:t.$I18n.trans("users.email")}}),t._v(" "),n("el-table-column",{attrs:{prop:"updatedAt",label:t.$I18n.trans("users.updated_at"),formatter:t.formatUpdatedAt}}),t._v(" "),n("el-table-column",{attrs:{fixed:"right",align:"right",width:"60",prop:"updatedAt"},scopedSlots:t._u([{key:"default",fn:function(e){return[n("el-button",{attrs:{type:"text",size:"small"},on:{click:function(n){return t.editDialogShow(e.row)}}},[t._v(t._s(t.$I18n.trans("users.edit")))])]}}])})],1)],1)]),t._v(" "),n("el-dialog",{attrs:{title:t.$I18n.trans("users.user_edit",{id:t.userToEdit.id||""}),visible:t.dialogVisible},on:{"update:visible":function(e){t.dialogVisible=e}}},[n("el-form",{ref:"formRef",attrs:{model:t.userToEdit,rules:t.dialogFormRules},on:{submit:t.editDialogSubmit},nativeOn:{submit:function(t){t.preventDefault()}}},[n("el-form-item",{attrs:{label:t.$I18n.trans("users.name"),prop:"name"}},[n("el-input",{attrs:{placeholder:t.$I18n.trans("users.name_cannot_be_empty")},model:{value:t.userToEdit.name,callback:function(e){t.$set(t.userToEdit,"name",e)},expression:"userToEdit.name"}})],1),t._v(" "),n("el-form-item",{attrs:{label:t.$I18n.trans("users.password"),prop:"password"}},[n("el-input",{attrs:{type:"password",autocomplete:"off"},model:{value:t.userToEdit.password,callback:function(e){t.$set(t.userToEdit,"password",e)},expression:"userToEdit.password"}})],1),t._v(" "),n("el-form-item",{attrs:{label:t.$I18n.trans("users.email")}},[n("el-input",{attrs:{type:"email",value:t.userToEdit.email,disabled:""}})],1)],1),t._v(" "),n("span",{staticClass:"dialog-footer",attrs:{slot:"footer"},slot:"footer"},[n("el-button",{on:{click:function(e){return t.closeDialog("formRef")}}},[t._v(t._s(t.$I18n.trans("common.cancel")))]),t._v(" "),n("el-button",{attrs:{type:"danger",disabled:t.userToEdit.id===t.currentUserId},on:{click:function(e){return t.deleteUser()}}},[t._v(t._s(t.$I18n.trans("users.delete")))]),t._v(" "),n("el-button",{attrs:{type:"primary"},on:{click:function(e){return t.editDialogSubmit("formRef")}}},[t._v(t._s(t.$I18n.trans("common.confirm")))])],1)],1)],1)}),[],!1,null,"21c16506",null).exports);function se(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}function ae(t){return function(){var e=this,n=arguments;return new Promise((function(r,s){var a=t.apply(e,n);function o(t){se(a,r,s,o,i,"next",t)}function i(t){se(a,r,s,o,i,"throw",t)}o(void 0)}))}}function oe(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(t);e&&(r=r.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,r)}return n}function ie(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?oe(Object(n),!0).forEach((function(e){ce(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):oe(Object(n)).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}function ce(t,e,n){return e in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}var ue={name:"UsersCreate",data:function(){return{rules:{name:[{required:!0,message:this.$I18n.trans("users.name_cannot_be_empty"),trigger:"blur"},{min:2,max:255,message:this.$I18n.trans("users.min_2_max_255"),trigger:"blur"}],email:[{required:!0,type:"email",message:this.$I18n.trans("users.please_set_valid_email"),trigger:"blur"}],password:[{required:!0,message:this.$I18n.trans("users.please_set_valid_password"),trigger:"blur"},{min:12,max:100,message:this.$I18n.trans("users.please_set_valid_password"),trigger:"blur"}]},user:{name:"",email:"",password:""}}},methods:ie(ie({},Object(a.b)(["addUser"])),{},{submitForm:function(t){var e=this;return ae(i.a.mark((function n(){return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:return n.next=2,e.validateForm(t);case 2:n.sent&&e.addUser(e.user).then((function(t){return e.notifyActionResult(t)}));case 4:case"end":return n.stop()}}),n)})))()},validateForm:function(t){var e=this;return ae(i.a.mark((function n(){return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:return n.abrupt("return",new Promise((function(n,r){e.$refs[t].validate((function(t){t&&n(!0),r(!1)}))})));case 1:case"end":return n.stop()}}),n)})))()},notifyActionResult:function(t){var e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:"users.created_success",n=arguments.length>2&&void 0!==arguments[2]?arguments[2]:"users.failed_to_create";t?(this.$notify.success(this.$I18n.trans(e)),this.$router.push(Le)):this.$notify.error(this.$I18n.trans(n))}})},le=Object(U.a)(ue,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"users-create"}},[n("div",{staticClass:"row"},[n("div",{staticClass:"col-12 mt-3 no-gutter-xs"},[n("el-form",{ref:"refCreateUserForm",attrs:{model:t.user,rules:t.rules}},[n("el-form-item",{attrs:{label:t.$I18n.trans("users.name"),prop:"name"}},[n("el-input",{model:{value:t.user.name,callback:function(e){t.$set(t.user,"name",e)},expression:"user.name"}})],1),t._v(" "),n("el-form-item",{attrs:{label:t.$I18n.trans("users.email"),prop:"email"}},[n("el-input",{attrs:{type:"email",autocomplete:"off"},model:{value:t.user.email,callback:function(e){t.$set(t.user,"email",e)},expression:"user.email"}})],1),t._v(" "),n("el-form-item",{attrs:{label:t.$I18n.trans("users.password"),prop:"password"}},[n("el-input",{attrs:{type:"password",autocomplete:"off"},model:{value:t.user.password,callback:function(e){t.$set(t.user,"password",e)},expression:"user.password"}})],1),t._v(" "),n("el-form-item",[n("el-button",{attrs:{type:"primary"},on:{click:function(e){return t.submitForm("refCreateUserForm")}}},[t._v(t._s(t.$I18n.trans("users.create")))])],1)],1)],1)])])}),[],!1,null,null,null).exports,de={name:"Navigation",data:function(){return{expensesRoute:Me,expensesCreateRoute:qe}}},pe={name:"Expenses",components:{Navigation:Object(U.a)(de,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"navigation"}},[n("div",{staticClass:"row justify-content-center top-submenu"},[n("div",{staticClass:"col-12 no-gutter-xs"},[n("div",{staticClass:"btn-group d-flex mb-2"},[n("router-link",{staticClass:"btn btn-group-sm w-100 inner-nav btn-secondary",attrs:{to:t.expensesRoute}},[t._v(t._s(t.$I18n.trans("expenses.expenses")))]),t._v(" "),n("router-link",{staticClass:"btn btn-group-sm w-100 inner-nav btn-secondary",attrs:{to:t.expensesCreateRoute}},[t._v(t._s(t.$I18n.trans("expenses.new_expense")))])],1)])])])}),[],!1,null,null,null).exports}},me=Object(U.a)(pe,(function(){var t=this.$createElement,e=this._self._c||t;return e("div",{attrs:{id:"expenses"}},[e("navigation"),this._v(" "),e("router-view")],1)}),[],!1,null,null,null).exports,fe=n("ZhIB"),ve=n.n(fe);function he(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(t);e&&(r=r.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,r)}return n}function be(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?he(Object(n),!0).forEach((function(e){ge(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):he(Object(n)).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}function ge(t,e,n){return e in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}var xe={name:"ExpensesList",data:function(){return{loading:!0}},computed:Object(a.c)(["expensesList"]),methods:be(be({},Object(a.b)(["fetchExpenses"])),{},{getSummaries:function(t){var e=this,n=t.columns,r=t.data,s=[];return n.forEach((function(t,n){if(0===n)s[n]=e.$I18n.trans("expenses.total");else if(1===n)s[n]="";else{var a=r.map((function(e){return Number(e[t.property])})).reduce((function(t,e){var n=Number(e);return isNaN(n)?t:t+e}),0);s[n]=ve()(a).format("$ 0,0[.]00")}})),s},formatCurr:function(t,e,n,r){return ve()(Number(n)).format("$ 0,0[.]00")},formatDate:function(t,e,n,r){var s=z()(n),a=s.format("YYYY-MM-DD"),o=s.fromNow();return"".concat(o," on ").concat(a)}}),mounted:function(){var t=this;this.fetchExpenses().then((function(e){return t.loading=!1}))}},ye=Object(U.a)(xe,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"expenses-list"}},[n("div",{staticClass:"row"},[n("div",{staticClass:"col-12 mt-3 no-gutter-xs"},[n("div",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}]},[n("el-table",{staticStyle:{width:"100%"},attrs:{data:t.expensesList,"show-summary":"","summary-method":t.getSummaries,"empty-text":t.$I18n.trans("expenses.failed_to_obtain_expenses")}},[n("el-table-column",{attrs:{prop:"description","min-width":"250",label:t.$I18n.trans("expenses.what")}}),t._v(" "),n("el-table-column",{attrs:{prop:"updated_at",formatter:t.formatDate,"min-width":"220",label:t.$I18n.trans("expenses.when")}}),t._v(" "),n("el-table-column",{attrs:{prop:"amount",align:"right",width:"100",label:t.$I18n.trans("expenses.amount"),formatter:t.formatCurr,fixed:"right"}})],1)],1)])])])}),[],!1,null,null,null).exports;function we(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}function _e(t){return function(){var e=this,n=arguments;return new Promise((function(r,s){var a=t.apply(e,n);function o(t){we(a,r,s,o,i,"next",t)}function i(t){we(a,r,s,o,i,"throw",t)}o(void 0)}))}}function ke(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(t);e&&(r=r.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,r)}return n}function je(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?ke(Object(n),!0).forEach((function(e){Ce(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):ke(Object(n)).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}function Ce(t,e,n){return e in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}var Oe={name:"ExpensesCreate",data:function(){return{expense:{description:null,amount:null},rules:{description:[{required:!0,message:this.$I18n.trans("expenses.expense_description_required"),trigger:"change"},{min:2,max:255,message:this.$I18n.trans("expenses.min_2_max_255"),trigger:"change"}],amount:[{required:!0,message:this.$I18n.trans("expenses.amount_is_required"),trigger:"change"},{type:"regexp",pattern:"/^\\d*\\.?\\d*$/",message:this.$I18n.trans("expenses.please_set_valid_amount"),trigger:"change"}]}}},methods:je(je({},Object(a.b)(["addExpense"])),{},{onSubmit:function(t){var e=this;return _e(i.a.mark((function n(){return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:e.$refs[t].validate(function(){var n=_e(i.a.mark((function n(r){return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:if(!r){n.next=9;break}return n.next=3,e.createExpense();case 3:return e.resetForm(t),n.next=6,e.$router.push(Me);case 6:return n.abrupt("return",!0);case 9:return n.abrupt("return",!1);case 10:case"end":return n.stop()}}),n)})));return function(t){return n.apply(this,arguments)}}());case 1:case"end":return n.stop()}}),n)})))()},createExpense:function(){var t=this;return _e(i.a.mark((function e(){var n,r;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,t.addExpense(t.expense);case 2:e.sent?(n=t.$I18n.trans("expenses.success_expense_added"),t.$notify.success(n),t.expense.description="",t.expense.amount=""):(r=t.$I18n.trans("expenses.fail_expense_added"),t.$notify.error(r));case 4:case"end":return e.stop()}}),e)})))()},resetForm:function(t){this.$refs[t].resetFields()}})},Ie=(n("RG2k"),Object(U.a)(Oe,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"expenses-add"},[n("div",{staticClass:"row"},[n("div",{staticClass:"col-12 mt-3 no-gutter-xs"},[n("el-form",{ref:"refExpenseForm",attrs:{model:t.expense,rules:t.rules}},[n("el-form-item",{attrs:{label:t.$I18n.trans("expenses.what"),prop:"description"}},[n("el-input",{attrs:{clearable:""},model:{value:t.expense.description,callback:function(e){t.$set(t.expense,"description",e)},expression:"expense.description"}})],1),t._v(" "),n("el-form-item",{attrs:{label:t.$I18n.trans("expenses.amount"),prop:"amount"}},[n("el-input",{attrs:{type:"number",inputmode:"decimal",decimal:"true",clearable:"",min:"0",step:"0.01"},model:{value:t.expense.amount,callback:function(e){t.$set(t.expense,"amount",e)},expression:"expense.amount"}})],1),t._v(" "),n("el-form-item",[n("el-button",{attrs:{type:"primary"},on:{click:function(e){return t.onSubmit("refExpenseForm")}}},[t._v(t._s(t.$I18n.trans("common.add")))])],1)],1)],1)])])}),[],!1,null,"0a4f35a8",null).exports);function $e(t,e,n,r,s,a,o){try{var i=t[a](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(r,s)}s.a.use(E.a);var Ee={path:"/authenticate",name:"Authenticate",component:F,meta:{requiresAuth:!1}},Pe={path:"/dashboard",alias:"/",name:"Dashboard",component:W,meta:{requiresAuth:!0}},Te={path:"new",name:"NotesCreate",component:xt,meta:{requiresAuth:!0}},Se={path:":id",name:"NotesShow",component:kt,props:!0,meta:{requiresAuth:!0}},Ne={path:"edit/:id",name:"NotesUpdate",component:It,props:!0,meta:{requiresAuth:!0}},De={path:"/notes",component:K,children:[{path:"",name:"Notes",component:dt,meta:{requiresAuth:!0}},Te,Se,Ne]},Ue={path:"",alias:"top",name:"HackerNewsTopPostsRoute",component:Mt,props:{type:"top"},meta:{requiresAuth:!0}},Fe={path:"best",name:"HackerNewsBestPostsRoute",component:Mt,props:{type:"best"},meta:{requiresAuth:!0}},Be={path:"bookmarks",name:"HackerNewsBookmarkedPostsRoute",component:Mt,props:{type:"bookmarks"},meta:{requiresAuth:!0}},Re={path:":id",name:"HackerNewsPost",component:Vt,props:!0,meta:{requiresAuth:!0}},Ae={path:"/hn",component:Tt,children:[Ue,Fe,Be,Re]},ze={path:"new",name:"UsersCreate",component:le,meta:{requiresAuth:!0}},Le={path:"/users",component:Xt,children:[{path:"",name:"UsersList",component:re,meta:{requiresAuth:!0}},ze]},qe={path:"new",name:"ExpensesCreate",component:Ie,meta:{requiresAuth:!0}},Me={path:"/expenses",component:me,children:[{path:"",name:"ExpensesList",component:ye,meta:{requiresAuth:!0}},qe]},Qe={path:"/404",name:"NotFound",component:R,props:!0,meta:{requiresAuth:!0}},Ye=new E.a({routes:[Ee,Pe,De,Ae,Le,Me,Qe,{path:"*",redirect:"/404",meta:{requiresAuth:!0}}]});Ye.beforeEach(function(){var t,e=(t=i.a.mark((function t(e,n,r){return i.a.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:if(e.matched.some((function(t){return t.meta.requiresAuth}))){t.next=4;break}return r(),t.abrupt("return");case 4:return t.next=6,k();case 6:if(!t.sent){t.next=11;break}return e.name===Ee.name&&r(Pe),r(),t.abrupt("return");case 11:r(Ee);case 12:case"end":return t.stop()}}),t)})),function(){var e=this,n=arguments;return new Promise((function(r,s){var a=t.apply(e,n);function o(t){$e(a,r,s,o,i,"next",t)}function i(t){$e(a,r,s,o,i,"throw",t)}o(void 0)}))});return function(t,n,r){return e.apply(this,arguments)}}());var He=n("XJYT"),Ve=n.n(He),Je=(n("D66Q"),n("stYL")),We=n.n(Je);function Xe(t){return(Xe="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t})(t)}function Ze(t,e){return function(t){if(Array.isArray(t))return t}(t)||function(t,e){if("undefined"==typeof Symbol||!(Symbol.iterator in Object(t)))return;var n=[],r=!0,s=!1,a=void 0;try{for(var o,i=t[Symbol.iterator]();!(r=(o=i.next()).done)&&(n.push(o.value),!e||n.length!==e);r=!0);}catch(t){s=!0,a=t}finally{try{r||null==i.return||i.return()}finally{if(s)throw a}}return n}(t,e)||function(t,e){if(!t)return;if("string"==typeof t)return Ke(t,e);var n=Object.prototype.toString.call(t).slice(8,-1);"Object"===n&&t.constructor&&(n=t.constructor.name);if("Map"===n||"Set"===n)return Array.from(t);if("Arguments"===n||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n))return Ke(t,e)}(t,e)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}()}function Ke(t,e){(null==e||e>t.length)&&(e=t.length);for(var n=0,r=new Array(e);n<e;n++)r[n]=t[n];return r}function Ge(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}function tn(t,e){for(var n=0;n<e.length;n++){var r=e[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(t,r.key,r)}}s.a.filter("diffForHumans",(function(t){if(t)return z()(t,"YYYY-MM-DD hh:mm:ss").fromNow()})),s.a.filter("domainFromUrl",(function(t){if(t)return new URL(t).hostname.replace("www.","")}));var en=function(){function t(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:"translations";Ge(this,t),this.key=e}var e,n,r;return e=t,(n=[{key:"trans",value:function(t){var e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{};return this._replace(this._extract(t),e)}},{key:"trans_choice",value:function(t){var e,n=this,r=arguments.length>1&&void 0!==arguments[1]?arguments[1]:1,s=arguments.length>2&&void 0!==arguments[2]?arguments[2]:{},a=this._extract(t,"|").split("|");return a.some((function(t){return e=n._match(t,r)})),e=(e=e||(r>1?a[1]:a[0])).replace(/\[.*?\]|\{.*?\}/,""),this._replace(e,s)}},{key:"_match",value:function(t,e){var n=t.match(/^[\{\[]([^\[\]\{\}]*)[\}\]](.*)/);if(n){if(n[1].includes(",")){var r=Ze(n[1].split(",",2),2),s=r[0],a=r[1];if("*"===a&&e>=s)return n[2];if("*"===s&&e<=a)return n[2];if(e>=s&&e<=a)return n[2]}return n[1]==e?n[2]:null}}},{key:"_replace",value:function(t,e){if("object"===Xe(t))return t;for(var n in e)t=t.toString().replace(":".concat(n),e[n]).replace(":".concat(n.toUpperCase()),e[n].toString().toUpperCase()).replace(":".concat(n.charAt(0).toUpperCase()).concat(n.slice(1)),e[n].toString().charAt(0).toUpperCase()+e[n].toString().slice(1));return t.toString().trim()}},{key:"_extract",value:function(t){var e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:null,n=t.toString().split("::"),r=n.pop().toString().split(".");return n.length>0&&(n[0]+="::"),n.concat(r).reduce((function(n,r){return n[r]||e||t}),window[this.key])}}])&&tn(e.prototype,n),r&&tn(e,r),t}(),nn=n("D/Jt"),rn={name:"MainNavigation",data:function(){return{dashboardRoute:Pe,menuOptions:[{route:De,text:this.$I18n.trans("notes.notes")},{route:Ae,text:this.$I18n.trans("hackernews.hackernews")},{route:Me,text:this.$I18n.trans("expenses.expenses")},{route:Le,text:this.$I18n.trans("users.users")}],showMenu:!0,menuCollapsed:!0,appRoutes:[]}},methods:{logout:function(t){var e=this;t.preventDefault(),u.a.post("/logout",{}).then((function(){window.localStorage.removeItem("ZE_UserInfo"),e.showMenu=!1,e.$router.push(Ee),e.showMenu=!1}))},toggleMenuVisibility:function(t){this.showMenu=Ee.name!==t.name}},watch:{$route:function(t,e){this.toggleMenuVisibility(t)}},mounted:function(){this.toggleMenuVisibility(this.$route)}},sn=(n("HF9f"),Object(U.a)(rn,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return t.showMenu?n("nav",{staticClass:"navbar navbar-expand-md navbar-light navbar-laravel",attrs:{id:"main-navigation"}},[n("div",{staticClass:"container"},[n("router-link",{staticClass:"navbar-brand",attrs:{exact:"",to:t.dashboardRoute}},[t._v("Zetetic Elench")]),t._v(" "),n("button",{staticClass:"navbar-toggler",attrs:{type:"button"},on:{click:function(e){t.menuCollapsed=!t.menuCollapsed}}},[n("span",{staticClass:"navbar-toggler-icon"})]),t._v(" "),n("div",{staticClass:"navbar-collapse",class:{collapse:t.menuCollapsed}},[n("ul",{staticClass:"navbar-nav mr-auto"},t._l(t.menuOptions,(function(e){return n("li",{staticClass:"nav-item"},[n("router-link",{staticClass:"nav-link text-right text-sm-left",attrs:{to:e.route}},[t._v(t._s(e.text))])],1)})),0),t._v(" "),n("ul",{staticClass:"navbar-nav ml-auto"},[n("li",{staticClass:"nav-item"},[n("a",{staticClass:"nav-link text-right text-sm-left",attrs:{href:"#"},on:{click:function(e){return t.logout(e)}}},[t._v(t._s(t.$I18n.trans("users.logout")))])])])])],1)]):t._e()}),[],!1,null,"590a6ef5",null).exports);n("XDMY");ve.a.locale("pt-pt"),u.a.defaults.headers.common["X-Requested-With"]="XMLHttpRequest",u.a.defaults.withCredentials=!0,s.a.component("main-navigation",sn),s.a.use(Ve.a,{locale:We.a}),s.a.use(nn.a),s.a.prototype.$I18n=new en;new s.a({store:$,router:Ye,el:"#app"})},deJl:function(t,e,n){var r=n("fSEL");"string"==typeof r&&(r=[[t.i,r,""]]);var s={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(r,s);r.locals&&(t.exports=r.locals)},eGZ0:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\n.divider[data-v-3d91250f] {\n  border-left: 1px solid #ccc;\n}\n",""])},fSEL:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,'\n.loader-text[data-v-7a7ca61a] {\n  background: #808080;\n  border-radius: 2px;\n}\n.loader-text--1x3[data-v-7a7ca61a] {\n  width: 33%;\n}\n.note[data-v-7a7ca61a] .footnotes {\n  font-size: 0.8rem;\n}\n.note[data-v-7a7ca61a] .footnotes p {\n  margin-bottom: 0.5rem;\n}\n.note[data-v-7a7ca61a] table.table th[align="left"] {\n  text-align: left;\n}\n.note[data-v-7a7ca61a] table.table th[align="center"] {\n  text-align: center;\n}\n.note[data-v-7a7ca61a] table.table th[align="right"] {\n  text-align: right;\n}\n',""])},fysm:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\n.bookmark-story[data-v-a624b0ec] {\n  position: relative;\n  top: 2px;\n  text-decoration: none;\n}\n.loader-text[data-v-a624b0ec] {\n  background: #808080;\n  border-radius: 2px;\n}\n.loader-text--1x3[data-v-a624b0ec] {\n  width: 33%;\n}\n.loader-text-top[data-v-a624b0ec] {\n  height: 22px;\n  margin-bottom: 6px;\n}\n.loader-text-bottom[data-v-a624b0ec] {\n  height: 18px;\n}\n",""])},gE6I:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\n.form-control[data-v-45329d92] {\n  width: auto;\n}\n",""])},iqeS:function(t,e,n){"use strict";var r=n("GFEm");n.n(r).a},mgPi:function(t,e,n){"use strict";var r=n("9bXS");n.n(r).a},nNO0:function(t,e,n){"use strict";var r=n("deJl");n.n(r).a},pM8S:function(t,e,n){"use strict";var r=n("A19t");n.n(r).a},pyCd:function(t,e){},wdM6:function(t,e,n){"use strict";var r=n("8O32");n.n(r).a},"yx/N":function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\ntextarea[data-v-401ef6a5] {\n  outline: 0;\n  background: #f4f4f4;\n  box-shadow: 1px 1px 0 #ccc;\n  border-color: transparent;\n  border-radius: 0;\n  font-family: 'Inconsolata', monospace;\n}\ntextarea[data-v-401ef6a5]:focus {\n  box-shadow: 1px 1px 0 #ccc;\n  outline: 0;\n  background: #f4f4f4;\n  border-color: transparent;\n}\n.new-tag-group[data-v-401ef6a5] {\n  font-size: 0;\n}\n.new-tag-group .input-new-tag[data-v-401ef6a5], .new-tag-group .button-new-tag[data-v-401ef6a5] {\n  vertical-align: middle;\n  width: 120px;\n}\n#note-update[data-v-401ef6a5] .el-checkbox-button--small .el-checkbox-button__inner {\n  border-radius: 4px;\n  border: 1px solid #DCDFE6;\n}\n#note-update[data-v-401ef6a5] .el-checkbox-button.is-checked .el-checkbox-button__inner {\n  border-color: #409EFF;\n}\n",""])},zTjN:function(t,e,n){"use strict";var r=n("D1a/");n.n(r).a}},[[0,1,2]]]);
