+++ public/mix-manifest.json
@@ -1,5 +1,5 @@
 {
-    "/js/app.js": "/js/app.js?id=008972496b7af1e626f5",
+    "/js/app.js": "/js/app.js?id=24608b94632a2ca150b1",
     "/css/app.css": "/css/app.css?id=b7e183f5706765b6bc71",
     "/js/manifest.js": "/js/manifest.js?id=3c768977c2574a34506e",
     "/js/vendor.js": "/js/vendor.js?id=de7eee010ae58ec1ec94"
