+++ resources/js/router.js
@@ -15,6 +15,8 @@ import Users from "./users/views/Users";
 import UsersList from "./users/views/UsersList";
 import UsersCreate from "./users/views/UsersCreate";
 import Expenses from "./expenses/views/Expenses";
+import ExpensesList from "./expenses/views/ExpensesList";
+import ExpensesCreate from "./expenses/views/ExpensesCreate";
 import { isUserAuthenticated } from "./helpers/user";
 
 Vue.use(Router);
@@ -164,13 +166,29 @@ export const UsersRoute = {
 //endregion
 
 //region expenses routes definitions
-export const ExpensesRoute = {
-  path: '/expenses',
-  name: 'Expenses',
-  component: Expenses,
+export const ExpensesListRoute = {
+  path: '',
+  name: 'ExpensesList',
+  component: ExpensesList,
   meta: {
     requiresAuth: true,
   },
+};
+export const ExpensesCreateRoute = {
+  path: 'new',
+  name: 'ExpensesCreate',
+  component: ExpensesCreate,
+  meta: {
+    requiresAuth: true,
+  },
+};
+export const ExpensesRoute = {
+  path: '/expenses',
+  component: Expenses,
+  children: [
+    ExpensesListRoute,
+    ExpensesCreateRoute,
+  ],
 };
 //endregion
